### 2025-08-04 1c3baa8
Keep default drive for SelectDirectoryTreeScreen as home. (#557)
* Keep default drive for SelectDirectoryTreeScreen as home.

* Fix tests under new conditions.
