### 2025-08-04 b95d9a4
Add rclone install check only if call fails.
