### 2025-08-04 b64b064
Add rclone install check only if call fails.
