### 2025-08-04 a43ccbe
Use better return code check.
