### 2025-08-04 bc9cb9d
Revert lsjson for local filesystem searches.
