### 2025-08-05 0847389
add: setup env variables from github secrets in the workflow file
