+++ .github/workflows/code_test_and_deploy.yml
@@ -60,6 +60,16 @@ jobs:
         # run SSH tests only on Linux because Windows and macOS
         # are already run within a virtual container and so cannot
         # run Linux containers because nested containerisation is disabled.
+      - name: Setup Google Drive and AWS Secrets
+        run: |
+          echo "GDRIVE_CLIENT_ID=${{secrets.GDRIVE_CLIENT_ID}}" >> $GITHUB_ENV
+          echo "GDRIVE_CLIENT_SECRET=${{secrets.GDRIVE_CLIENT_SECRET}}" >> $GITHUB_ENV
+          echo "GDRIVE_ROOT_FOLDER_ID=${{secrets.GDRIVE_ROOT_FOLDER_ID}}" >> $GITHUB_ENV
+          echo "GDRIVE_CONFIG_TOKEN=${{secrets.GDRIVE_CONFIG_TOKEN}}" >> $GITHUB_ENV
+          echo "AWS_ACCESS_KEY_ID=${{secrets.AWS_ACCESS_KEY_ID}}" >> $GITHUB_ENV
+          echo "AWS_REGION=${{secrets.AWS_REGION}}" >> $GITHUB_ENV
+          echo "AWS_BUCKET_NAME=${{secrets.AWS_BUCKET_NAME}}" >> $GITHUB_ENV
+          echo "AWS_SECRET_ACCESS_KEY=${{secrets.AWS_SECRET_ACCESS_KEY}}" >> $GITHUB_ENV
       - name: Test SSH (Linux only)
         if: runner.os == 'Linux'
         run: |
