### 2025-08-05 30bbd6c
Fix recursion.
