### 2025-08-05 8bb936b
Fix recursion.
