### 2025-08-05 4dba40d
Use internal error function to raise error.
