### 2025-08-05 03f1115
Use internal error function to raise error.
