+++ datashuttle/utils/rclone.py
@@ -195,12 +195,14 @@ def log_rclone_config_output() -> None:
 def prompt_rclone_download_if_does_not_exist() -> None:
     """Check that rclone is installed."""
     if not check_rclone_with_default_call():
+
         newline = "" if "PYTEST_CURRENT_TEST" in os.environ else "\n"
 
-        raise BaseException(
+        utils.log_and_raise_error(
             f"RClone installation not found. Install by entering "
             f"the following into your terminal:{newline}"
-            f"  conda install -c conda-forge rclone"
+            f"  conda install -c conda-forge rclone",
+            RuntimeError
         )
 
 
