+++ tests/tests_transfers/gdrive/gdrive_test_utils.py
@@ -1,6 +1,8 @@
 import builtins
 import copy
 import os
+import random
+import string
 
 from datashuttle import DataShuttle
 from datashuttle.utils import gdrive
@@ -10,9 +12,12 @@ def setup_project_for_gdrive(project: DataShuttle):
     """
     Set up a project with configs for Google Drive transfers.
     """
+    characters = string.ascii_letters + string.digits
+    random_string = "".join(random.choices(characters, k=15))
+
     project.update_config_file(
         connection_method="gdrive",
-        central_path=f"/main/{project.project_name}",
+        central_path=f"/main/{random_string}/{project.project_name}",
         gdrive_client_id=os.environ["GDRIVE_CLIENT_ID"],
         gdrive_root_folder_id=os.environ["GDRIVE_ROOT_FOLDER_ID"],
     )
