+++ tests/tests_integration/test_search_methods.py
@@ -1,5 +1,7 @@
 from pathlib import Path
 
+import pytest
+
 from .. import test_utils
 from ..base import BaseTest
 
@@ -9,47 +11,35 @@ from ..base import BaseTest
 
 
 class TestSubSesSearches(BaseTest):
-    def test_local_vs_central_search_methods(self, project, monkeypatch):
+    # TODO: return full path
+    @pytest.mark.parametrize("return_full_path", [True, False])
+    def test_local_vs_central_search_methods(
+        self, project, monkeypatch, return_full_path
+    ):
         """ """
         central_path = project.get_central_path()
 
+        # fmt: off
         paths_to_make = []
         for i in range(1, 4):
             paths_to_make.append(Path(f"rawdata/sub-00{i}/ses-001/behav"))
-            paths_to_make.append(
-                Path(f"rawdata/sub-00{i}/ses-002_date-20250402/behav")
-            )
-            paths_to_make.append(
-                Path(f"rawdata/sub-00{i}/ses-002_date-20250402/anat")
-            )
+            paths_to_make.append(Path(f"rawdata/sub-00{i}/ses-002_date-20250402/behav"))
+            paths_to_make.append(Path(f"rawdata/sub-00{i}/ses-002_date-20250402/anat"))
 
-        paths_to_make.append(
-            Path("rawdata/sub-003_condition-test/ses-001_hello-world/ephys")
-        )
-        paths_to_make.append(
-            Path("rawdata/sub-004_condition-test/ses-002_hello-world/funcimg")
-        )
+        paths_to_make.append(Path("rawdata/sub-003_condition-test/ses-001_hello-world/ephys"))
+        paths_to_make.append(Path("rawdata/sub-004_condition-test/ses-002_hello-world/funcimg"))
 
         for path_ in paths_to_make:
             (central_path / path_).mkdir(parents=True)
-            breakpoint()
-            test_utils.write_file(
-                central_path / path_ / f"{path_.name}.md",
-                contents="hello_world",
-            )
-            test_utils.write_file(
-                central_path / path_.parent / f"{path_.parent.name}.md",
-                contents="hello_world",
-            )
-            test_utils.write_file(
-                central_path
-                / path_.parent.parent
-                / f"{path_.parent.parent.name}.md",
-                contents="hello_world",
-            )
+            test_utils.write_file(central_path / path_ / f"{path_.name}_file.md", contents="hello_world",)
+            test_utils.write_file(central_path / path_.parent / f"{path_.name}.md", contents="hello_world",)
+            test_utils.write_file(central_path / path_.parent.parent / f"{path_.parent.name}.md", contents="hello_world",)
+            test_utils.write_file(central_path / path_.parent.parent.parent / f"{path_.parent.parent.name}.md", contents="hello_world",)
+        # fmt: on
 
         from datashuttle.utils.folders import (
             search_central_via_connection,
+            search_local_filesystem,
         )
 
         # -- monkeypatch cfg.get_rclone_config to return dummy config
@@ -66,11 +56,38 @@ class TestSubSesSearches(BaseTest):
             shell=True,
         )
 
-        hello, world = search_central_via_connection(
-            project.cfg,
-            central_path / "rawdata",
-            "sub-*",
-            return_full_path=True,
-        )
+        for search_path, search_str in (
+            (central_path / "rawdata", "*"),
+            (central_path / "rawdata", "sub-*"),
+            (central_path / "rawdata" / "sub-003_condition-test", "ses-*"),
+            (central_path / "rawdata/sub-001/ses-002_date-20250402", "behav"),
+            (
+                central_path / "rawdata/sub-001/ses-002_date-20250402/behav",
+                "behav_file.md",
[…]
