+++ tests/tests_integration/test_search_methods.py
@@ -1,6 +1,12 @@
+import subprocess
 from pathlib import Path
 
 import pytest
+
+from datashuttle.utils.folders import (
+    search_central_via_connection,
+    search_local_filesystem,
+)
 
 from .. import test_utils
 from ..base import BaseTest
@@ -11,14 +17,26 @@ from ..base import BaseTest
 
 
 class TestSubSesSearches(BaseTest):
-    # TODO: return full path
     @pytest.mark.parametrize("return_full_path", [True, False])
     def test_local_vs_central_search_methods(
         self, project, monkeypatch, return_full_path
     ):
-        """ """
+        """
+        Test the `search_local_filesystem` and `search_central_via_connection`
+        functions. These functions perform the same function but `search_local_filesystem`
+        is used for local filesystem for speed. Here we check the outputs of these
+        functions match.
+
+        These functions are individually tested in many places, primarily in
+        transfer tests. Local filesystem transfer tests are very extensive,
+        because they are quicker, while central connection tests are less
+        thorough. We test the outputs of these two functions directly
+        under a range of test cases to ensure they are matched under many conditions.
+
+        """
         central_path = project.get_central_path()
 
+        # Create a set of folders and files
         # fmt: off
         paths_to_make = []
         for i in range(1, 4):
@@ -37,42 +55,36 @@ class TestSubSesSearches(BaseTest):
             test_utils.write_file(central_path / path_.parent.parent.parent / f"{path_.parent.parent.name}.md", contents="hello_world",)
         # fmt: on
 
-        from datashuttle.utils.folders import (
-            search_central_via_connection,
-            search_local_filesystem,
-        )
-
-        # -- monkeypatch cfg.get_rclone_config to return dummy config
+        # Monkeycatch `get_rclone_config_name` to return `local` and set `local`
+        # as a rclone config entry associated with the local filesystem. By this
+        # method we can hijack `search_central_via_connection` to run locally
+        # (though it is set up in practice to run via ssh, gdrive or aws).
         monkeypatch.setattr(
             project.cfg,
             "get_rclone_config_name",
             lambda connection_method: "local",
         )
 
-        import subprocess
-
         subprocess.run(
             ["rclone", "config", "create", "local", "local", "nounc", "true"],
             shell=True,
         )
 
+        # Perform a range of checks across folders and files
+        # and check the outputs of both approaches match.
+        # fmt: off
         for search_path, search_str in (
             (central_path / "rawdata", "*"),
             (central_path / "rawdata", "sub-*"),
             (central_path / "rawdata" / "sub-003_condition-test", "ses-*"),
             (central_path / "rawdata/sub-001/ses-002_date-20250402", "behav"),
-            (
-                central_path / "rawdata/sub-001/ses-002_date-20250402/behav",
-                "behav_file.md",
-            ),  # ses-002_date-20250402
-            (
-                central_path / "rawdata/sub-001/ses-002_date-20250402/behav",
-                "*",
-            ),
+            (central_path / "rawdata/sub-001/ses-002_date-20250402/behav", "behav_file.md",),
+            (central_path / "rawdata/sub-001/ses-002_date-20250402/behav", "*",),
             (central_path / "rawdata/sub-002", "*"),
             (central_path / "rawdata/sub-002/ses-002_date-20250402", "*"),
             (central_path / "rawdata", "sub-003*"),
         ):
+        # fmt: on
             central_method_folders, central_method_files = (
                 search_central_via_connection(
                     project.cfg,
