### 2025-08-06 aa226a8
Minor changes for null central path and distinct folders (#572)
* Use different folder name per tests.

* Play around with AWS see if it works.

* Edit workflow.
