### 2025-08-06 d7472a9
Try sorting outputs to fix tests.
