+++ datashuttle/utils/folders.py
@@ -660,7 +660,7 @@ def search_local_filesystem(
         elif item.is_file():
             all_filenames.append(to_append)
 
-    return all_folder_names, all_filenames
+    return sorted(all_folder_names), sorted(all_filenames)
 
 
 def search_central_via_connection(
