### 2025-08-06 e7d0c2f
fix: handle None central_path in tempfile path and correct aws bucket checking logic
