### 2025-08-06 35b5534
Sort all outputs for complete cross-platform equivalence.
