### 2025-08-06 e0d2b0b
Add warning, revert chaning copy to CTRL+Q.
