+++ tests/tests_tui/test_tui_directorytree.py
@@ -114,7 +114,7 @@ class TestTuiCreateDirectoryTree(TuiBase):
     async def test_create_folders_directorytree_clipboard(
         self, setup_project_paths
     ):
-        """Check that pressing CTRL+Q on the directorytree copies the
+        """Check that pressing CTRL+C on the directorytree copies the
         hovered folder to the clipboard (using pyperclip).
         """
         tmp_config_path, tmp_path, project_name = setup_project_paths.values()
@@ -134,7 +134,7 @@ class TestTuiCreateDirectoryTree(TuiBase):
                 pilot,
                 "#create_folders_directorytree",
                 hover_line=2,
-                press_string="ctrl+q",
+                press_string="ctrl+c",
             )
 
             pasted_path = pyperclip.paste()
@@ -178,7 +178,7 @@ class TestTuiCreateDirectoryTree(TuiBase):
                 pilot,
                 "#create_folders_directorytree",
                 hover_line=2,
-                press_string="ctrl+q",
+                press_string="ctrl+c",
             )
 
             # Check that the error message is displayed.
