### 2025-08-06 1042ba1
central_path may only be None for gdrive (or local-only mode).
