### 2025-08-06 55de02d
fix: location constraint, add: failure test for aws setup via tui
