+++ tests/tests_tui/test_tui_setup_aws.py
@@ -3,6 +3,7 @@ import os
 import pytest
 
 from datashuttle.tui.app import TuiApp
+from datashuttle.tui.screens.project_manager import ProjectManagerScreen
 
 from .tui_base import TuiBase
 
@@ -15,34 +16,13 @@ class TestTuiSetupAws(TuiBase):
         app = TuiApp()
 
         async with app.run_test(size=self.tui_size()) as pilot:
-            await self.setup_aws_project(pilot, project_name)
-
-            await self.scroll_to_click_pause(
-                pilot, "#configs_setup_connection_button"
+            await self.check_and_click_onto_existing_project(
+                pilot, project_name
             )
 
-            # Start connection setup
-            assert (
-                "Ready to setup AWS connection. Press OK to proceed"
-                in pilot.app.screen.query_one(
-                    "#setup_aws_messagebox_message"
-                ).renderable
+            await self.setup_aws_project_and_run_connection_setup(
+                pilot, os.environ["AWS_SECRET_ACCESS_KEY"]
             )
-            await self.scroll_to_click_pause(pilot, "#setup_aws_ok_button")
-
-            # Fill secret access key
-            assert (
-                "Please Enter your AWS Secret Access Key"
-                in pilot.app.screen.query_one(
-                    "#setup_aws_messagebox_message"
-                ).renderable
-            )
-            await self.fill_input(
-                pilot,
-                "#setup_aws_secret_access_key_input",
-                os.environ["AWS_SECRET_ACCESS_KEY"],
-            )
-            await self.scroll_to_click_pause(pilot, "#setup_aws_ok_button")
 
             assert (
                 "AWS Connection Successful!"
@@ -51,8 +31,64 @@ class TestTuiSetupAws(TuiBase):
                 ).renderable
             )
 
-    async def setup_aws_project(self, pilot, project_name):
-        await self.check_and_click_onto_existing_project(pilot, project_name)
+    @pytest.mark.asyncio
+    async def test_aws_connection_setup_failed(self, setup_project_paths):
+        tmp_config_path, tmp_path, project_name = setup_project_paths.values()
+
+        app = TuiApp()
+
+        async with app.run_test(size=self.tui_size()) as pilot:
+            await self.check_and_click_onto_existing_project(
+                pilot, project_name
+            )
+
+            await self.setup_aws_project_and_run_connection_setup(
+                pilot, "some-random-client-secret"
+            )
+
+            assert (
+                "AWS setup failed. Please check your configs and secret access key"
+                in pilot.app.screen.query_one(
+                    "#setup_aws_messagebox_message"
+                ).renderable
+            )
+
+    async def setup_aws_project_and_run_connection_setup(
+        self, pilot, secret_access_key
+    ):
+        await self.setup_aws_project(pilot)
+
+        await self.scroll_to_click_pause(
+            pilot, "#configs_setup_connection_button"
+        )
+
+        # Start connection setup
+        assert (
+            "Ready to setup AWS connection. Press OK to proceed"
+            in pilot.app.screen.query_one(
+                "#setup_aws_messagebox_message"
+            ).renderable
+        )
+        await self.scroll_to_click_pause(pilot, "#setup_aws_ok_button")
+
+        # Fill secret access key
+        assert (
+            "Please Enter your AWS Secret Access Key"
+            in pilot.app.screen.query_one(
+                "#setup_aws_messagebox_message"
+            ).renderable
+        )
+        await self.fill_input(
+            pilot,
+            "#setup_aws_secret_access_key_input",
+            secret_access_key,
+        )
+
+        await self.scroll_to_click_pause(pilot, "#setup_aws_ok_button")
+
[…]
