### 2025-08-06 9c97b46
Fix unecessary changes and tests, tidy up app.py
