### 2025-08-06 3463aa7
Remove assert False.
