### 2025-08-06 93218b0
Random attempt using pilot.pause().
