+++ datashuttle/utils/validation.py
@@ -148,6 +148,7 @@ def validate_list_of_names(
     prefix: Prefix,
     name_templates: Optional[Dict] = None,
     check_value_lengths: bool = True,
+    ALLOW_ALPHANUMERIC: bool = False,
 ) -> List[str]:
     """Validate a list of subject or session names against NeuroBlueprint.
 
@@ -182,7 +183,7 @@ def validate_list_of_names(
         path_, name = get_path_and_name(path_or_name)
 
         error_messages += prefix_is_duplicate_or_has_bad_values(
-            name, prefix, path_
+            name, prefix, path_, ALLOW_ALPHANUMERIC
         )
         error_messages += name_begins_with_bad_key(name, prefix, path_)
         error_messages += names_include_special_characters(name, path_)
@@ -200,14 +201,14 @@ def validate_list_of_names(
     # Note this called functions again loop over the list (O(n^2)) so
     # this is not very efficient but these lists should never be that long.
     stripped_path_or_names_list = strip_uncheckable_names(
-        path_or_name_list, prefix
+        path_or_name_list, prefix, ALLOW_ALPHANUMERIC
     )
 
     for path_or_name in stripped_path_or_names_list:
         path_, name = get_path_and_name(path_or_name)
 
         error_messages += new_name_duplicates_existing(
-            name, stripped_path_or_names_list, prefix
+            name, stripped_path_or_names_list, prefix, ALLOW_ALPHANUMERIC
         )
 
     if check_value_lengths:
@@ -219,12 +220,16 @@ def validate_list_of_names(
 
 
 def prefix_is_duplicate_or_has_bad_values(
-    name: str, prefix: Prefix, path_: Path | None
+    name: str, prefix: Prefix, path_: Path | None, ALLOW_ALPHANUMERIC: bool
 ) -> List[str]:
     """Check the sub- or ses- prefix.
 
     Ensure it is found only once in the name and
     that its value can be converted to integer.
+
+    These are tested together because we cannot test
+    for integer sub- or ses- prefix without being
+    sure it exists and is not duplicated.
 
     Parameters
     ----------
@@ -250,17 +255,21 @@ def prefix_is_duplicate_or_has_bad_values(
     if len(value) > 1:
         return [get_duplicate_prefix_error(name, prefix, path_)]
 
-    try:
-        int(value[0])
+    if ALLOW_ALPHANUMERIC:
         return []
-    except ValueError:
-        return [get_bad_value_error(name, prefix, path_)]
+    else:
+        try:
+            int(value[0])
+            return []
+        except ValueError:
+            return [get_bad_value_error(name, prefix, path_)]
 
 
 def new_name_duplicates_existing(
     new_name: str,
     existing_path_or_name_list: List[Path] | List[str],
     prefix: Prefix,
+    ALLOW_ALPHANUMERIC: bool,
 ) -> List[str]:
     """Check that a subject or session value does not duplicate an existing value.
 
@@ -286,9 +295,11 @@ def new_name_duplicates_existing(
         A list of validation errors.
 
     """
+    return_as_int = not ALLOW_ALPHANUMERIC
+
     # Make a list of matches between `new_name` and any in `existing_names`
     new_name_id = utils.get_values_from_bids_formatted_name(
-        [new_name], prefix, return_as_int=True
+        [new_name], prefix, return_as_int=return_as_int
     )[0]
 
     error_messages = []
@@ -296,7 +307,7 @@ def new_name_duplicates_existing(
         exist_path, exist_name = get_path_and_name(exist_path_or_name)
 
         exist_name_id = utils.get_values_from_bids_formatted_name(
-            [exist_name], prefix, return_as_int=True
+            [exist_name], prefix, return_as_int=return_as_int
         )[0]
 
         if exist_name_id == new_name_id:
[…]
