+++ datashuttle/tui/interface.py
@@ -125,6 +125,8 @@ class Interface:
         ]
         bypass_validation = self.tui_settings["bypass_validation"]
 
+        ALLOW_ALPHANUMERIC = self.tui_settings["ALLOW_ALPHANUMERIC"]
+
         try:
             self.project.create_folders(
                 top_level_folder,
@@ -132,6 +134,7 @@ class Interface:
                 ses_names=ses_names,
                 datatype=datatype,
                 bypass_validation=bypass_validation,
+                ALLOW_ALPHANUMERIC=ALLOW_ALPHANUMERIC,
             )
             return True, None
 
@@ -139,7 +142,10 @@ class Interface:
             return False, str(e)
 
     def validate_names(
-        self, sub_names: List[str], ses_names: Optional[List[str]]
+        self,
+        sub_names: List[str],
+        ses_names: Optional[List[str]],
+        ALLOW_ALPHANUMERIC: bool,
     ) -> InterfaceOutput:
         """Validate a list of subject / session names.
 
@@ -168,6 +174,7 @@ class Interface:
                 ses_names,
                 self.get_name_templates(),
                 bypass_validation=False,
+                ALLOW_ALPHANUMERIC=ALLOW_ALPHANUMERIC,
             )
 
             return True, {
@@ -183,6 +190,7 @@ class Interface:
         top_level_folder: list[str] | None,
         include_central: bool,
         strict_mode: bool,
+        ALLOW_ALPHANUMERIC: bool,
     ) -> tuple[bool, list[str] | str]:
         """Wrap the validate project function.
 
@@ -213,6 +221,7 @@ class Interface:
                 display_mode="print",  # unused
                 include_central=include_central,
                 strict_mode=strict_mode,
+                ALLOW_ALPHANUMERIC=ALLOW_ALPHANUMERIC,
             )
             return True, results
 
