### 2025-08-07 20cea4d
add: gdrive tui connection setup failure tests
