+++ tests/tests_tui/test_tui_setup_gdrive.py
@@ -11,14 +11,19 @@ from .tui_base import TuiBase
 class TestTuiSetupGdrive(TuiBase):
     @pytest.mark.asyncio
     async def test_gdrive_connection_setup_without_browser(
-        self, setup_project_paths, mocker
+        self, setup_project_paths
     ):
         tmp_config_path, tmp_path, project_name = setup_project_paths.values()
 
         app = TuiApp()
 
         async with app.run_test(size=self.tui_size()) as pilot:
-            await self.setup_gdrive_project(pilot, project_name)
+            await self.setup_gdrive_project(
+                pilot,
+                project_name,
+                os.environ["GDRIVE_CLIENT_ID"],
+                os.environ["GDRIVE_ROOT_FOLDER_ID"],
+            )
 
             await self.setup_gdrive_connection_via_tui(
                 pilot, with_browser=False
@@ -53,13 +58,120 @@ class TestTuiSetupGdrive(TuiBase):
             )
 
     @pytest.mark.asyncio
+    async def test_gdrive_connection_setup_incorrect_config_token(
+        self, setup_project_paths
+    ):
+        tmp_config_path, tmp_path, project_name = setup_project_paths.values()
+
+        app = TuiApp()
+
+        async with app.run_test(size=self.tui_size()) as pilot:
+            await self.setup_gdrive_project(
+                pilot,
+                project_name,
+                os.environ["GDRIVE_CLIENT_ID"],
+                os.environ["GDRIVE_ROOT_FOLDER_ID"],
+            )
+
+            await self.setup_gdrive_connection_via_tui(
+                pilot, with_browser=False
+            )
+
+            assert (
+                "Press shift+click to copy."
+                in pilot.app.screen.query_one(
+                    "#gdrive_setup_messagebox_message"
+                ).renderable
+            )
+
+            # Fill the config token
+            await self.fill_input(
+                pilot,
+                "#setup_gdrive_generic_input_box",
+                "placeholder",
+            )
+            await self.scroll_to_click_pause(
+                pilot, "#setup_gdrive_enter_button"
+            )
+
+            await test_utils.await_task_by_name_if_present(
+                "setup_gdrive_connection_without_browser_task"
+            )
+
+            assert (
+                "Google Drive setup failed. Please check your credentials"
+                in pilot.app.screen.query_one(
+                    "#gdrive_setup_messagebox_message"
+                ).renderable
+            )
+
+    @pytest.mark.asyncio
+    async def test_gdrive_connection_setup_incorrect_root_folder_id(
+        self, setup_project_paths
+    ):
+        tmp_config_path, tmp_path, project_name = setup_project_paths.values()
+
+        app = TuiApp()
+
+        async with app.run_test(size=self.tui_size()) as pilot:
+            await self.setup_gdrive_project(
+                pilot,
+                project_name,
+                os.environ["GDRIVE_CLIENT_ID"],
+                "placeholder",
+            )
+
+            await self.setup_gdrive_connection_via_tui(
+                pilot, with_browser=False
+            )
+
+            assert (
+                "Press shift+click to copy."
+                in pilot.app.screen.query_one(
+                    "#gdrive_setup_messagebox_message"
+                ).renderable
+            )
+
+            # Fill the config token
+            await self.fill_input(
+                pilot,
+                "#setup_gdrive_generic_input_box",
+                os.environ["GDRIVE_CONFIG_TOKEN"],
+            )
+            await self.scroll_to_click_pause(
+                pilot, "#setup_gdrive_enter_button"
+            )
+
+            await test_utils.await_task_by_name_if_present(
+                "setup_gdrive_connection_without_browser_task"
[…]
