### 2025-08-07 4a70bfc
Try not pausing between each click for double click.
