### 2025-08-07 4ae5bb7
add: switch connection radiobutton test
