### 2025-08-07 3ba1dcc
Try 0.5 seconds for the pilot pause.
