### 2025-08-07 61861d8
Make central function, remove pause() outside of this main use case.
