### 2025-08-07 6eb57c0
Tidy up a bit more, add back pilot pauses in some places.
