### 2025-08-08 6e494c6
Add confirm close screen.
