+++ tests/tests_integration/test_settings.py
@@ -179,6 +179,7 @@ class TestPersistentSettings(BaseTest):
                 "custom_transfer": "rawdata",
             },
             "bypass_validation": False,
+            "ALLOW_ALPHANUMERIC": False,
             "overwrite_existing_files": "never",
             "dry_run": False,
             "suggest_next_sub_ses_central": False,
@@ -219,6 +220,7 @@ class TestPersistentSettings(BaseTest):
                 "custom_transfer": "derivatives",
             },
             "bypass_validation": True,
+            "ALLOW_ALPHANUMERIC": True,
             "overwrite_existing_files": "always",
             "dry_run": True,
             "suggest_next_sub_ses_central": True,
