+++ datashuttle/datashuttle_class.py
@@ -124,7 +124,6 @@ class DataShuttle:
         ses_names: Optional[Union[str, List[str]]] = None,
         datatype: Union[str, List[str]] = "",
         bypass_validation: bool = False,
-        ALLOW_ALPHANUMERIC: bool = False,
         log: bool = True,
     ) -> Dict[str, List[Path]]:
         """Create a folder tree in the project folder.
@@ -159,12 +158,6 @@ class DataShuttle:
         bypass_validation
             If `True`, folders will be created even if they are not
             valid to NeuroBlueprint style.
-
-        ALLOW_ALPHANUMERIC
-            If `False`, non-integer sub- or ses- labels will raise an error, and duplicate
-            value checks include checks on the numerical value(e.g. sub-01 and sub-001_date-20240101 are
-            considered duplicate). If `True`, alphanumeric labels (e.g. sub-abc) will not raise an error
-            and duplicate label checks are only based on alphanumeric values.
 
         log
             If `True`, details of folder creation will be logged.
@@ -224,7 +217,6 @@ class DataShuttle:
             ses_names,
             name_templates,
             bypass_validation,
-            ALLOW_ALPHANUMERIC=ALLOW_ALPHANUMERIC,
             log=True,
         )
 
@@ -261,25 +253,16 @@ class DataShuttle:
         ses_names: Optional[Union[str, List[str]]],
         name_templates: Dict,
         bypass_validation: bool,
-        ALLOW_ALPHANUMERIC: bool = False,
         log: bool = True,
     ) -> Tuple[List[str], List[str]]:
         """Central method to format and validate subject and session names."""
         format_sub = formatting.check_and_format_names(
-            sub_names,
-            "sub",
-            name_templates,
-            bypass_validation,
-            ALLOW_ALPHANUMERIC=ALLOW_ALPHANUMERIC,
+            sub_names, "sub", name_templates, bypass_validation
         )
 
         if ses_names is not None:
             format_ses = formatting.check_and_format_names(
-                ses_names,
-                "ses",
-                name_templates,
-                bypass_validation,
-                ALLOW_ALPHANUMERIC=ALLOW_ALPHANUMERIC,
+                ses_names, "ses", name_templates, bypass_validation
             )
         else:
             format_ses = []
@@ -294,7 +277,6 @@ class DataShuttle:
                 display_mode="error",
                 log=log,
                 name_templates=name_templates,
-                ALLOW_ALPHANUMERIC=ALLOW_ALPHANUMERIC,
             )
 
         return format_sub, format_ses
@@ -1295,7 +1277,6 @@ class DataShuttle:
         display_mode: DisplayMode,
         include_central: bool = False,
         strict_mode: bool = False,
-        ALLOW_ALPHANUMERIC: bool = False,
     ) -> List[str]:
         """Perform validation on the project.
 
@@ -1324,12 +1305,6 @@ class DataShuttle:
             starting with sub- or ses- prefix are checked. In ``Strict Mode``,
             any folder not prefixed with sub-, ses- or a valid datatype will
             raise a validation issue.
-
-        ALLOW_ALPHANUMERIC
-            If `False`, non-integer sub- or ses- labels will raise an error, and duplicate
-            value checks include checks on the numerical value(e.g. sub-01 and sub-001_date-20240101 are
-            considered duplicate). If `True`, alphanumeric labels (e.g. sub-abc) will not raise an error
-            and duplicate label checks are only based on alphanumeric values.
 
         Returns
         -------
@@ -1369,7 +1344,6 @@ class DataShuttle:
             display_mode=display_mode,
             name_templates=name_templates,
             strict_mode=strict_mode,
[…]
