### 2025-08-08 65da92c
extend: update tui config test for gdrive and aws
