+++ tests/tests_tui/test_tui_configs.py
@@ -2,6 +2,7 @@ import copy
 from pathlib import Path
 from time import monotonic
 from typing import List
+from uuid import uuid4
 
 import pytest
 from textual.widget import Widget
@@ -200,66 +201,113 @@ class TestTuiConfigs(TuiBase):
             # Make sure they are all different to the existing configs,
             # then save and check the configs on the DataShuttle instance
             # and file are updated.
-            local_path = tmp_path / f"some-random-path/{project_name}"
-            central_path = tmp_path / f"some-random-path2/{project_name}"
-
-            local_path.mkdir(parents=True)
-            central_path.mkdir(parents=True)
-
+            
             new_kwargs = {
-                "local_path": local_path.as_posix(),
-                "central_path": central_path.as_posix(),
+                "local_path": self.make_and_get_random_project_path(tmp_path, project_name),
+                "central_path": self.make_and_get_random_project_path(tmp_path, project_name),
                 "connection_method": "ssh",
                 "central_host_id": "random_host",
                 "central_host_username": "random_username",
             }
-
-            for key in new_kwargs:
-                # The purpose is to update to completely new configs
-                assert new_kwargs[key] != project_cfg[key]
-
-            await self.set_configs_content_widgets(pilot, new_kwargs)
-
-            await self.check_configs_widgets_match_configs(
-                configs_content, new_kwargs
+            await self.edit_configs_and_check_widgets(
+                pilot, 
+                tmp_config_path, 
+                project_name, 
+                new_kwargs, 
+                project_cfg
             )
 
-            await self.scroll_to_click_pause(
-                pilot,
-                "#configs_save_configs_button",
-            )
-            assert (
-                pilot.app.screen.query_one(
-                    "#messagebox_message_label"
-                ).renderable
-                == "Configs saved."
-            )
-            await self.close_messagebox(pilot)
-
-            test_utils.check_configs(
-                pilot.app.screen.interface.project,
-                new_kwargs,
-                tmp_config_path / project_name / "config.yaml",
+            project_cfg = copy.deepcopy(pilot.app.screen.interface.project.cfg)
+            new_kwargs = {
+                "local_path": self.make_and_get_random_project_path(tmp_path, project_name),
+                "central_path": self.make_and_get_random_project_path(tmp_path, project_name),
+                "connection_method": "gdrive",
+                "gdrive_root_folder_id": "random-folder-id",
+                "gdrive_client_id": "random-client-id",
+            }
+            await self.edit_configs_and_check_widgets(
+                pilot, 
+                tmp_config_path, 
+                project_name, 
+                new_kwargs, 
+                project_cfg
             )
 
-            # Finally, use "Main Menu" button to go back to the home screen,
-            # navigate back to the project and check the new configs are now
-            # displayed.
-            await self.scroll_to_click_pause(pilot, "#all_main_menu_buttons")
-            assert pilot.app.screen.id == "_default"
+            project_cfg = copy.deepcopy(pilot.app.screen.interface.project.cfg)
+            new_kwargs = {
+                "local_path": self.make_and_get_random_project_path(tmp_path, project_name),
+                "central_path": self.make_and_get_random_project_path(tmp_path, project_name),
+                "connection_method": "aws",
+                "aws_access_key_id": "random-access-key-id",
+                "aws_region": "us-east-1"
+            }
+            await self.edit_configs_and_check_widgets(
+                pilot, 
+                tmp_config_path, 
+                project_name, 
+                new_kwargs, 
+                project_cfg
+            )
 
-            await self.check_and_click_onto_existing_project(
-                pilot, project_name
-            )
[…]
