+++ tests/tests_tui/test_tui_configs.py
@@ -201,61 +201,61 @@ class TestTuiConfigs(TuiBase):
             # Make sure they are all different to the existing configs,
             # then save and check the configs on the DataShuttle instance
             # and file are updated.
-            
+
             new_kwargs = {
-                "local_path": self.make_and_get_random_project_path(tmp_path, project_name),
-                "central_path": self.make_and_get_random_project_path(tmp_path, project_name),
+                "local_path": self.make_and_get_random_project_path(
+                    tmp_path, project_name
+                ),
+                "central_path": self.make_and_get_random_project_path(
+                    tmp_path, project_name
+                ),
                 "connection_method": "ssh",
                 "central_host_id": "random_host",
                 "central_host_username": "random_username",
             }
             await self.edit_configs_and_check_widgets(
-                pilot, 
-                tmp_config_path, 
-                project_name, 
-                new_kwargs, 
-                project_cfg
+                pilot, tmp_config_path, project_name, new_kwargs, project_cfg
             )
 
             project_cfg = copy.deepcopy(pilot.app.screen.interface.project.cfg)
             new_kwargs = {
-                "local_path": self.make_and_get_random_project_path(tmp_path, project_name),
-                "central_path": self.make_and_get_random_project_path(tmp_path, project_name),
+                "local_path": self.make_and_get_random_project_path(
+                    tmp_path, project_name
+                ),
+                "central_path": self.make_and_get_random_project_path(
+                    tmp_path, project_name
+                ),
                 "connection_method": "gdrive",
                 "gdrive_root_folder_id": "random-folder-id",
                 "gdrive_client_id": "random-client-id",
             }
             await self.edit_configs_and_check_widgets(
-                pilot, 
-                tmp_config_path, 
-                project_name, 
-                new_kwargs, 
-                project_cfg
+                pilot, tmp_config_path, project_name, new_kwargs, project_cfg
             )
 
             project_cfg = copy.deepcopy(pilot.app.screen.interface.project.cfg)
             new_kwargs = {
-                "local_path": self.make_and_get_random_project_path(tmp_path, project_name),
-                "central_path": self.make_and_get_random_project_path(tmp_path, project_name),
+                "local_path": self.make_and_get_random_project_path(
+                    tmp_path, project_name
+                ),
+                "central_path": self.make_and_get_random_project_path(
+                    tmp_path, project_name
+                ),
                 "connection_method": "aws",
                 "aws_access_key_id": "random-access-key-id",
-                "aws_region": "us-east-1"
+                "aws_region": "us-east-1",
             }
             await self.edit_configs_and_check_widgets(
-                pilot, 
-                tmp_config_path, 
-                project_name, 
-                new_kwargs, 
-                project_cfg
+                pilot, tmp_config_path, project_name, new_kwargs, project_cfg
             )
 
     async def edit_configs_and_check_widgets(
-        self, 
-        pilot, 
-        tmp_config_path, 
-        project_name, 
-        new_kwargs, 
-        prev_project_cfg
+        self,
+        pilot,
+        tmp_config_path,
+        project_name,
+        new_kwargs,
+        prev_project_cfg,
     ):
         for key in new_kwargs:
             # The purpose is to update to completely new configs
@@ -276,9 +276,7 @@ class TestTuiConfigs(TuiBase):
             "#configs_save_configs_button",
         )
         assert (
-            pilot.app.screen.query_one(
-                "#messagebox_message_label"
-            ).renderable
[…]
