### 2025-08-08 ef86f32
add: backwards compability tests for config.yaml
