### 2025-08-08 eebaf34
Small fixes, add first tests.
