### 2025-08-08 c387581
Small fixes, add first tests.
