+++ datashuttle/datashuttle_class.py
@@ -124,6 +124,7 @@ class DataShuttle:
         ses_names: Optional[Union[str, List[str]]] = None,
         datatype: Union[str, List[str]] = "",
         bypass_validation: bool = False,
+        ALLOW_ALPHANUMERIC: bool = False,
         log: bool = True,
     ) -> Dict[str, List[Path]]:
         """Create a folder tree in the project folder.
@@ -158,6 +159,13 @@ class DataShuttle:
         bypass_validation
             If `True`, folders will be created even if they are not
             valid to NeuroBlueprint style.
+
+        ALLOW_ALPHANUMERIC
+            If `True`, any alphanumeric character are allowed for sub- or ses- labels. Otherwise,
+            labels must be integer and the following additional checks are performed:
+                - Identical numbers are considered the same value even if padded with different number of zeros
+                    (e.g. sub-01 and sub-001_date-20240101 are considered duplicate).
+                - Labels must be the same length (e.g. sub-01 and sub-002 is invalid).
 
         log
             If `True`, details of folder creation will be logged.
@@ -217,6 +225,7 @@ class DataShuttle:
             ses_names,
             name_templates,
             bypass_validation,
+            ALLOW_ALPHANUMERIC,
             log=True,
         )
 
@@ -253,16 +262,25 @@ class DataShuttle:
         ses_names: Optional[Union[str, List[str]]],
         name_templates: Dict,
         bypass_validation: bool,
+        ALLOW_ALPHANUMERIC: bool,
         log: bool = True,
     ) -> Tuple[List[str], List[str]]:
         """Central method to format and validate subject and session names."""
         format_sub = formatting.check_and_format_names(
-            sub_names, "sub", name_templates, bypass_validation
+            sub_names,
+            "sub",
+            name_templates,
+            bypass_validation,
+            ALLOW_ALPHANUMERIC,
         )
 
         if ses_names is not None:
             format_ses = formatting.check_and_format_names(
-                ses_names, "ses", name_templates, bypass_validation
+                ses_names,
+                "ses",
+                name_templates,
+                bypass_validation,
+                ALLOW_ALPHANUMERIC,
             )
         else:
             format_ses = []
@@ -274,6 +292,7 @@ class DataShuttle:
                 format_sub,
                 format_ses,
                 include_central=False,
+                ALLOW_ALPHANUMERIC=ALLOW_ALPHANUMERIC,  # TODO: CHECK ORDER
                 display_mode="error",
                 log=log,
                 name_templates=name_templates,
@@ -1277,6 +1296,7 @@ class DataShuttle:
         display_mode: DisplayMode,
         include_central: bool = False,
         strict_mode: bool = False,
+        ALLOW_ALPHANUMERIC: bool = False,
     ) -> List[str]:
         """Perform validation on the project.
 
@@ -1305,6 +1325,13 @@ class DataShuttle:
             starting with sub- or ses- prefix are checked. In ``Strict Mode``,
             any folder not prefixed with sub-, ses- or a valid datatype will
             raise a validation issue.
+
+        ALLOW_ALPHANUMERIC
+            If `True`, any alphanumeric character are allowed for sub- or ses- labels. Otherwise,
+            labels must be integer and the following additional checks are performed:
+                - Identical numbers are considered the same value even if padded with different number of zeros
+                    (e.g. sub-01 and sub-001_date-20240101 are considered duplicate).
+                - Labels must be the same length (e.g. sub-01 and sub-002 is invalid).
 
         Returns
         -------
@@ -1344,6 +1371,7 @@ class DataShuttle:
             display_mode=display_mode,
             name_templates=name_templates,
[…]
