### 2025-08-08 4d02006
Small fixes, add first tests.
