+++ tests/tests_tui/test_tui_widgets_and_defaults.py
@@ -420,6 +420,20 @@ class TestTuiWidgets(TuiBase):
                 is False
             )
 
+            # ALLOW_ALPHANUMERIC
+            assert (
+                pilot.app.screen.query_one(
+                    "#create_folders_ALLOW_ALPHANUMERIC_checkbox"
+                ).label._text
+                == "ALLOW_ALPHANUMERIC"
+            )
+            assert (
+                pilot.app.screen.query_one(
+                    "#create_folders_ALLOW_ALPHANUMERIC_checkbox"
+                ).value
+                is False
+            )
+
             # Template validation
             assert (
                 pilot.app.screen.query_one(
@@ -654,11 +668,25 @@ class TestTuiWidgets(TuiBase):
             await pilot.pause()
 
     @pytest.mark.asyncio
-    async def test_bypass_validation_settings(self, setup_project_paths):
-        """Test all configs that underly the 'bypass validation'
+    @pytest.mark.parametrize(
+        "parameter_name", ["bypass_validation", "ALLOW_ALPHANUMERIC"]
+    )
+    async def test_create_folderes_validation_settings(
+        self, setup_project_paths, parameter_name
+    ):
+        """Test all configs that underly the 'bypass validation' and `ALLOW_ALPHANUMERIC`
         setting are updated correctly by the widget.
+
+        These two options are similar and are both default off, we so we can test
+        them using the same process (turning them on and checking the underlying
+        parameters in the stored dictionaries are also changed as expected).
         """
         tmp_config_path, tmp_path, project_name = setup_project_paths.values()
+
+        if parameter_name == "bypass_validation":
+            checkbox_id = "#create_folders_settings_bypass_validation_checkbox"
+        else:
+            checkbox_id = "#create_folders_ALLOW_ALPHANUMERIC_checkbox"
 
         app = TuiApp()
         async with app.run_test(size=self.tui_size()) as pilot:
@@ -669,30 +697,17 @@ class TestTuiWidgets(TuiBase):
                 pilot, "#create_folders_settings_button"
             )
 
+            assert pilot.app.screen.query_one(checkbox_id).value is False
             assert (
-                pilot.app.screen.query_one(
-                    "#create_folders_settings_bypass_validation_checkbox"
-                ).value
-                is False
-            )
-            assert (
-                pilot.app.screen.interface.tui_settings["bypass_validation"]
+                pilot.app.screen.interface.tui_settings[parameter_name]
                 is False
             )
 
-            await self.scroll_to_click_pause(
-                pilot, "#create_folders_settings_bypass_validation_checkbox"
-            )
+            await self.scroll_to_click_pause(pilot, checkbox_id)
 
+            assert pilot.app.screen.query_one(checkbox_id).value is True
             assert (
-                pilot.app.screen.query_one(
-                    "#create_folders_settings_bypass_validation_checkbox"
-                ).value
-                is True
-            )
-            assert (
-                pilot.app.screen.interface.tui_settings["bypass_validation"]
-                is True
+                pilot.app.screen.interface.tui_settings[parameter_name] is True
             )
 
             await self.scroll_to_click_pause(
@@ -705,15 +720,9 @@ class TestTuiWidgets(TuiBase):
                 pilot, "#create_folders_settings_button"
             )
 
+            assert pilot.app.screen.query_one(checkbox_id).value is True
             assert (
-                pilot.app.screen.query_one(
-                    "#create_folders_settings_bypass_validation_checkbox"
-                ).value
-                is True
-            )
-            assert (
-                pilot.app.screen.interface.tui_settings["bypass_validation"]
-                is True
+                pilot.app.screen.interface.tui_settings[parameter_name] is True
             )
 
             await pilot.pause()
