+++ tests/tests_tui/test_tui_validate.py
@@ -104,6 +104,7 @@ class TestTuiValidate(TuiBase):
                 "ses": None,
             }
             assert kwargs_["strict_mode"] is False
+            assert kwargs_["ALLOW_ALPHANUMERIC"] is False
 
             # Then, change all arguments and check these are
             # changed at the level of the called function.
@@ -116,6 +117,10 @@ class TestTuiValidate(TuiBase):
                 pilot, "#validate_include_central_checkbox"
             )
             await self.change_checkbox(pilot, "#validate_strict_mode_checkbox")
+
+            await self.change_checkbox(
+                pilot, "#validate_ALLOW_ALPHANUMERIC_checkbox"
+            )
 
             await self.scroll_to_click_pause(
                 pilot, "#validate_validate_button"
@@ -146,6 +151,7 @@ class TestTuiValidate(TuiBase):
                 "ses": None,
             }
             assert kwargs_["strict_mode"] is False
+            assert kwargs_["ALLOW_ALPHANUMERIC"] is True
 
             # Check the widgets are hidden as expected.
             # Path widgets are not shown for Transfer tab
@@ -176,6 +182,10 @@ class TestTuiValidate(TuiBase):
             )
             await self.fill_input(pilot, "#validate_path_input", project_path)
 
+            await self.change_checkbox(
+                pilot, "#validate_ALLOW_ALPHANUMERIC_checkbox"
+            )
+
             # Spy the function and click 'validate' button
             spy_validate = mocker.spy(
                 datashuttle.tui.shared.validate_content,
@@ -194,6 +204,7 @@ class TestTuiValidate(TuiBase):
             assert args_[0] == project_path
             assert kwargs_["top_level_folder"] == "rawdata"
             assert kwargs_["strict_mode"] is False
+            assert kwargs_["ALLOW_ALPHANUMERIC"] is True
 
             # Check removed widgets, this should be removed because always local
             with pytest.raises(textual.css.query.InvalidQueryFormat):
