### 2025-08-14 9391982
Updated test cases to use basemodel and also added more test cases to cover all edges
