+++ tests/test_date_search_range.py
@@ -1,217 +1,340 @@
-import glob
 import os
 import re
 import shutil
-import tempfile
-from pathlib import Path
-from typing import List
-
 import pytest
+from pathlib import Path
 
-from datashuttle.utils.folders import search_with_tags
+from datashuttle.configs import canonical_tags
 
-
-# Dummy implementation for canonical_tags
-class DummyCanonicalTags:
-    @staticmethod
-    def tags(x: str) -> str:
-        tags_dict = {
-            "*": "@*@",
-            "DATETO": "@DATETO@",
-            "TIMETO": "@TIMETO@",
-            "DATETIMETO": "@DATETIMETO@",
-        }
-        return tags_dict.get(x, x)
-
-    @staticmethod
-    def get_datetime_format(format_type: str) -> str:
-        formats = {
-            "datetime": "%Y%m%dT%H%M%S",
-            "time": "%H%M%S",
-            "date": "%Y%m%d",
-        }
-        if format_type not in formats:
-            raise ValueError(f"Invalid format type: {format_type}")
-        return formats[format_type]
+from . import test_utils
+from .base import BaseTest
 
 
-# Patch canonical_tags
-@pytest.fixture(autouse=True)
-def patch_canonical_tags(monkeypatch):
-    from datashuttle.configs import canonical_tags
+class TestDateSearchRange(BaseTest):
+    """Test date/time range search functionality with real datashuttle projects."""
 
-    monkeypatch.setattr(canonical_tags, "tags", DummyCanonicalTags.tags)
-    monkeypatch.setattr(
-        canonical_tags,
-        "get_datetime_format",
-        DummyCanonicalTags.get_datetime_format,
-    )
+    def test_simple_wildcard_first(self, project):
+        """Test basic wildcard functionality before testing date ranges."""
+        subs = ["sub-001", "sub-002"] 
+        sessions = ["ses-001", "ses-002"]
+        
+        datatypes_used = test_utils.get_all_broad_folders_used(value=False)
+        datatypes_used.update({"behav": True})
+        test_utils.make_and_check_local_project_folders(
+            project, "rawdata", subs, sessions, ["behav"], datatypes_used
+        )
+        
+        project.upload_custom(
+            "rawdata",
+            sub_names=[f"sub-{canonical_tags.tags('*')}"],
+            ses_names=[f"ses-{canonical_tags.tags('*')}"],
+            datatype=["behav"],
+        )
+        
+        central_path = project.get_central_path() / "rawdata"
+        transferred_subs = [sub.name for sub in central_path.glob("sub-*")]
+        
+        expected_subs = ["sub-001", "sub-002"]
+        assert sorted(transferred_subs) == sorted(expected_subs)
+        
+        for sub_name in expected_subs:
+            sub_path = central_path / sub_name
+            transferred_sessions = [ses.name for ses in sub_path.glob("ses-*")]
+            expected_sessions = ["ses-001", "ses-002"]
+            assert sorted(transferred_sessions) == sorted(expected_sessions)
 
+    def test_date_range_transfer(self, project):
+        """Test that date range patterns correctly filter folders during transfer."""
+        subs = ["sub-001", "sub-002"]
+        sessions = [
+            "ses-001_date-20240301",
+            "ses-002_date-20240315", 
+            "ses-003_date-20240401",
+            "ses-004_date-20240415",
+            "ses-005_date-20240501",
+        ]
+        
+        datatypes_used = test_utils.get_all_broad_folders_used(value=False)
+        datatypes_used.update({"behav": True, "ephys": True})
+        test_utils.make_and_check_local_project_folders(
+            project, "rawdata", subs, sessions, ["behav", "ephys"], datatypes_used
+        )
+        
+        project.upload_custom(
[…]
