### 2025-08-14 23679eb
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
