+++ tests/test_date_search_range.py
@@ -1,8 +1,7 @@
 import os
-import re
 import shutil
+
 import pytest
-from pathlib import Path
 
 from datashuttle.configs import canonical_tags
 
@@ -15,28 +14,28 @@ class TestDateSearchRange(BaseTest):
 
     def test_simple_wildcard_first(self, project):
         """Test basic wildcard functionality before testing date ranges."""
-        subs = ["sub-001", "sub-002"] 
+        subs = ["sub-001", "sub-002"]
         sessions = ["ses-001", "ses-002"]
-        
+
         datatypes_used = test_utils.get_all_broad_folders_used(value=False)
         datatypes_used.update({"behav": True})
         test_utils.make_and_check_local_project_folders(
             project, "rawdata", subs, sessions, ["behav"], datatypes_used
         )
-        
+
         project.upload_custom(
             "rawdata",
             sub_names=[f"sub-{canonical_tags.tags('*')}"],
             ses_names=[f"ses-{canonical_tags.tags('*')}"],
             datatype=["behav"],
         )
-        
+
         central_path = project.get_central_path() / "rawdata"
         transferred_subs = [sub.name for sub in central_path.glob("sub-*")]
-        
+
         expected_subs = ["sub-001", "sub-002"]
         assert sorted(transferred_subs) == sorted(expected_subs)
-        
+
         for sub_name in expected_subs:
             sub_path = central_path / sub_name
             transferred_sessions = [ses.name for ses in sub_path.glob("ses-*")]
@@ -48,33 +47,43 @@ class TestDateSearchRange(BaseTest):
         subs = ["sub-001", "sub-002"]
         sessions = [
             "ses-001_date-20240301",
-            "ses-002_date-20240315", 
+            "ses-002_date-20240315",
             "ses-003_date-20240401",
             "ses-004_date-20240415",
             "ses-005_date-20240501",
         ]
-        
+
         datatypes_used = test_utils.get_all_broad_folders_used(value=False)
         datatypes_used.update({"behav": True, "ephys": True})
         test_utils.make_and_check_local_project_folders(
-            project, "rawdata", subs, sessions, ["behav", "ephys"], datatypes_used
+            project,
+            "rawdata",
+            subs,
+            sessions,
+            ["behav", "ephys"],
+            datatypes_used,
         )
-        
+
         project.upload_custom(
             "rawdata",
             sub_names=subs,
-            ses_names=[f"ses-{canonical_tags.tags('*')}_20240315{canonical_tags.tags('DATETO')}20240401"],
+            ses_names=[
+                f"ses-{canonical_tags.tags('*')}_20240315{canonical_tags.tags('DATETO')}20240401"
+            ],
             datatype=["behav", "ephys"],
         )
-        
+
         central_path = project.get_central_path() / "rawdata"
         transferred_subs = list(central_path.glob("sub-*"))
-        
+
         assert len(transferred_subs) == 2
-        
+
         for sub_path in transferred_subs:
             transferred_sessions = [ses.name for ses in sub_path.glob("ses-*")]
-            expected_sessions = ["ses-002_date-20240315", "ses-003_date-20240401"]
+            expected_sessions = [
+                "ses-002_date-20240315",
+                "ses-003_date-20240401",
+            ]
             assert sorted(transferred_sessions) == sorted(expected_sessions)
 
     def test_time_range_transfer(self, project):
@@ -82,27 +91,29 @@ class TestDateSearchRange(BaseTest):
         subs = ["sub-001"]
         sessions = [
             "ses-001_time-080000",
-            "ses-002_time-120000",  
+            "ses-002_time-120000",
             "ses-003_time-160000",
             "ses-004_time-200000",
         ]
-        
+
[…]
