### 2025-08-14 044f82c
Revert handling None case in get_aws_bucket_name
