### 2025-08-16 5abcf90
add: docs for setting up google drive and aws connection
