### 2025-08-17 a285311
add: helpful links
