### 2025-08-18 7c27837
Attempt to fix random CI test errors (#574)
* Random attempt using pilot.pause().

* Also clear inputs.

* Try await task every time it is used.

* Try more pauses.

* Pause for 2 seconds and add after fill input too.

* Try not pausing between each click for double click.

* TO REVERT: only run key tests.

* Try 0.5 seconds for the pilot pause.

* Make central function, remove pause() outside of this main use case.

* Some tidyng up.

* Tidy up a bit more, add back pilot pauses in some places.

* Finaly tidy up and revert workflow changes.

* Fix description.
