### 2025-08-20 e05ea59
Apply suggestions from code review
Co-authored-by: Joe Ziminski <55797454+JoeZiminski@users.noreply.github.com>
