### 2025-08-20 3ac2057
fix: env in tui tests
