### 2025-08-20 c26fa87
use inbuilt rclone recursive search
