+++ tests/test_utils.py
@@ -422,40 +422,29 @@ def check_config_file(config_path, *kwargs):
 
 def recursive_search_central(project: DataShuttle):
     """
-    A convenience function to recursively search a
-    project for files on remote folders using `rclone
-    lsjson`
+    A convenience function to search  project for files on remote folders
+    using rclone's recursive search.
     """
     all_filenames: list[str] = []
-    rclone_recursive_search(
-        project,
-        (project.cfg["central_path"] / "rawdata").as_posix(),
-        all_filenames,
-    )
 
-    return all_filenames
+    path_ = (project.cfg["central_path"] / "rawdata").as_posix()
 
-
-def rclone_recursive_search(
-    project: DataShuttle, path_: str, all_filenames: list[str]
-):
+    # -R flag searches recursively
     output = rclone.call_rclone(
-        f"lsjson {project.cfg.get_rclone_config_name()}:{path_}", pipe_std=True
+        f"lsjson -R {project.cfg.get_rclone_config_name()}:{path_}",
+        pipe_std=True,
     )
 
     files_or_folders = json.loads(output.stdout)
 
     for file_or_folder in files_or_folders:
-        name = file_or_folder["Name"]
         is_dir = file_or_folder.get("IsDir", False)
-        if is_dir:
-            rclone_recursive_search(
-                project,
-                path_ + "/" + name,
-                all_filenames,
-            )
-        else:
-            all_filenames.append(path_ + "/" + name)
+
+        if not is_dir:
+            file_path = file_or_folder["Path"]
+            all_filenames.append(path_ + "/" + file_path)
+
+    return all_filenames
 
 
 # -----------------------------------------------------------------------------
