### 2025-08-20 030d11f
add: docstrings to backwards compatiblity tests
