### 2025-08-20 b963119
use rclone recursive search in ssh tests
