+++ tests/tests_transfers/ssh/ssh_test_utils.py
@@ -1,10 +1,7 @@
 import builtins
 import copy
-import stat
 import subprocess
 import sys
-
-import paramiko
 
 from datashuttle.utils import rclone, ssh, utils
 
@@ -69,48 +66,6 @@ def setup_ssh_connection(project, setup_ssh_key_pair=True):
     )
 
     return verified
-
-
-def recursive_search_central(project):
-    """
-    A convenience function to recursively search a
-    project for files through SSH, used  during testing
-    across an SSH connection to collected names of
-    files that were transferred.
-    """
-    with paramiko.SSHClient() as client:
-        ssh.connect_client_core(client, project.cfg)
-
-        sftp = client.open_sftp()
-
-        all_filenames = []
-
-        sftp_recursive_file_search(
-            sftp,
-            (project.cfg["central_path"] / "rawdata").as_posix(),
-            all_filenames,
-        )
-    return all_filenames
-
-
-def sftp_recursive_file_search(sftp, path_, all_filenames):
-    """Append all filenames found within a folder,
-    when searching over a sftp connection.
-    """
-    try:
-        sftp.stat(path_)
-    except FileNotFoundError:
-        return
-
-    for file_or_folder in sftp.listdir_attr(path_):
-        if stat.S_ISDIR(file_or_folder.st_mode):
-            sftp_recursive_file_search(
-                sftp,
-                path_ + "/" + file_or_folder.filename,
-                all_filenames,
-            )
-        else:
-            all_filenames.append(path_ + "/" + file_or_folder.filename)
 
 
 def docker_is_running():
