### 2025-08-20 78572de
use central function for ssh wildcard tests
