+++ tests/tests_transfers/ssh/test_ssh_transfer.py
@@ -1,13 +1,8 @@
 import fnmatch
 import platform
-import shutil
 
-import paramiko
 import pytest
 
-from datashuttle.utils import ssh
-
-from ... import test_utils
 from . import ssh_test_utils
 from .base_ssh import BaseSSHTransfer
 
@@ -71,8 +66,12 @@ class TestSSHTransfer(BaseSSHTransfer):
         """
         pathtable, project = ssh_setup
 
+        expected_transferred_paths = self.get_expected_transferred_paths(
+            pathtable, sub_names, ses_names, datatype
+        )
+
         self.run_and_check_transfers(
-            project, pathtable, sub_names, ses_names, datatype
+            project, sub_names, ses_names, datatype, expected_transferred_paths
         )
 
     # Test Wildcards
@@ -102,7 +101,7 @@ class TestSSHTransfer(BaseSSHTransfer):
 
         expected_transferred_paths = pathtable["path"]
 
-        self.check_transfer_ssh(
+        self.run_and_check_transfers(
             project, sub_names, ses_names, datatype, expected_transferred_paths
         )
 
@@ -126,7 +125,7 @@ class TestSSHTransfer(BaseSSHTransfer):
 
         expected_transferred_paths = pathtable["path"]
 
-        self.check_transfer_ssh(
+        self.run_and_check_transfers(
             project, sub_names, ses_names, datatype, expected_transferred_paths
         )
 
@@ -155,95 +154,6 @@ class TestSSHTransfer(BaseSSHTransfer):
 
         expected_transferred_paths = pathtable["path"]
 
-        self.check_transfer_ssh(
+        self.run_and_check_transfers(
             project, sub_names, ses_names, datatype, expected_transferred_paths
         )
-
-    def check_transfer_ssh(
-        self,
-        project,
-        sub_names,
-        ses_names,
-        datatype,
-        expected_transferred_paths,
-    ):
-        """Transfer the data and check the transferred files match the
-        `expected_transferred_paths`.
-
-        In test setup, the entire project is created in the `local_path` and
-        is uploaded to `central_path`. So we only need to set up once per test,
-        upload and download is to temporary folders and these temporary folders
-        are cleaned at the end of each parameterization.
-        """
-        # Upload data from the setup local project to a temporary
-        # central directory.
-        true_central_path = project.cfg["central_path"]
-        tmp_central_path = (
-            project.cfg["central_path"] / "tmp" / project.project_name
-        )
-        self.remake_logging_path(project)
-
-        project.update_config_file(central_path=tmp_central_path)
-
-        project.upload_custom(
-            "rawdata", sub_names, ses_names, datatype, init_log=False
-        )
-
-        # Search the paths that were transferred and tidy them up,
-        # then check against the paths that were expected to be transferred.
-        transferred_files = test_utils.recursive_search_central(project)
-        paths_to_transferred_files = self.remove_path_before_rawdata(
-            transferred_files
-        )
-
-        assert sorted(paths_to_transferred_files) == sorted(
-            expected_transferred_paths
-        )
-
-        # Now, move data from the central path where the project is
-        # setup, to a temp local folder to test download.
-        true_local_path = project.cfg["local_path"]
-        tmp_local_path = (
-            project.cfg["local_path"] / "tmp" / project.project_name
-        )
-        tmp_local_path.mkdir(exist_ok=True, parents=True)
-
-        project.update_config_file(local_path=tmp_local_path)
-        project.update_config_file(central_path=true_central_path)
-
-        project.download_custom(
-            "rawdata", sub_names, ses_names, datatype, init_log=False
-        )
-
-        # Find the transferred paths, tidy them up
-        # and check expected paths were transferred.
[…]
