+++ tests/tests_transfers/gdrive/test_gdrive_transfer.py
@@ -47,6 +47,10 @@ class TestGdriveTransfer(BaseTransfer):
     ):
         pathtable, project = gdrive_setup
 
+        expected_transferred_paths = self.get_expected_transferred_paths(
+            pathtable, sub_names, ses_names, datatype
+        )
+
         self.run_and_check_transfers(
-            project, pathtable, sub_names, ses_names, datatype
+            project, sub_names, ses_names, datatype, expected_transferred_paths
         )
