### 2025-08-20 7b25c64
fix: CI error on deleting temp.txt
