+++ datashuttle/utils/rclone.py
@@ -335,10 +335,12 @@ def check_successful_connection_and_raise_error_on_fail(cfg: Configs) -> None:
     If the command fails, it raises a ConnectionError. The created file is
     deleted thereafter.
     """
+    filename = f"{utils.get_random_string()}_temp.txt"
+
     if cfg["central_path"] is None:
-        tempfile_path = "temp.txt"
+        tempfile_path = filename
     else:
-        tempfile_path = (cfg["central_path"] / "temp.txt").as_posix()
+        tempfile_path = (cfg["central_path"] / filename).as_posix()
 
     output = call_rclone(
         f"touch {cfg.get_rclone_config_name()}:{tempfile_path}", pipe_std=True
