### 2025-08-20 9fcbc38
add: docstrings to transfer tests
