+++ tests/tests_transfers/gdrive/gdrive_test_utils.py
@@ -7,8 +7,13 @@ from datashuttle.utils import gdrive, utils
 
 
 def setup_project_for_gdrive(project: DataShuttle):
-    """
-    Set up a project with configs for Google Drive transfers.
+    """Set up a project with configs for Google Drive transfers.
+
+    The connection credentials are fetched from the environment which
+    the developer shall set themselves to test locally. In the CI, these
+    are set using the github secrets. A random string is added to the
+    central path so that the test project paths do not interfere while
+    running multiple test instances simultaneously in CI.
     """
     random_string = utils.get_random_string()
 
@@ -24,6 +29,11 @@ def setup_gdrive_connection(project: DataShuttle):
     """
     Convenience function to set up the Google Drive connection by
     mocking user input.
+
+    The mock input is triggered twice. First, to deny the presence of
+    a browser. Second, to enter a `GDRIVE_CONFIG_TOKEN` needed to set up
+    connection without a browser. The credentials are set in the environment
+    by the CI. To run tests locally, the developer must set them themselves.
     """
     state = {"first": True}
 
