+++ tests/tests_transfers/aws/aws_test_utils.py
@@ -6,6 +6,14 @@ from datashuttle.utils import aws, utils
 
 
 def setup_project_for_aws(project: DataShuttle):
+    """Update the config file for an AWS connection.
+
+    The connection credentials are fetched from the environment which
+    the developer shall set themselves to test locally. In the CI, these
+    are set using the github secrets. A random string is added to the
+    central path so that the test project paths do not interfere while
+    running multiple test instances simultaneously in CI.
+    """
     aws_bucket_name = os.environ["AWS_BUCKET_NAME"]
 
     random_string = utils.get_random_string()
@@ -22,6 +30,9 @@ def setup_aws_connection(project: DataShuttle):
     """
     Convenience function to set up the AWS connection by
     mocking the `aws.get_aws_secret_access_key` function.
+
+    The `AWS_SECRET_ACCESS_KEY` is set in the environment by the CI while
+    testing. For testing locally, the developer must set it themselves.
     """
     original_get_secret = copy.deepcopy(aws.get_aws_secret_access_key)
     aws.get_aws_secret_access_key = lambda *args, **kwargs: os.environ[
