### 2025-08-20 dd80989
add: docstrings to tui tests
