### 2025-08-20 1a4ae15
add: docstrings to unit test
