### 2025-08-21 64f96dc
Apply suggestions from code review
Co-authored-by: Joe Ziminski <55797454+JoeZiminski@users.noreply.github.com>
