### 2025-08-21 8bcec21
Minor changes to docstrings.
