+++ datashuttle/tui/interface.py
@@ -130,7 +130,9 @@ class Interface:
         ]
         bypass_validation = self.tui_settings["bypass_validation"]
 
-        ALLOW_ALPHANUMERIC = self.tui_settings["ALLOW_ALPHANUMERIC"]
+        allow_alphanumeric_sub_ses_values = self.tui_settings[
+            "allow_alphanumeric_sub_ses_values"
+        ]
 
         try:
             self.project.create_folders(
@@ -139,7 +141,7 @@ class Interface:
                 ses_names=ses_names,
                 datatype=datatype,
                 bypass_validation=bypass_validation,
-                ALLOW_ALPHANUMERIC=ALLOW_ALPHANUMERIC,
+                allow_alphanumeric_sub_ses_values=allow_alphanumeric_sub_ses_values,
             )
             return True, None
 
@@ -150,7 +152,7 @@ class Interface:
         self,
         sub_names: List[str],
         ses_names: Optional[List[str]],
-        ALLOW_ALPHANUMERIC: bool,
+        allow_alphanumeric_sub_ses_values: bool,
     ) -> InterfaceOutput:
         """Validate a list of subject / session names.
 
@@ -167,7 +169,7 @@ class Interface:
         ses_names
             List of session names to format.
 
-        ALLOW_ALPHANUMERIC
+        allow_alphanumeric_sub_ses_values
             If `True`, alphanumeric values will not raise an error.
 
         """
@@ -182,7 +184,7 @@ class Interface:
                 ses_names,
                 self.get_name_templates(),
                 bypass_validation=False,
-                ALLOW_ALPHANUMERIC=ALLOW_ALPHANUMERIC,
+                allow_alphanumeric_sub_ses_values=allow_alphanumeric_sub_ses_values,
             )
 
             return True, {
@@ -198,7 +200,7 @@ class Interface:
         top_level_folder: list[str] | None,
         include_central: bool,
         strict_mode: bool,
-        ALLOW_ALPHANUMERIC: bool,
+        allow_alphanumeric_sub_ses_values: bool,
     ) -> tuple[bool, list[str] | str]:
         """Wrap the validate project function.
 
@@ -213,7 +215,7 @@ class Interface:
             If `True`, the central project is also validated.
         strict_mode
             If `True`, validation will be run in strict mode.
-        ALLOW_ALPHANUMERIC
+        allow_alphanumeric_sub_ses_values
             If `True`, alphanumeric values will not raise an error.
 
         Returns
@@ -231,7 +233,7 @@ class Interface:
                 display_mode="print",  # unused
                 include_central=include_central,
                 strict_mode=strict_mode,
-                ALLOW_ALPHANUMERIC=ALLOW_ALPHANUMERIC,
+                allow_alphanumeric_sub_ses_values=allow_alphanumeric_sub_ses_values,
             )
             return True, results
 
