+++ datashuttle/tui/tooltips.py
@@ -135,8 +135,8 @@ def get_tooltip(id: str) -> str:
             "Allow folder creation even when there is a validation error."
         )
 
-    elif id == "#create_folders_ALLOW_ALPHANUMERIC_checkbox":
-        tooltip = "ALLOW_ALPHANUMERIC"
+    elif id == "#create_folders_allow_alphanumeric_sub_ses_values_checkbox":
+        tooltip = "allow_alphanumeric_sub_ses_values"
 
     # template validation checkbox
     elif id == "#template_settings_validation_on_checkbox":
