+++ datashuttle/datashuttle_class.py
@@ -121,7 +121,7 @@ class DataShuttle:
         ses_names: Optional[Union[str, List[str]]] = None,
         datatype: Union[str, List[str]] = "",
         bypass_validation: bool = False,
-        ALLOW_ALPHANUMERIC: bool = False,
+        allow_alphanumeric_sub_ses_values: bool = False,
         log: bool = True,
     ) -> Dict[str, List[Path]]:
         """Create a folder tree in the project folder.
@@ -157,7 +157,7 @@ class DataShuttle:
             If `True`, folders will be created even if they are not
             valid to NeuroBlueprint style.
 
-        ALLOW_ALPHANUMERIC
+        allow_alphanumeric_sub_ses_values
             If `True`, any alphanumeric character are allowed for sub- or ses- labels. Otherwise,
             labels must be integer and the following additional checks are performed:
                 - Identical numbers are considered the same value even if padded with different number of zeros
@@ -222,7 +222,7 @@ class DataShuttle:
             ses_names,
             name_templates,
             bypass_validation,
-            ALLOW_ALPHANUMERIC,
+            allow_alphanumeric_sub_ses_values,
             log=True,
         )
 
@@ -259,7 +259,7 @@ class DataShuttle:
         ses_names: Optional[Union[str, List[str]]],
         name_templates: Dict,
         bypass_validation: bool,
-        ALLOW_ALPHANUMERIC: bool,
+        allow_alphanumeric_sub_ses_values: bool,
         log: bool = True,
     ) -> Tuple[List[str], List[str]]:
         """Central method to format and validate subject and session names."""
@@ -268,7 +268,7 @@ class DataShuttle:
             "sub",
             name_templates,
             bypass_validation,
-            ALLOW_ALPHANUMERIC,
+            allow_alphanumeric_sub_ses_values,
         )
 
         if ses_names is not None:
@@ -277,7 +277,7 @@ class DataShuttle:
                 "ses",
                 name_templates,
                 bypass_validation,
-                ALLOW_ALPHANUMERIC,
+                allow_alphanumeric_sub_ses_values,
             )
         else:
             format_ses = []
@@ -289,7 +289,7 @@ class DataShuttle:
                 format_sub,
                 format_ses,
                 include_central=False,
-                ALLOW_ALPHANUMERIC=ALLOW_ALPHANUMERIC,  # TODO: CHECK ORDER
+                allow_alphanumeric_sub_ses_values=allow_alphanumeric_sub_ses_values,  # TODO: CHECK ORDER
                 display_mode="error",
                 log=log,
                 name_templates=name_templates,
@@ -1168,7 +1168,7 @@ class DataShuttle:
         display_mode: DisplayMode,
         include_central: bool = False,
         strict_mode: bool = False,
-        ALLOW_ALPHANUMERIC: bool = False,
+        allow_alphanumeric_sub_ses_values: bool = False,
     ) -> List[str]:
         """Perform validation on the project.
 
@@ -1198,7 +1198,7 @@ class DataShuttle:
             any folder not prefixed with sub-, ses- or a valid datatype will
             raise a validation issue.
 
-        ALLOW_ALPHANUMERIC
+        allow_alphanumeric_sub_ses_values
             If `True`, any alphanumeric character are allowed for sub- or ses- labels. Otherwise,
             labels must be integer and the following additional checks are performed:
                 - Identical numbers are considered the same value even if padded with different number of zeros
@@ -1243,7 +1243,7 @@ class DataShuttle:
             display_mode=display_mode,
             name_templates=name_templates,
             strict_mode=strict_mode,
[…]
