### 2025-08-21 d21674c
Rename placeholder variable.
