+++ datashuttle/utils/validation.py
@@ -42,10 +42,15 @@ def get_missing_prefix_error(name: str, prefix, path_: Path | None) -> str:
 
 
 def get_bad_value_error(
-    name: str, prefix, path_: Path | None, ALLOW_ALPHANUMERIC: bool
+    name: str,
+    prefix,
+    path_: Path | None,
+    allow_alphanumeric_sub_ses_values: bool,
 ) -> str:
     """Return error message when the value for a prefix is not an integer."""
-    type_ = "alphanumeric" if ALLOW_ALPHANUMERIC else "an integer"
+    type_ = (
+        "alphanumeric" if allow_alphanumeric_sub_ses_values else "an integer"
+    )
 
     return handle_path(
         f"BAD_VALUE: The value for prefix {prefix} in name {name} is not {type_}.",
@@ -152,7 +157,7 @@ def validate_list_of_names(
     prefix: Prefix,
     name_templates: Optional[Dict] = None,
     check_value_lengths: bool = True,
-    ALLOW_ALPHANUMERIC: bool = False,
+    allow_alphanumeric_sub_ses_values: bool = False,
 ) -> List[str]:
     """Validate a list of subject or session names against NeuroBlueprint.
 
@@ -171,9 +176,10 @@ def validate_list_of_names(
         If `True`, check that the prefix-<value> value lengths
         are consistent across the passed list.
 
-    ALLOW_ALPHANUMERIC
-        If `True`, any alphanumeric character are allowed for sub- or ses- labels. Otherwise,
-        labels must be integer and the following additional checks are performed:
+    allow_alphanumeric_sub_ses_values
+        If `True`, any alphanumeric character are allowed for the values associated
+        with sub- or ses-  keys. Otherwise, values must be integer
+        and the following additional checks are performed:
             - Identical numbers are considered the same value even if padded with different number of zeros
                 (e.g. sub-01 and sub-001_date-20240101 are considered duplicate).
             - Labels must be the same length (e.g. sub-01 and sub-002 is invalid).
@@ -194,7 +200,7 @@ def validate_list_of_names(
         path_, name = get_path_and_name(path_or_name)
 
         error_messages += prefix_is_duplicate_or_has_bad_values(
-            name, prefix, path_, ALLOW_ALPHANUMERIC
+            name, prefix, path_, allow_alphanumeric_sub_ses_values
         )
         error_messages += name_begins_with_bad_key(name, prefix, path_)
         error_messages += names_include_special_characters(name, path_)
@@ -212,17 +218,20 @@ def validate_list_of_names(
     # Note this called functions again loop over the list (O(n^2)) so
     # this is not very efficient but these lists should never be that long.
     stripped_path_or_names_list = strip_uncheckable_names(
-        path_or_name_list, prefix, ALLOW_ALPHANUMERIC
+        path_or_name_list, prefix, allow_alphanumeric_sub_ses_values
     )
 
     for path_or_name in stripped_path_or_names_list:
         path_, name = get_path_and_name(path_or_name)
 
         error_messages += new_name_duplicates_existing(
-            name, stripped_path_or_names_list, prefix, ALLOW_ALPHANUMERIC
+            name,
+            stripped_path_or_names_list,
+            prefix,
+            allow_alphanumeric_sub_ses_values,
         )
 
-    if not ALLOW_ALPHANUMERIC and check_value_lengths:
+    if not allow_alphanumeric_sub_ses_values and check_value_lengths:
         error_messages += value_lengths_are_inconsistent(
             stripped_path_or_names_list, prefix
         )
@@ -231,7 +240,10 @@ def validate_list_of_names(
 
 
 def prefix_is_duplicate_or_has_bad_values(
[…]
