+++ tests/tests_tui/test_tui_widgets_and_defaults.py
@@ -420,16 +420,16 @@ class TestTuiWidgets(TuiBase):
                 is False
             )
 
-            # ALLOW_ALPHANUMERIC
+            # allow_alphanumeric_sub_ses_values
             assert (
                 pilot.app.screen.query_one(
-                    "#create_folders_ALLOW_ALPHANUMERIC_checkbox"
+                    "#create_folders_allow_alphanumeric_sub_ses_values_checkbox"
                 ).label._text
-                == "ALLOW_ALPHANUMERIC"
+                == "allow_alphanumeric_sub_ses_values"
             )
             assert (
                 pilot.app.screen.query_one(
-                    "#create_folders_ALLOW_ALPHANUMERIC_checkbox"
+                    "#create_folders_allow_alphanumeric_sub_ses_values_checkbox"
                 ).value
                 is False
             )
@@ -669,12 +669,13 @@ class TestTuiWidgets(TuiBase):
 
     @pytest.mark.asyncio
     @pytest.mark.parametrize(
-        "parameter_name", ["bypass_validation", "ALLOW_ALPHANUMERIC"]
+        "parameter_name",
+        ["bypass_validation", "allow_alphanumeric_sub_ses_values"],
     )
     async def test_create_folderes_validation_settings(
         self, setup_project_paths, parameter_name
     ):
-        """Test all configs that underly the 'bypass validation' and `ALLOW_ALPHANUMERIC`
+        """Test all configs that underly the 'bypass validation' and `allow_alphanumeric_sub_ses_values`
         setting are updated correctly by the widget.
 
         These two options are similar and are both default off, we so we can test
@@ -686,7 +687,9 @@ class TestTuiWidgets(TuiBase):
         if parameter_name == "bypass_validation":
             checkbox_id = "#create_folders_settings_bypass_validation_checkbox"
         else:
-            checkbox_id = "#create_folders_ALLOW_ALPHANUMERIC_checkbox"
+            checkbox_id = (
+                "#create_folders_allow_alphanumeric_sub_ses_values_checkbox"
+            )
 
         app = TuiApp()
         async with app.run_test(size=self.tui_size()) as pilot:
