+++ datashuttle/configs/canonical_configs.py
@@ -279,7 +279,7 @@ def get_tui_config_defaults() -> Dict:
             "overwrite_existing_files": "never",
             "dry_run": False,
             "suggest_next_sub_ses_central": False,
-            "ALLOW_ALPHANUMERIC": False,
+            "allow_alphanumeric_sub_ses_values": False,
         }
     }
 

[…]
