### 2025-08-21 c16cf68
Add tooltips.
