### 2025-08-21 1ccf582
remove: backwards compatibility test for incorrect configs
