+++ tests/tests_regression/old_version_configs/v0.6.0_bad/config.yaml
@@ -1,7 +1,0 @@
-local_path: old_ver
-central_path: old_ver
-connection_method: gdrive
-central_host_id: null
-central_host_username: null
-gdrive_client_id: null
-gdrive_root_folder_id: null
