### 2025-08-21 fb0938b
Tidying up TUI, trying to get create settings page nice.
