### 2025-08-21 f8f93ee
Tidying up TUI, trying to get create settings page nice.
