+++ tests/tests_tui/test_tui_widgets_and_defaults.py
@@ -419,13 +419,13 @@ class TestTuiWidgets(TuiBase):
             # allow_alphanumeric_sub_ses_values
             assert (
                 pilot.app.screen.query_one(
-                    "#create_folders_allow_alphanumeric_sub_ses_values_checkbox"
+                    "#create_folders_settings_allow_alphanumeric_checkbox"
                 ).label._text
                 == "allow_alphanumeric_sub_ses_values"
             )
             assert (
                 pilot.app.screen.query_one(
-                    "#create_folders_allow_alphanumeric_sub_ses_values_checkbox"
+                    "#create_folders_settings_allow_alphanumeric_checkbox"
                 ).value
                 is False
             )
@@ -684,7 +684,7 @@ class TestTuiWidgets(TuiBase):
             checkbox_id = "#create_folders_settings_bypass_validation_checkbox"
         else:
             checkbox_id = (
-                "#create_folders_allow_alphanumeric_sub_ses_values_checkbox"
+                "#create_folders_settings_allow_alphanumeric_checkbox"
             )
 
         app = TuiApp()
