### 2025-08-21 54efb6e
Tidying up TUI, trying to get create settings page nice.
