### 2025-08-21 c8edfdb
add: make new project configs tests for gdrive and aws
