+++ tests/tests_tui/test_tui_configs.py
@@ -1,25 +1,19 @@
 import copy
 from pathlib import Path
 from time import monotonic
-from typing import List
-from uuid import uuid4
 
 import pytest
-from textual.widget import Widget
 
 from datashuttle.configs import load_configs
-from datashuttle.configs.canonical_configs import get_connection_methods_list
 from datashuttle.tui.app import TuiApp
 from datashuttle.tui.screens.modal_dialogs import (
     SelectDirectoryTreeScreen,
 )
-from datashuttle.tui.screens.project_manager import ProjectManagerScreen
 
-from .. import test_utils
-from .tui_base import TuiBase
+from .tui_configs_base import TuiConfigsBase
 
 
-class TestTuiConfigs(TuiBase):
+class TestTuiConfigs(TuiConfigsBase):
     # -------------------------------------------------------------------------
     # Test New Project Configs
     # -------------------------------------------------------------------------
@@ -71,87 +65,13 @@ class TestTuiConfigs(TuiBase):
 
         app = TuiApp()
         async with app.run_test(size=self.tui_size()) as pilot:
-            # Select a new project, check NewProjectScreen is
-            # displayed correctly.
-            await self.scroll_to_click_pause(
-                pilot, "#mainwindow_new_project_button"
-            )
-
-            # Get the ConfigsContent and check all configs are displayed
-            # correctly. `check_new_project_configs` checks empty defaults
-            # are displayed, then updates with the kwargs and checks.
-            configs_content = pilot.app.screen.query_one(
-                "#new_project_configs_content"
-            )
-
-            await self.check_new_project_configs(
-                pilot, project_name, configs_content, kwargs
-            )
-
-            # Save the configs and check the correct messages are shown.
-            await self.scroll_to_click_pause(
+            await self.run_and_test_new_project_configs(
                 pilot,
-                "#configs_save_configs_button",
+                project_name,
+                tmp_config_path,
+                connection_method_name="SSH" if kwargs_set == 2 else "",
+                config_kwargs=kwargs,
             )
-
-            # if SSH is set, then the config window remains up and the
-            # 'setup ssh' button is enabled. Otherwise, the screen
-            # will automatically move to the project page.
-            if kwargs["connection_method"] == "ssh":
-                assert (
-                    pilot.app.screen.query_one(
-                        "#messagebox_message_label"
-                    ).renderable
-                    == "A datashuttle project has now been created.\n\n Next, "
-                    "setup the SSH connection. Once complete, navigate to "
-                    "the 'Main Menu' and proceed to the project page, "
-                    "where you will be able to create and transfer "
-                    "project folders."
-                )
-                await self.close_messagebox(pilot)
-                assert (
-                    pilot.app.screen.query_one(
-                        "#configs_setup_connection_button"
-                    ).label
-                    == "Setup SSH Connection"
-                )
-            else:
-                assert (
-                    pilot.app.screen.query_one(
-                        "#messagebox_message_label"
-                    ).renderable
-                    == "A datashuttle project has now been created.\n\n "
-                    "Next proceed to the project page, where you will "
-                    "be able to create and transfer project folders."
-                )
-                await self.close_messagebox(pilot)
-
-            assert (
-                pilot.app.screen.query_one(
-                    "#configs_go_to_project_screen_button"
-                ).visible
-                is True
-            )
-            await self.scroll_to_click_pause(
-                pilot, "#configs_go_to_project_screen_button"
-            )
-            assert isinstance(pilot.app.screen, ProjectManagerScreen)
-
-            project = pilot.app.screen.interface.project
-
-            assert (
-                pilot.app.screen.interface.project.project_name == project_name
[…]
