+++ tests/tests_tui/tui_configs_base.py
@@ -1,0 +1,389 @@
+from pathlib import Path
+from typing import List
+from uuid import uuid4
+
+from textual.widget import Widget
+
+from datashuttle.configs.canonical_configs import get_connection_methods_list
+from datashuttle.tui.screens.project_manager import ProjectManagerScreen
+from datashuttle.utils import utils
+
+from .. import test_utils
+from .tui_base import TuiBase
+
+
+class TuiConfigsBase(TuiBase):
+    """Contains helper functions for TUI tests of the configs tab."""
+
+    async def run_and_test_new_project_configs(
+        self,
+        pilot,
+        project_name,
+        tmp_config_path,
+        connection_method_name,
+        config_kwargs,
+    ):
+        """Check the ConfigsContent widgets are displayed correctly are displayed
+        correctly when making a new project. Fill these widgets with `config_kwargs`,
+        save them, and check the interface.project and saved configs match the new
+        settings.
+        """
+        assert pilot.app.screen.id == "_default"
+
+        # Select a new project, check NewProjectScreen is
+        # displayed correctly.
+        await self.scroll_to_click_pause(
+            pilot, "#mainwindow_new_project_button"
+        )
+
+        # Get the ConfigsContent and check all configs are displayed
+        # correctly. `check_new_project_configs` checks empty defaults
+        # are displayed, then updates with the kwargs and checks.
+        configs_content = pilot.app.screen.query_one(
+            "#new_project_configs_content"
+        )
+
+        await self.check_new_project_configs(
+            pilot, project_name, configs_content, config_kwargs
+        )
+
+        # Save the configs and check the correct messages are shown.
+        await self.scroll_to_click_pause(
+            pilot,
+            "#configs_save_configs_button",
+        )
+
+        if config_kwargs["connection_method"] == "local_filesystem":
+            assert (
+                pilot.app.screen.query_one(
+                    "#messagebox_message_label"
+                ).renderable
+                == "A datashuttle project has now been created.\n\n "
+                "Next proceed to the project page, where you will "
+                "be able to create and transfer project folders."
+            )
+            await self.close_messagebox(pilot)
+        else:
+            assert (
+                pilot.app.screen.query_one(
+                    "#messagebox_message_label"
+                ).renderable
+                == utils.get_project_created_message_template().format(
+                    method_name=connection_method_name
+                )
+            )
+
+            await self.close_messagebox(pilot)
+            assert (
+                pilot.app.screen.query_one(
+                    "#configs_setup_connection_button"
+                ).label
+                == f"Setup {connection_method_name} Connection"
+            )
+
+        assert (
+            pilot.app.screen.query_one(
+                "#configs_go_to_project_screen_button"
+            ).visible
+            is True
+        )
+        await self.scroll_to_click_pause(
+            pilot, "#configs_go_to_project_screen_button"
+        )
+        assert isinstance(pilot.app.screen, ProjectManagerScreen)
+
+        project = pilot.app.screen.interface.project
+
+        assert pilot.app.screen.interface.project.project_name == project_name
+
+        # After saving, check all configs are correct on the DataShuttle
+        # instance as well as the stored configs.
+        test_utils.check_configs(
+            project,
+            config_kwargs,
+            tmp_config_path / project_name / "config.yaml",
+        )
+
+        await pilot.pause()
+
+    async def check_new_project_configs(
+        self, pilot, project_name, configs_content, kwargs
+    ):
+        """Check the configs displayed on the TUI match those found in `kwargs`.
+        Also, check the widgets unique to ConfigsContent on the
+        configs selection for a new project.
+        """
+        # Project Name --------------------------------------------------------
+
[…]
