### 2025-08-21 c2e88dd
add: test for None central path in gdrive
