### 2025-08-21 b1c6623
add: central path fixtures to tui tests and tweak central path for transfer tests
