+++ tests/tests_tui/gdrive/test_tui_setup_gdrive.py
@@ -4,16 +4,30 @@ import pytest
 
 from datashuttle.tui.app import TuiApp
 from datashuttle.tui.screens.project_manager import ProjectManagerScreen
+from datashuttle.utils import rclone, utils
 
 from ... import test_utils
 from ..tui_base import TuiBase
 
 
 class TestTuiSetupGdrive(TuiBase):
+    @pytest.fixture(scope="function")
+    def central_path_and_project(self, setup_project_paths):
+        tmp_config_path, tmp_path, project_name = setup_project_paths.values()
+
+        random_prefix = utils.get_random_string()
+        central_path = f"{random_prefix}/{tmp_path.as_posix().lstrip('/')}"
+
+        yield central_path, project_name
+
+        rclone.call_rclone(
+            f"purge central_{project_name}_gdrive:{random_prefix}"
+        )
+
     @pytest.mark.parametrize("central_path_none", [True, False])
     @pytest.mark.asyncio
     async def test_gdrive_connection_setup_without_browser(
-        self, central_path_none, setup_project_paths
+        self, central_path_none, central_path_and_project
     ):
         """Test Google Drive connection setup via the TUI.
 
@@ -23,7 +37,7 @@ class TestTuiSetupGdrive(TuiBase):
         setup is tested without a browser. The credentials in the environment are
         set by the CI. For testing locally, the developer must set these themselves.
         """
-        tmp_config_path, tmp_path, project_name = setup_project_paths.values()
+        central_path, project_name = central_path_and_project
 
         app = TuiApp()
 
@@ -33,7 +47,7 @@ class TestTuiSetupGdrive(TuiBase):
                 project_name,
                 os.environ["GDRIVE_CLIENT_ID"],
                 os.environ["GDRIVE_ROOT_FOLDER_ID"],
-                central_path_none=central_path_none,
+                central_path=central_path if not central_path_none else "",
             )
 
             await self.setup_gdrive_connection_via_tui(
@@ -230,7 +244,7 @@ class TestTuiSetupGdrive(TuiBase):
         project_name,
         gdrive_client_id,
         root_folder_id,
-        central_path_none: bool = False,
+        central_path: str = "",
     ):
         """Navigate to the configs tab, fill in the Google Drive config credentials and save them."""
 
@@ -246,9 +260,9 @@ class TestTuiSetupGdrive(TuiBase):
             pilot, "#configs_gdrive_root_folder_id_input", root_folder_id
         )
 
-        if central_path_none:
-            pilot.app.screen.query_one("#configs_central_path_input").clear()
-            await pilot.pause()
+        await self.fill_input(
+            pilot, "#configs_central_path_input", central_path
+        )
 
         await self.scroll_to_click_pause(pilot, "#configs_save_configs_button")
         await self.close_messagebox(pilot)
