### 2025-08-21 c29b1dc
fix: ci error on windows
