### 2025-08-22 2cb6f38
Apply suggestions from code review
Co-authored-by: Joe Ziminski <55797454+JoeZiminski@users.noreply.github.com>
