### 2025-08-22 8df7948
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
