### 2025-08-22 13d20b3
add extra check to ssh setup and move message template function to tui_utils.py
