+++ docs/source/pages/get_started/set-up-a-project.md
@@ -109,11 +109,12 @@ If you would also like to transfer files to a central machine, see the next sect
 ## Set up a project for transfer
 
 Above, we have set up a ``datashuttle`` project by providing the **project name**
-and **local path**. Transfer across the local filesystem or via SSH is supported.
-Therefore, we will need to provide:
+and **local path**. To set up a project for transfer, we need to provide
+additional information:
 
-1) **central path**: location of the project on the central storage machine.
-2) Connection-specific settings (e.g. if using SSH).
+1) **central path**: location of the project on the central storage.
+2) Connection-specific settings (e.g. if using a mounted drive, SSH, Google Drive or
+Amazon Web Services (AWS)).
 
 ```{image} /_static/datashuttle-overview-dark.png
    :align: center
@@ -128,10 +129,12 @@ Therefore, we will need to provide:
 <br>
 
 How the **central path** is set depends on whether your connection to
-central storage is a
-[mounted drive](new-project-mounted-drive)
-or via
-[SSH](new-project-ssh).
+central storage is connected to using a
+[mounted drive](new-project-mounted-drive), via
+[SSH](new-project-ssh) or is an
+[AWS S3 Bucket](new-project-gdrive)
+or
+[Google Drive](new-project-aws).
 
 If you are unsure of your connection method, speak to your lab administrator
 or IT department.
@@ -238,17 +241,19 @@ project.make_config_file(
 (new-project-ssh)=
 ### Connecting to central storage through SSH
 
-Another common method of connecting to a central storage machine is via
-[SSH](https://www.ssh.com/academy/ssh/protocol).
+A common method of connecting to a central storage machine is using [SSH](https://www.ssh.com/academy/ssh/protocol).
 
-Following details are required in the configs before setting up SSH connection:
+Following details must be set in the project configs prior to setting up the connection:
 
 1) **central_host_id:** This is the address of the server you want to connect to.
 
 2) **central_host_username:** This is your profile username on the server you want to
 connect to.
 
-3) **central path**: This is the path to the project *on the server*.
+3) **central_path**: This is the path to the project *on the server*.
+
+Once the configs are saved, we can set up the connection by clicking `Set Up SSH Connection`
+(through the TUI) or running the function [](setup_ssh_connection) in Python.
 
 :::{dropdown} SSH Example
 :color: info
@@ -275,9 +280,6 @@ the **project name**, it will be automatically included.
 Note that Linux-based shortcuts (e.g. `~` for home directory) are not permitted.
 
 ::::
-
-Once the configs are saved, we can set up the connection by clicking Set Up SSH Connection
-(through the TUI) or running the function `project.setup_ssh_connection()`.
 
 ::::{tab-set}
 
@@ -334,7 +336,7 @@ Next, a one-time command to set up the SSH connection must be run:
 project.setup_ssh_connection()
 ```
 
-Running `setup_ssh_connection()` will require verification
+Running [](setup_ssh_connection) will require verification
 that the SSH server connected to is correct (pressing `y` to proceed).
 
 Finally, your password to the central server will be requested (you will
@@ -346,28 +348,32 @@ only need to do this once).
 (new-project-gdrive)=
 ### Connecting to central storage through Google Drive
 
-Another common method of connecting to a central storage machine is via
-[Google Drive](https://drive.google.com).
+To transfer data to a [Google Drive](https://drive.google.com),
+the following details must be set in the project configs prior to setting up the connection:
 
-Following details are required in the configs before setting up Google Drive connection:
+1) **gdrive_root_folder_id:** This is the Google Drive folder ID of the project folder.
+It is the alphanumeric code in the Google Drive URL when located in the folder (after `/folders/`).
 
[…]
