+++ tests/tests_tui/test_tui_widgets_and_defaults.py
@@ -393,7 +393,7 @@ class TestTuiWidgets(TuiBase):
                 pilot.app.screen.query_one(
                     "#suggest_next_sub_ses_central_checkbox"
                 ).label._text
-                == "Search Central For Suggestions"
+                == "Search central for suggestions"
             )
             assert (
                 pilot.app.screen.query_one(
@@ -416,12 +416,11 @@ class TestTuiWidgets(TuiBase):
                 is False
             )
 
-            # allow_alphanumeric_sub_ses_values
             assert (
                 pilot.app.screen.query_one(
                     "#create_folders_settings_allow_alphanumeric_checkbox"
                 ).label._text
-                == "allow_alphanumeric_sub_ses_values"
+                == "Allow alphanumeric sub- and ses- values"
             )
             assert (
                 pilot.app.screen.query_one(
