+++ docs/source/pages/get_started/set-up-a-project.md
@@ -2,13 +2,12 @@
 # Set up a project
 
 The first section of this guide will
-set up a "local-only" project that can manage creation
-and validation of project folders. This requires
+set up a "local-only" project that can create
+and validate project folders. This requires
 only minimal configuration to get started.
 
 To see how a datashuttle project can be set up for transfer,
-visit [Set up a project for transfer](set-up-a-project-for-transfer) section.
-
+visit the [Set up a project for transfer](set-up-a-project-for-transfer) section.
 
 ::::{tab-set}
 
@@ -18,8 +17,8 @@ visit [Set up a project for transfer](set-up-a-project-for-transfer) section.
 Selecting `Make New Project` will take you to the project set up screen.
 
 Enter the name of your project, the path to your project folder and
-select `No connection (local only)` (note that the central-path option
-is now disabled).
+select `No connection (local only)` (note that the `Central Path` option
+will be disabled).
 
 ```{image} /_static/screenshots/how-to-make-local-project-configs-dark.png
    :align: center
@@ -75,26 +74,11 @@ project.make_config_file(
 
 ```
 \
-The project is now ready for use, and in future can be instantiated only
+The project is now ready for use, and in future can be instantiated
 with the line ``project = DataShuttle("my_project_name")`` (i.e. you will not
 have to set the `local_path` again).
 
 If you wish to change the project settings at a later time, use ``project.update_config_file()``.
-
-For example, it is possible to immediately validate the project (if it already exists):
-
-```python
-project.validate_project("rawdata", error_or_warn="warn")
-```
-
-Setting ``error_or_warn`` will display all validation issues, otherwise
-it will error on the first one encountered.
-
-New project folders can also be created in the local folder:
-
-```python
-project.create_folders("rawdata", "sub-001", "ses-001_@DATE@", datatype=["ephys", "behav"])
-```
 
 :::
 ::::
@@ -103,7 +87,7 @@ Now, this project is ready for creating and validating
 folders to the [NeuroBlueprint](https://neuroblueprint.neuroinformatics.dev/latest/index.html) standard. See [create folders](how-to-create-folders)
 and [validate folders](tutorial-validation) for details.
 
-If you would also like to transfer files to a central machine, see the next section.
+If you would also like to transfer files to a central storage, see the next section.
 
 (set-up-a-project-for-transfer)=
 ## Set up a project for transfer
@@ -129,12 +113,12 @@ Amazon Web Services (AWS)).
 <br>
 
 How the **central path** is set depends on whether your connection to
-central storage is connected to using a
+central storage is connected to through a
 [mounted drive](new-project-mounted-drive), via
 [SSH](new-project-ssh) or is an
-[AWS S3 Bucket](new-project-gdrive)
+[AWS S3 Bucket](new-project-aws)
 or
-[Google Drive](new-project-aws).
+[Google Drive](new-project-gdrive).
 
 If you are unsure of your connection method, speak to your lab administrator
 or IT department.
@@ -156,12 +140,12 @@ Imagine your central storage is a remote server that is mounted to
 your machine at `X:\username`. You want your project folder to
 be located at `X:\username\my_projects`.
 
-In this case, you can set the **central_path** to `X:\username\my_projects`
-and with **connection_method** to **local filesystem**.
+In this case, you can set the **central path** to `X:\username\my_projects`
+and with **connection method** to **local filesystem**.
 
 The project folder will be located
 at `X:\username\my_projects\my_project_name`.
-You may pass the local or central path without the **project name**,
+You may pass the central path without the **project name**,
 (it will be automatically included).
 
 :::
@@ -192,20 +176,14 @@ Setting up ``datashuttle`` is as simple as entering the `Project name`,
 `Local Path` and `Central Path` into the relevant input boxes.
 
[…]
