### 2025-08-25 967d5a0
ADD: Tests for Google Drive and AWS Connection Methods (#570)
* add: gdrive and aws transfer tests

* add: tui tests for setting up gdrive and aws connections

* add: setup env variables from github secrets in the workflow file

* Minor changes for null central path and distinct folders (#572)

* Use different folder name per tests.

* Play around with AWS see if it works.

* Edit workflow.

* fix: handle None central_path in tempfile path and correct aws bucket checking logic

* fix: location constraint, add: failure test for aws setup via tui

* minor change

* add: gdrive tui connection set
[…]
