### 2025-08-25 a0b4122
Additional newline in tooltip.
