### 2025-08-25 0ee1851
Uncapitalise datashuttle, rename AWS S3 label to AWS S3 Bucket.
