+++ datashuttle/tui/screens/setup_gdrive.py
@@ -138,7 +138,7 @@ class SetupGdriveScreen(ModalScreen):
     def ask_user_for_browser(self) -> None:
         """Ask the user if their machine has access to a browser."""
         message = (
-            "Are you running Datashuttle on a machine "
+            "Are you running datashuttle on a machine "
             "that can open a web browser?"
         )
         self.update_message_box_message(message)
@@ -185,7 +185,9 @@ class SetupGdriveScreen(ModalScreen):
         The connection setup is asynchronous so that the user is able to
         cancel the setup if anything goes wrong without quitting datashuttle altogether.
         """
-        message = "Please authenticate through browser."
+        message = (
+            "Please authenticate through browser (it should open shortly)."
+        )
         self.update_message_box_message(message)
 
         asyncio.create_task(
