### 2025-08-26 2e3bcd7
Update datashuttle/utils/folders.py
Co-authored-by: Joe Ziminski <55797454+JoeZiminski@users.noreply.github.com>
