### 2025-08-26 735c636
Extend tooltips and add example placeholders.
