### 2025-08-26 102a691
Partially revert lsjson (#571)
* Revert lsjson for local filesystem searches.

* working on tests.

* Still working on test.

* Document tests.

* Fix tests.

* Try sorting outputs to fix tests.

* Sort all outputs for complete cross-platform equivalence.

* Fix tests on linux.

* Minor changes to docstrings.

* Apply feedback from review.
