### 2025-08-26 954ea65
Uncomment test parameterisation.
