+++ docs/source/pages/get_started/set-up-a-project.md
@@ -2,13 +2,12 @@
 # Set up a project
 
 The first section of this guide will
-set up a "local-only" project that can manage creation
-and validation of project folders. This requires
+set up a "local-only" project that can create
+and validate project folders. This requires
 only minimal configuration to get started.
 
 To see how a datashuttle project can be set up for transfer,
-visit [Set up a project for transfer](set-up-a-project-for-transfer) section.
-
+visit the [Set up a project for transfer](set-up-a-project-for-transfer) section.
 
 ::::{tab-set}
 
@@ -18,8 +17,8 @@ visit [Set up a project for transfer](set-up-a-project-for-transfer) section.
 Selecting `Make New Project` will take you to the project set up screen.
 
 Enter the name of your project, the path to your project folder and
-select `No connection (local only)` (note that the central-path option
-is now disabled).
+select `No connection (local only)` (note that the `Central Path` option
+will be disabled).
 
 ```{image} /_static/screenshots/how-to-make-local-project-configs-dark.png
    :align: center
@@ -75,26 +74,11 @@ project.make_config_file(
 
 ```
 \
-The project is now ready for use, and in future can be instantiated only
+The project is now ready for use, and in future can be instantiated
 with the line ``project = DataShuttle("my_project_name")`` (i.e. you will not
 have to set the `local_path` again).
 
 If you wish to change the project settings at a later time, use ``project.update_config_file()``.
-
-For example, it is possible to immediately validate the project (if it already exists):
-
-```python
-project.validate_project("rawdata", error_or_warn="warn")
-```
-
-Setting ``error_or_warn`` will display all validation issues, otherwise
-it will error on the first one encountered.
-
-New project folders can also be created in the local folder:
-
-```python
-project.create_folders("rawdata", "sub-001", "ses-001_@DATE@", datatype=["ephys", "behav"])
-```
 
 :::
 ::::
@@ -103,17 +87,18 @@ Now, this project is ready for creating and validating
 folders to the [NeuroBlueprint](https://neuroblueprint.neuroinformatics.dev/latest/index.html) standard. See [create folders](how-to-create-folders)
 and [validate folders](tutorial-validation) for details.
 
-If you would also like to transfer files to a central machine, see the next section.
+If you would also like to transfer files to a central storage, see the next section.
 
 (set-up-a-project-for-transfer)=
 ## Set up a project for transfer
 
 Above, we have set up a ``datashuttle`` project by providing the **project name**
-and **local path**. Transfer across the local filesystem or via SSH is supported.
-Therefore, we will need to provide:
+and **local path**. To set up a project for transfer, we need to provide
+additional information:
 
-1) **central path**: location of the project on the central storage machine.
-2) Connection-specific settings (e.g. if using SSH).
+1) **central path**: location of the project on the central storage.
+2) Connection-specific settings (e.g. if using a mounted drive, SSH, Google Drive or
+Amazon Web Services (AWS)).
 
 ```{image} /_static/datashuttle-overview-dark.png
    :align: center
@@ -128,10 +113,12 @@ Therefore, we will need to provide:
 <br>
 
 How the **central path** is set depends on whether your connection to
-central storage is a
-[mounted drive](new-project-mounted-drive)
-or via
-[SSH](new-project-ssh).
+central storage is connected to through a
+[mounted drive](new-project-mounted-drive), via
+[SSH](new-project-ssh) or is an
+[AWS S3 Bucket](new-project-aws)
+or
+[Google Drive](new-project-gdrive).
 
 If you are unsure of your connection method, speak to your lab administrator
 or IT department.
[…]
