### 2025-08-27 22bc636
Small changes to TUI for AWS / GDrive PR (#583)
* Use local only mode as default.

* Capitalise Google Drive.

* Capitalise Rclone.

* Additional newline in tooltip.

* Uncapitalise datashuttle, rename AWS S3 label to AWS S3 Bucket.

* Extend tooltips and add example placeholders.

* Small fixes

* Fix missed uppercase datashuttle.

* Uncomment test parameterisation.

* Fix tests.
