+++ .pre-commit-config.yaml
@@ -6,7 +6,7 @@ exclude: 'conf.py|test_file_conflicts_pathtable.py'
 
 repos:
     - repo: https://github.com/pre-commit/pre-commit-hooks
-      rev: v5.0.0
+      rev: v6.0.0
       hooks:
           - id: check-added-large-files
           - id: check-docstring-first
@@ -31,7 +31,7 @@ repos:
         - id: rst-directive-colons
         - id: rst-inline-touching-normal
     - repo: https://github.com/astral-sh/ruff-pre-commit
-      rev: v0.12.7
+      rev: v0.12.11
       hooks:
         - id: ruff
         - id: ruff-format
