+++ package/packaging_utils.py
@@ -11,7 +11,8 @@ def get_wezterm_version():
 
 def download_wezterm(vendored_dir, wezterm_foldername):
     """ """
-    wezterm_url = f"https://github.com/wezterm/wezterm/releases/download/{get_wezterm_version()}/{wezterm_foldername}.zip"
+    wezterm_url = f"https://github.com/wezterm/wezterm/releases/download/{get_wezterm_version()}/{wezterm_foldername}"  # .zip TODO
+    print(wezterm_url)
 
     wezterm_extracted_dir = vendored_dir / wezterm_foldername
     wezterm_zip_path = vendored_dir / f"{wezterm_foldername}.zip"
