### 2025-09-02 3f1ef5c
Start playing around with Linux 7.
