+++ package/packaging_utils.py
@@ -42,7 +42,7 @@ def download_wezterm(vendored_dir, wezterm_foldername):
             )
         elif platform == "linux":
             subprocess.run(
-                f"chmod +x wezterm_zip_path; {wezterm_zip_path} --appimage-extract",
+                f"chmod +x {wezterm_zip_path}; {wezterm_zip_path} --appimage-extract",
                 shell=True,
             )
         else:
