### 2025-09-02 f72c412
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
