### 2025-09-04 a175d0f
Skip AWS and GDRIVE tests when environment variables not set.
