+++ tests/tests_transfers/gdrive/test_gdrive_transfer.py
@@ -2,10 +2,15 @@ import pytest
 
 from datashuttle.utils import rclone
 
+from ... import test_utils
 from ..base_transfer import BaseTransfer
 from . import gdrive_test_utils
 
 
+@pytest.mark.skipif(
+    not test_utils.has_gdrive_environment_variables(),
+    reason="Google Drive set up environment variables must be set.",
+)
 class TestGdriveTransfer(BaseTransfer):
     @pytest.fixture(
         scope="class",
