### 2025-09-04 22f16ca
Rename setup_google_drive_connection -> setup_gdrive_connection.
