+++ tests/tests_transfers/gdrive/gdrive_test_utils.py
@@ -56,3 +56,15 @@ def setup_gdrive_connection(project: DataShuttle):
 
     builtins.input = original_input
     gdrive.get_client_secret = original_get_secret
+
+
+def has_gdrive_environment_variables():
+    for key in [
+        "GDRIVE_CLIENT_ID",
+        "GDRIVE_ROOT_FOLDER_ID",
+        "GDRIVE_CONFIG_TOKEN",
+        "GDRIVE_CLIENT_SECRET",
+    ]:
+        if key not in os.environ:
+            return False
+    return True
