+++ tests/tests_transfers/gdrive/test_gdrive_transfer.py
@@ -2,13 +2,12 @@ import pytest
 
 from datashuttle.utils import rclone
 
-from ... import test_utils
 from ..base_transfer import BaseTransfer
 from . import gdrive_test_utils
 
 
 @pytest.mark.skipif(
-    not test_utils.has_gdrive_environment_variables(),
+    not gdrive_test_utils.has_gdrive_environment_variables(),
     reason="Google Drive set up environment variables must be set.",
 )
 class TestGdriveTransfer(BaseTransfer):
