+++ datashuttle/datashuttle_functions.py
@@ -63,9 +63,10 @@ def validate_project_from_path(
         If `True`, any alphanumeric character are allowed for the values associated
         with sub- or ses-  keys. Otherwise, values must be integer
         and the following additional checks are performed:
-            - Identical numbers are considered the same value even if padded with different number of zeros
-                (e.g. sub-01 and sub-001_date-20240101 are considered duplicate).
-            - Labels must be the same length (e.g. sub-01 and sub-002 is invalid).
+
+        - Identical numbers are considered the same value even if padded with different number of zeros
+          (e.g. sub-01 and sub-001_date-20240101 are considered duplicate).
+        - Labels must be the same length (e.g. sub-01 and sub-002 is invalid).
 
     Returns
     -------
