+++ docs/source/pages/user_guides/create-folders.md
@@ -124,18 +124,25 @@ the top-level folder, and bypass validation.
 ```{image} /_static/screenshots/how-to-create-folders-settings-dark.png
    :align: center
    :class: only-dark
-   :width: 500px
+   :width: 900px
 ```
 ```{image} /_static/screenshots/how-to-create-folders-settings-light.png
    :align: center
    :class: only-light
-   :width: 500px
+   :width: 900px
+
 ```
 <br>
 
 Top level folder
 : This dropdown box will set whether folders are created in the
 `rawdata` or `derivatives` top-level folder.
+
+Search central for suggestions
+: By default, suggesting the next `sub-` or `ses-` number will
+be based only on the local project. Select this to also search
+the project on central. If a `local only` project, this will have
+no effect.
 
 Bypass validation
 : This setting will allow folder creation even if the names
