### 2025-09-05 58dc123
Change names of similar functions in interface.py and elsewhere.
