### 2025-09-05 de71250
Refactor for TUI tests, skip AWS and GDrive TUI set up tests if environment variables not set.
