+++ tests/tests_tui/test_tui_aws_configs.py
@@ -4,7 +4,7 @@ import pytest
 
 from datashuttle.tui.app import TuiApp
 
-from ..tui_configs_base import TuiConfigsBase
+from .tui_configs_base import TuiConfigsBase
 
 
 class TestTuiAwsConfigs(TuiConfigsBase):
