### 2025-09-05 de34018
Set functions to API links.
