### 2025-09-05 3fc5e26
Set functions to API links.
