### 2025-09-05 6d64759
Set functions to API links.
