+++ .github/workflows/docs_build_and_deploy.yml
@@ -19,7 +19,12 @@ jobs:
     name: Build Sphinx Docs
     runs-on: ubuntu-latest
     steps:
-      - uses: neuroinformatics-unit/actions/build_sphinx_docs@v2
+      - uses: neuroinformatics-unit/actions/build_sphinx_docs@main
+        with:
+          python-version: 3.12
+          use-make: false
+          fetch-tags: false
+          use-artifactci: lazy
 
   deploy_sphinx_docs:
     name: Deploy Sphinx Docs
