### 2025-09-08 801fea1
Rename 'name templates' to 'validation templates'.
