### 2025-09-08 1aa0f2e
Rename 'name templates' to 'validation templates'.
