### 2025-09-08 fbf15eb
Rename 'name templates' to 'validation templates'.
