+++ datashuttle/datashuttle_class.py
@@ -218,13 +218,13 @@ class DataShuttle:
         utils.log("\nFormatting Names...")
         ds_logger.log_names(["sub_names", "ses_names"], [sub_names, ses_names])
 
-        name_templates = self.get_name_templates()
+        validation_templates = self.get_validation_templates()
 
         format_sub, format_ses = self._format_and_validate_names(
             top_level_folder,
             sub_names,
             ses_names,
-            name_templates,
+            validation_templates,
             bypass_validation,
             allow_letters_in_sub_ses_values,
             log=True,
@@ -261,7 +261,7 @@ class DataShuttle:
         top_level_folder: TopLevelFolder,
         sub_names: Union[str, List[str]],
         ses_names: Optional[Union[str, List[str]]],
-        name_templates: Dict,
+        validation_templates: Dict,
         bypass_validation: bool,
         allow_letters_in_sub_ses_values: bool,
         log: bool = True,
@@ -270,7 +270,7 @@ class DataShuttle:
         format_sub = formatting.check_and_format_names(
             sub_names,
             "sub",
-            name_templates,
+            validation_templates,
             bypass_validation,
             allow_letters_in_sub_ses_values,
         )
@@ -279,7 +279,7 @@ class DataShuttle:
             format_ses = formatting.check_and_format_names(
                 ses_names,
                 "ses",
-                name_templates,
+                validation_templates,
                 bypass_validation,
                 allow_letters_in_sub_ses_values,
             )
@@ -295,7 +295,7 @@ class DataShuttle:
                 include_central=False,
                 display_mode="error",
                 log=log,
-                name_templates=name_templates,
+                validation_templates=validation_templates,
                 allow_letters_in_sub_ses_values=allow_letters_in_sub_ses_values,
             )
 
@@ -1177,9 +1177,9 @@ class DataShuttle:
         The next subject ID.
 
         """
-        name_template = self.get_name_templates()
-        name_template_regexp = (
-            name_template["sub"] if name_template["on"] else None
+        validation_template = self.get_validation_templates()
+        validation_template_regexp = (
+            validation_template["sub"] if validation_template["on"] else None
         )
 
         if self.is_local_project():
@@ -1192,7 +1192,7 @@ class DataShuttle:
             include_central=include_central,
             return_with_prefix=return_with_prefix,
             search_str="sub-*",
-            name_template_regexp=name_template_regexp,
+            validation_template_regexp=validation_template_regexp,
         )
 
     @check_configs_set
@@ -1226,9 +1226,9 @@ class DataShuttle:
         The next session ID.
 
         """
-        name_template = self.get_name_templates()
-        name_template_regexp = (
-            name_template["ses"] if name_template["on"] else None
+        validation_template = self.get_validation_templates()
+        validation_template_regexp = (
+            validation_template["ses"] if validation_template["on"] else None
         )
 
         if self.is_local_project():
@@ -1241,7 +1241,7 @@ class DataShuttle:
             include_central=include_central,
             return_with_prefix=return_with_prefix,
             search_str="ses-*",
-            name_template_regexp=name_template_regexp,
+            validation_template_regexp=validation_template_regexp,
         )
 
     @check_configs_set
@@ -1256,36 +1256,38 @@ class DataShuttle:
     # Name Templates
     # -------------------------------------------------------------------------
 
-    def get_name_templates(self) -> Dict:
+    def get_validation_templates(self) -> Dict:
         """Return the regexp templates used for validation.
 
         If the "on" key is set to `False`, template validation is not performed.
 
         Returns
         -------
-        name_templates
-            e.g. {"name_templates": {"on": False, "sub": None, "ses": None}}
+        validation_templates
[…]
