### 2025-09-08 09c52c9
Rename quick_validate_project.
