### 2025-09-09 e2940b2
Cover DATETO etc. cases in check_and_format_names so they are not validated.
