+++ datashuttle/utils/formatting.py
@@ -63,7 +63,13 @@ def check_and_format_names(
 
     names_to_format, reserved_keywords = [], []
     for name in names:
-        if name in canonical_reserved_keywords() or tags("*") in name:
+        if (
+            name in canonical_reserved_keywords()
+            or tags("*") in name
+            or tags("DATETO") in name
+            or tags("TIMETO") in name
+            or tags("DATETIMETO") in name
+        ):
             if tags("to") in name:
                 # handle an edge case where use searches with both tags
                 reserved_keywords += update_names_with_range_to_flag(
