### 2025-09-09 e7a341f
Fix get_connection_methods_list().
