### 2025-09-09 b7c94c5
Fix radiobutton_id_from_connection_method in TUI.
