### 2025-09-09 c0f4a9c
Revert changes and instead use 'local_only' internally for connection method is local only.
