+++ datashuttle/datashuttle_class.py
@@ -971,6 +971,7 @@ class DataShuttle:
 
         connection_method
             The method used to connect to the central project filesystem,
+            ``None`` is an alias for ``"local_only"``.
             e.g. ``"local_filesystem"`` (e.g. mounted drive) or ``"ssh"``
 
         central_host_id
@@ -1007,6 +1008,10 @@ class DataShuttle:
             local_vars=locals(),
             store_in_temp_folder=True,
         )
+
+        if connection_method is None:
+            # For backward compatibility
+            connection_method = "local_only"
 
         if self._config_path.is_file():
             utils.log_and_raise_error(
@@ -1063,6 +1068,11 @@ class DataShuttle:
             "update-config-file",
             local_vars=locals(),
         )
+
+        if "connection_method" in kwargs:
+            if kwargs["connection_method"] is None:
+                # For backward compatibility
+                kwargs["connection_method"] = "local_only"
 
         new_cfg = copy.deepcopy(self.cfg)
         new_cfg.update(**kwargs)
