+++ datashuttle/tui/shared/configs_content.py
@@ -6,6 +6,7 @@ from typing import TYPE_CHECKING, Any, Dict, List, Literal, Optional, Union
 if TYPE_CHECKING:
     from textual.app import ComposeResult
 
+    from datashuttle.concifgs.canonical_configs import ConnectionMethods
     from datashuttle.tui.interface import Interface
     from datashuttle.tui.screens.new_project import NewProjectScreen
     from datashuttle.tui.screens.project_manager import ProjectManagerScreen
@@ -171,7 +172,9 @@ class ConfigsContent(Container):
             RadioSet(
                 RadioButton(
                     "No connection (local only)",
-                    id=self.radiobutton_id_from_connection_method(None),
+                    id=self.radiobutton_id_from_connection_method(
+                        "local_only"
+                    ),
                 ),
                 RadioButton(
                     "Local Filesystem",
@@ -278,7 +281,7 @@ class ConfigsContent(Container):
         else:
             self.query_one("#configs_local_only_radiobutton").value = True
 
-            self.setup_widgets_to_display(connection_method=None)
+            self.setup_widgets_to_display(connection_method="local_only")
 
         # Setup tooltips
         if not self.interface:
@@ -337,9 +340,6 @@ class ConfigsContent(Container):
         self, connection_method: str | None
     ) -> str:
         """Create a canonical radiobutton textual ID from the connection method."""
-        if connection_method is None:
-            connection_method = "local_only"
-
         return f"configs_{connection_method}_radiobutton"
 
     def connection_method_from_radiobutton_id(
@@ -352,21 +352,15 @@ class ConfigsContent(Container):
         connection_method = radiobutton_id[
             len("configs_") : -len("_radiobutton")
         ]
-        if connection_method == "local_only":
-            connection_method = None
-
         return connection_method
 
     def set_central_path_input_tooltip(
         self, connection_method: str | None
     ) -> None:
         """Set tooltip depending on the connection method."""
-        if connection_method is None:
-            tooltip = get_tooltip("config_central_path_input_mode-local_only")
-        else:
-            tooltip = get_tooltip(
-                f"config_central_path_input_mode-{connection_method}"
-            )
+        tooltip = get_tooltip(
+            f"config_central_path_input_mode-{connection_method}"
+        )
         self.query_one("#configs_central_path_input").tooltip = tooltip
 
     def get_platform_dependent_example_paths(
@@ -661,7 +655,7 @@ class ConfigsContent(Container):
             "configs_aws_radiobutton":
                 cfg_to_load["connection_method"] == "aws",
             "configs_local_only_radiobutton":
-                cfg_to_load["connection_method"] is None,
+                cfg_to_load["connection_method"] == "local_only",
         }
         # fmt: on
 
@@ -743,7 +737,9 @@ class ConfigsContent(Container):
         )
         select.value = value
 
-    def setup_widgets_to_display(self, connection_method: str | None) -> None:
+    def setup_widgets_to_display(
+        self, connection_method: ConnectionMethods
+    ) -> None:
         """Set up widgets to display based on the chosen `connection_method` on the radiobutton.
 
         The widgets pertaining to the chosen connection method will be displayed.
@@ -756,7 +752,7 @@ class ConfigsContent(Container):
         Called on mount, on radiobuttons' switch and upon saving project configs.
         """
         assert connection_method in get_connection_methods_list(), (
-            "Unexpected Connection Method"
+            "Unexpected connection method."
         )
 
         # Connection specific widgets
@@ -773,7 +769,7 @@ class ConfigsContent(Container):
             for widget in connection_widgets:
                 widget.display = connection_method == name
 
-        has_connection_method = connection_method is not None
+        has_connection_method = connection_method != "local_only"
 
[…]
