### 2025-09-09 fcc3994
Remove unecessary arguments.
