### 2025-09-09 8674958
Fix  connection method unit test.
