+++ tests/tests_unit/test_unit.py
@@ -244,7 +244,7 @@ class TestUnit:
             "local_filesystem",
             "gdrive",
             "aws",
-            None,
+            "local_only",
         ]
 
     # -------------------------------------------------------------------------
