### 2025-09-09 ead1185
Make ConnectionMethods a custom type, fix make_config_file default arg.
