+++ tests/tests_transfers/ssh/base_ssh.py
@@ -25,7 +25,7 @@ class BaseSSHTransfer(BaseTransfer):
     @pytest.fixture(
         scope="class",
     )
-    def setup_ssh_container(self):
+    def setup_ssh_container_fixture(self):
         """
         Set up the Dockerfile container for SSH tests and
         delete it on teardown.
