### 2025-09-10 8cfdec7
Rename `quick_validate_project` to `validate_project_from_path` (#593)
* Rename quick_validate_project.

* Update TUI to from 'at' to 'from'.
