+++ docs/source/pages/get_started/quick-validate-project.md
@@ -1,5 +1,5 @@
 :orphan:
-(quick-validate-projects)=
+(validate-project-from-path)=
 
 # Validate a project from a filepath
 
@@ -14,7 +14,7 @@ to any problematic folders.
 :sync: gui
 
 To quickly validate a project, start the terminal user interface with
-``datashuttle launch`` and click ``Validate Project at Path``.
+``datashuttle launch`` and click ``Validate Project From Path``.
 
 The screen below will show. To validate an existing project,
 enter the full filepath to the project folder in the top input box
@@ -54,12 +54,12 @@ raise a validation issue.
 :sync: python
 
 To validate a project using the Python API, pass the path
-to the project to validate to ``quick_validate_project``:
+to the project to validate to ``validate_project_from_path``:
 
 ```python
-from datashuttle import quick_validate_project
+from datashuttle import validate_project_from_path
 
-quick_validate_project(
+validate_project_from_path(
     project_path="/mydrive/path/to/project/project_name",
     display_mode="error",
 )
@@ -69,7 +69,7 @@ quick_validate_project(
 In this case, `display_mode=error` will result in an error on the first encountered validation issue.
 Otherwise, `"warn"` will show a python warning for all detected issues, while `"print"` will print directly to the console.
 
-See the [](datashuttle.quick_validate_project) API documentation
+See the [](validate_project_from_path()) API documentation
 for full details of parameters, including the important argument ``strict_mode``
 that controls how validation is performed.
 
