### 2025-09-10 10826a3
Small tidy ups.
