### 2025-09-10 a5168ba
Turn off validation during data transfers, add test.
