+++ tests/test_utils.py
@@ -334,7 +334,13 @@ def check_datatype_sub_ses_uploaded_correctly(
 
 
 def make_and_check_local_project_folders(
-    project, top_level_folder, subs, sessions, datatype, datatypes_used=None
+    project,
+    top_level_folder,
+    subs,
+    sessions,
+    datatype,
+    datatypes_used=None,
+    bypass_validation=False,
 ):
     """Make a local project folder tree with the specified datatype,
     subs, sessions and check it is made successfully.
@@ -347,7 +353,7 @@ def make_and_check_local_project_folders(
         datatypes_used = get_all_broad_folders_used()
 
     make_local_folders_with_files_in(
-        project, top_level_folder, subs, sessions, datatype
+        project, top_level_folder, subs, sessions, datatype, bypass_validation
     )
 
     check_folder_tree_is_correct(
@@ -359,9 +365,20 @@ def make_and_check_local_project_folders(
 
 
 def make_local_folders_with_files_in(
-    project, top_level_folder, subs, sessions=None, datatype=""
+    project,
+    top_level_folder,
+    subs,
+    sessions=None,
+    datatype="",
+    bypass_validation=False,
 ):
-    project.create_folders(top_level_folder, subs, sessions, datatype)
+    project.create_folders(
+        top_level_folder,
+        subs,
+        sessions,
+        datatype,
+        bypass_validation=bypass_validation,
+    )
     for root, dirs, _ in os.walk(project.cfg["local_path"]):
         if not dirs:
             path_ = Path(root) / "placeholder_file.txt"
