### 2025-09-10 0c99066
Tidy up validation.
