### 2025-09-10 a2f8fa9
Explain better the leading zero case, do not mention this generally because inconsnstent zeros are not allowed anyway.
