+++ datashuttle/tui/tooltips.py
@@ -188,8 +188,6 @@ def get_tooltip(id: str) -> str:
             "If `True`, any alphanumeric character are allowed for the values associated "
             "with sub- or ses- keys. \n\n"
             "Otherwise, values must be integer and the following additional checks are performed:\n"
-            " - Identical numbers are considered the same value even if padded with different number of zeros "
-            "   (e.g. sub-01 and sub-001_date-20240101 are considered duplicate).\n"
             " - Labels must be the same length (e.g. sub-01 and sub-002 is invalid)."
         )
 
