### 2025-09-10 7692f1c
Small fixes to docs and docstrings.
