### 2025-09-10 e82d675
Add test docstring.
