### 2025-09-10 5d8d813
Refactor cannoincal_config checks for readability.
