### 2025-09-11 4234912
Refactor SSH method.
