+++ tests/test_utils.py
@@ -40,11 +40,7 @@ def setup_project_default_configs(
 
     project.make_config_file(**default_configs)
 
-    rclone.setup_rclone_config_for_ssh(
-        project.cfg,
-        project.cfg.get_rclone_config_name("ssh"),
-        project.cfg.ssh_key_path,
-    )
+    project._setup_rclone_central_local_filesystem_config()
 
     if local_path:
         os.makedirs(local_path, exist_ok=True)
