+++ datashuttle/utils/rclone.py
@@ -3,8 +3,6 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Dict, List, Literal, Optional
 
 if TYPE_CHECKING:
-    from pathlib import Path
-
     from datashuttle.configs.config_class import Configs
     from datashuttle.utils.custom_types import TopLevelFolder
 
@@ -176,7 +174,7 @@ def setup_rclone_config_for_local_filesystem(
 def setup_rclone_config_for_ssh(
     cfg: Configs,
     rclone_config_name: str,
-    ssh_key_path: Path,
+    private_key_str: str,
     log: bool = True,
 ) -> None:
     """Set the RClone remote config for ssh.
@@ -194,24 +192,26 @@ def setup_rclone_config_for_ssh(
         canonical config name, generated by
         datashuttle.cfg.get_rclone_config_name()
 
-    ssh_key_path
-        path to the ssh key used for connecting to
-        ssh central filesystem
+    private_key_str
+        PEM encoded sssh private key to pass to RClone.
 
     log
         whether to log, if True logger must already be initialised.
 
     """
-    call_rclone(
+    key_escaped = private_key_str.replace("\n", "\\n")
+
+    command = (
         f"config create "
         f"{rclone_config_name} "
         f"sftp "
         f"host {cfg['central_host_id']} "
         f"user {cfg['central_host_username']} "
         f"port {canonical_configs.get_default_ssh_port()} "
-        f"key_file {ssh_key_path.as_posix()}",
-        pipe_std=True,
+        f'-- key_pem "{key_escaped}"'
     )
+
+    call_rclone(command, pipe_std=True)  # TODO: handle bad return code...
 
     if log:
         log_rclone_config_output()
@@ -274,7 +274,7 @@ def setup_rclone_config_for_gdrive(
         f"{client_secret_key_value}"
         f"scope drive "
         f"root_folder_id {cfg['gdrive_root_folder_id']} "
-        f"{extra_args}",
+        f"{extra_args}"
     )
 
     return process
