### 2025-09-12 bb212ae
Add separate config files for all connection methods.
