+++ datashuttle/tui/shared/validate_content.py
@@ -101,9 +101,9 @@ class ValidateContent(Container):
                     id="validate_strict_mode_checkbox",
                 ),
                 Checkbox(
-                    "Allow alphanumeric sub- and ses- values",
+                    "Allow letters in sub- and ses- values",
                     value=False,
-                    id="validate_allow_alphanumeric_sub_ses_values_checkbox",
+                    id="validate_allow_letters_in_sub_ses_values_checkbox",
                 ),
                 id="validate_arguments_horizontal",
             ),
@@ -121,7 +121,7 @@ class ValidateContent(Container):
             "validate_top_level_folder_select",
             "validate_include_central_checkbox",
             "validate_strict_mode_checkbox",
-            "validate_allow_alphanumeric_sub_ses_values_checkbox",
+            "validate_allow_letters_in_sub_ses_values_checkbox",
         ]:
             self.query_one(f"#{id}").tooltip = get_tooltip(id)
 
@@ -160,8 +160,8 @@ class ValidateContent(Container):
                 "#validate_strict_mode_checkbox"
             ).value
 
-            allow_alphanumeric_sub_ses_values = self.query_one(
-                "#validate_allow_alphanumeric_sub_ses_values_checkbox"
+            allow_letters_in_sub_ses_values = self.query_one(
+                "#validate_allow_letters_in_sub_ses_values_checkbox"
             ).value
 
             if self.interface:
@@ -176,7 +176,7 @@ class ValidateContent(Container):
                     top_level_folder=top_level_folder,
                     include_central=include_central,
                     strict_mode=strict_mode,
-                    allow_alphanumeric_sub_ses_values=allow_alphanumeric_sub_ses_values,
+                    allow_letters_in_sub_ses_values=allow_letters_in_sub_ses_values,
                 )
                 if not success:
                     self.parent_class.mainwindow.show_modal_error_dialog(
@@ -206,7 +206,7 @@ class ValidateContent(Container):
                     path_,
                     top_level_folder=top_level_folder,
                     strict_mode=strict_mode,
-                    allow_alphanumeric_sub_ses_values=allow_alphanumeric_sub_ses_values,
+                    allow_letters_in_sub_ses_values=allow_letters_in_sub_ses_values,
                 )
                 self.write_results_to_richlog(output)
 
