### 2025-09-15 f53d00e
Update tests/tests_tui/test_tui_widgets_and_defaults.py
Co-authored-by: Shrey Singh <96627769+cs7-shrey@users.noreply.github.com>
