### 2025-09-15 60dface
Use allow_letters_in_sub_ses_values in test.
