### 2025-09-15 730c147
Get allow_letters_in_sub_ses_values from tui_settings in interface.
