### 2025-09-15 8fbc4e6
Add comment to check_and_format_names.
