### 2025-09-16 aef36f9
Support alphanumeric sub or ses labels. (#575)
* Support alphanumeric sub or ses labels.

* Fix linting.

* Fix accidentally included dash in check.

* Fix tests.

* Fixing tests.

* Small fixes, add first tests.

* Add tests.

* Rename placeholder variable.

* Add tooltips.

* Tidying up TUI, trying to get create settings page nice.

* Fix tests.

* Fix tests again.

* Update docs text.

* Update tcss.

* Update documentation.

* Small tidy ups.

* Turn off validation during data transfers, add test.

* Tidy up validation.

* Explain better the leading zero case, do not mention this generally
[…]
