+++ datashuttle/utils/validation.py
@@ -41,10 +41,17 @@ def get_missing_prefix_error(name: str, prefix, path_: Path | None) -> str:
     )
 
 
-def get_bad_value_error(name: str, prefix, path_: Path | None) -> str:
+def get_bad_value_error(
+    name: str,
+    prefix,
+    path_: Path | None,
+    allow_letters_in_sub_ses_values: bool,
+) -> str:
     """Return error message when the value for a prefix is not an integer."""
+    type_ = "alphanumeric" if allow_letters_in_sub_ses_values else "an integer"
+
     return handle_path(
-        f"BAD_VALUE: The value for prefix {prefix} in name {name} is not an integer.",
+        f"BAD_VALUE: The value for prefix {prefix} in name {name} is not {type_}.",
         path_,
     )
 
@@ -148,6 +155,7 @@ def validate_list_of_names(
     prefix: Prefix,
     name_templates: Optional[Dict] = None,
     check_value_lengths: bool = True,
+    allow_letters_in_sub_ses_values: bool = False,
 ) -> List[str]:
     """Validate a list of subject or session names against NeuroBlueprint.
 
@@ -166,6 +174,13 @@ def validate_list_of_names(
         If `True`, check that the prefix-<value> value lengths
         are consistent across the passed list.
 
+    allow_letters_in_sub_ses_values
+        If `True`, any alphanumeric character are allowed for the values associated
+        with sub- or ses-  keys. Otherwise, values must be integer
+        and the following additional checks are performed:
+
+        - Labels must be the same length (e.g. sub-01 and sub-002 is invalid).
+
     Returns
     -------
     error_messages
@@ -182,7 +197,7 @@ def validate_list_of_names(
         path_, name = get_path_and_name(path_or_name)
 
         error_messages += prefix_is_duplicate_or_has_bad_values(
-            name, prefix, path_
+            name, prefix, path_, allow_letters_in_sub_ses_values
         )
         error_messages += name_begins_with_bad_key(name, prefix, path_)
         error_messages += names_include_special_characters(name, path_)
@@ -200,17 +215,20 @@ def validate_list_of_names(
     # Note this called functions again loop over the list (O(n^2)) so
     # this is not very efficient but these lists should never be that long.
     stripped_path_or_names_list = strip_uncheckable_names(
-        path_or_name_list, prefix
+        path_or_name_list, prefix, allow_letters_in_sub_ses_values
     )
 
     for path_or_name in stripped_path_or_names_list:
         path_, name = get_path_and_name(path_or_name)
 
         error_messages += new_name_duplicates_existing(
-            name, stripped_path_or_names_list, prefix
+            name,
+            stripped_path_or_names_list,
+            prefix,
+            allow_letters_in_sub_ses_values,
         )
 
-    if check_value_lengths:
+    if not allow_letters_in_sub_ses_values and check_value_lengths:
         error_messages += value_lengths_are_inconsistent(
             stripped_path_or_names_list, prefix
         )
@@ -219,12 +237,19 @@ def validate_list_of_names(
 
 
 def prefix_is_duplicate_or_has_bad_values(
-    name: str, prefix: Prefix, path_: Path | None
+    name: str,
+    prefix: Prefix,
+    path_: Path | None,
+    allow_letters_in_sub_ses_values: bool,
 ) -> List[str]:
     """Check the sub- or ses- prefix.
 
     Ensure it is found only once in the name and
     that its value can be converted to integer.
+
+    These are tested together because we cannot test
+    for integer sub- or ses- prefix without being
+    sure it exists and is not duplicated.
 
     Parameters
     ----------
@@ -237,12 +262,16 @@ def prefix_is_duplicate_or_has_bad_values(
     path_
[…]
