### 2025-09-16 3f88acb
improve test_bad_setup.
