+++ tests/tests_integration/test_local_only_mode.py
@@ -30,25 +30,65 @@ class TestLocalOnlyProject(BaseTest):
             in str(e.value)
         )
 
-        for connection_method in ["ssh"]:  # , "local_filesystem"]:
-            with pytest.raises(ConfigError) as e:
-                project.make_config_file(
-                    local_path,
-                    connection_method=connection_method,
-                    central_host_id="",
-                    central_host_username="",
-                )
-
-            assert (
-                f"'central_host_id' and 'central_host_username' are required if 'connection_method' is '{connection_method}'."
-                in str(e.value)
+        with pytest.raises(ConfigError) as e:
+            project.make_config_file(
+                local_path, connection_method="local_filesystem"
             )
+
+        assert (
+            "`central_path` must be set if `connection_method` is local_filesystem"
+            in str(e.value)
+        )
+
+        with pytest.raises(ConfigError) as e:
+            project.make_config_file(
+                local_path,
+                connection_method="ssh",
+            )
+
+        assert (
+            "central_host_id' and 'central_host_username' are required if 'connection_method' is 'ssh'."
+            in str(e.value)
+        )
+
+        with pytest.raises(ConfigError) as e:
+            project.make_config_file(
+                local_path,
+                connection_method="ssh",
+                central_host_id="@",
+                central_host_username="@",
+            )
+
+        assert (
+            "`central_path` must be set if `connection_method` is ssh"
+            in str(e.value)
+        )
 
         with pytest.raises(ConfigError) as e:
             project.make_config_file(local_path, connection_method="aws")
         assert (
             "`central_path` cannot be `None` when `connection_method` is 'aws'. `central_path` must include the s3 bucket name."
             in str(e.value)
+        )
+
+        with pytest.raises(ConfigError) as e:
+            project.make_config_file(
+                local_path, central_path="@", connection_method="aws"
+            )
+
+        assert (
+            str(e.value)
+            == "Both aws_access_key_id and aws_region must be present for AWS connection."
+        )
+
+        with pytest.raises(ConfigError) as e:
+            project.make_config_file(
+                local_path, central_path="@", connection_method="gdrive"
+            )
+
+        assert (
+            str(e.value)
+            == "'gdrive_root_folder_id' is required if 'connection_method' is 'gdrive'."
         )
 
     @pytest.mark.parametrize("project", ["local"], indirect=True)
