### 2025-09-16 038f2f8
Add AWS and GDrive to paper.
