### 2025-09-18 bfbb3eb
Reduce CI usage.
