+++ .github/workflows/code_test_and_deploy.yml
@@ -6,7 +6,7 @@ on:
       - main
     tags:
       - 'v*'
-  workflow_dispatch:
+  pull_request:
   schedule:
     # Cron runs on the 1st and 15th of every month.
     # This will only run on main by default.
