### 2025-09-19 3fa053a
Reduce CI usage. (#608)
* Reduce CI usage.

* Run on workflow dispatch.

* Revert workflow dispatch.

* Don't run tests on draft PRs.
