### 2025-09-19 b171139
Update readme, rebuild paper.
