+++ joss-paper/README.md
@@ -1,3 +1,3 @@
-Build with:
+On windows, build with the below command when in the `joss_paper` folder:
 
 docker run --rm -v ${PWD}:/data openjournals/paperdraft
