### 2025-09-19 036ac6b
Refactor load_old_version_yamls.
