### 2025-09-24 f02d6f1
Add password machinery for windows.
