### 2025-09-24 cfe9d71
Add password machinery for windows.
