### 2025-09-26 5fa0d09
Improve error message.
