### 2025-09-26 21ac6f4
Change `None` to `"local_only"` for connection method and expand connection tests (#596)
* Fix get_connection_methods_list().

* Fix radiobutton_id_from_connection_method in TUI.

* Revert changes and instead use 'local_only' internally for connection method is local only.

* Update backwards compatability tests.

* Fix  connection method unit test.

* Make ConnectionMethods a custom type, fix make_config_file default arg.

* Refactor cannoincal_config checks for readability.

* Fix central path error tests.

* improve test_bad_setup.

* Fix type hint.

* Refactor load_old_version_yamls.

* Ad
[…]
