+++ datashuttle/utils/custom_types.py
@@ -9,3 +9,7 @@ OverwriteExistingFiles = Literal["never", "always", "if_source_newer"]
 Prefix = Literal["sub", "ses"]
 
 InterfaceOutput = Tuple[bool, Any]
+
+ConnectionMethods = Literal[
+    "ssh", "local_filesystem", "gdrive", "aws", "local_only"
+]
