+++ .github/workflows/code_test_and_deploy.yml
@@ -6,7 +6,8 @@ on:
       - main
     tags:
       - 'v*'
-  pull_request: [opened, synchronize, reopened, ready_for_review]
+  pull_request:
+    types: [opened, synchronize, reopened, ready_for_review]
   schedule:
     # Cron runs on the 1st and 15th of every month.
     # This will only run on main by default.
