### 2025-09-29 5aa7037
Prototype input for set up ssh connection.
