### 2025-09-29 bcc5da4
Prototype input for set up ssh connection.
