### 2025-09-29 0c4017e
Refactor rclone call functions to handle password.
