+++ datashuttle/utils/rclone.py
@@ -48,7 +48,17 @@ def call_rclone(command: str, pipe_std: bool = False) -> CompletedProcess:
     return output
 
 
-def call_rclone_through_script(command: str) -> CompletedProcess:
+def call_rclone_for_central_connection(
+    cfg, command: str, pipe_std: bool = False
+) -> CompletedProcess:
+    return run_function_that_may_require_central_connection_password(
+        cfg, lambda: call_rclone(command, pipe_std)
+    )
+
+
+def call_rclone_through_script_for_central_connection(
+    cfg, command: str
+) -> CompletedProcess:
     """Call rclone through a script.
 
     This is to avoid limits on command-line calls (in particular on Windows).
@@ -84,11 +94,15 @@ def call_rclone_through_script(command: str) -> CompletedProcess:
         if system != "Windows":
             os.chmod(tmp_script_path, 0o700)
 
-        output = subprocess.run(
+        lambda_func = lambda: subprocess.run(
             [tmp_script_path],
             stdout=subprocess.PIPE,
             stderr=subprocess.PIPE,
             shell=False,
+        )
+
+        output = run_function_that_may_require_central_connection_password(
+            cfg, lambda_func
         )
 
         if output.returncode != 0:
@@ -100,7 +114,9 @@ def call_rclone_through_script(command: str) -> CompletedProcess:
     return output
 
 
-def call_rclone_with_popen(command: str) -> subprocess.Popen:
+def call_rclone_with_popen_for_central_connection(
+    command: str,
+) -> subprocess.Popen:
     """Call rclone using `subprocess.Popen` for control over process termination.
 
     It is not possible to kill a process while running it using `subprocess.run`.
@@ -116,7 +132,7 @@ def call_rclone_with_popen(command: str) -> subprocess.Popen:
     return process
 
 
-def await_call_rclone_with_popen_raise_on_fail(
+def await_call_rclone_with_popen_for_central_connection_raise_on_fail(
     process: subprocess.Popen, log: bool = True
 ):
     """Await rclone the subprocess.Popen call.
@@ -124,13 +140,35 @@ def await_call_rclone_with_popen_raise_on_fail(
     Calling `process.communicate()` waits for the process to complete and returns
     the stdout and stderr.
     """
-    stdout, stderr = process.communicate()
+    lambda_func = lambda: process.communicate()
+
+    stdout, stderr = run_function_that_may_require_central_connection_password(
+        cfg, lambda_func
+    )
 
     if process.returncode != 0:
         utils.log_and_raise_error(stderr.decode("utf-8"), ConnectionError)
 
     if log:
         log_rclone_config_output()
+
+
+def run_function_that_may_require_central_connection_password(
+    cfg, lambda_func
+):
+    """ """
+    set_password = cfg.backend_has_password[cfg["connection_method"]]
+
+    if set_password:
+        config_filepath = rclone_password.get_password_filepath(cfg)
+        rclone_password.set_credentials_as_password_command(config_filepath)
+
+    results = lambda_func()
+
+    if set_password:
+        rclone_password.remove_credentials_as_password_command()
+
+    return results
 
 
 # -----------------------------------------------------------------------------
@@ -262,7 +300,7 @@ def setup_rclone_config_for_gdrive(
 ) -> subprocess.Popen:
     """Set up rclone config for connections to Google Drive.
 
-    This function uses `call_rclone_with_popen` instead of `call_rclone`. This
+    This function uses `call_rclone_with_popen_for_central_connection` instead of `call_rclone`. This
     is done to have more control over the setup process in case the user wishes to
[…]
