### 2025-09-29 a59f96f
Refactor rclone call functions to handle password.
