+++ datashuttle/utils/aws.py
@@ -10,7 +10,8 @@ def check_if_aws_bucket_exists(cfg: Configs) -> bool:
 
     The first part of`cfg["central_path"] should be an existing bucket name.
     """
-    output = rclone.call_rclone(
+    output = rclone.call_rclone_for_central_connection(
+        cfg,
         f"lsjson {cfg.get_rclone_config_name()}: {rclone.get_config_arg(cfg)}",
         pipe_std=True,
     )
