### 2025-10-01 fb5232f
Update paper for AWS / Google Drive (#606)
* Add AWS and GDrive to paper.

* Small edits.

* Add shrey affiliation.

* Update readme, rebuild paper.

* Tighten up the section

* slightly reword sentence

* fix typo

* use 'via' instead of 'through'

* Update paper pdf.
