### 2025-10-01 5329c58
Add to macOS.
