### 2025-10-01 a6d1763
Add to macOS.
