### 2025-10-02 366d67b
Refactoring.
