+++ datashuttle/datashuttle_class.py
@@ -4,6 +4,7 @@ import copy
 import glob
 import json
 import os
+import platform
 import shutil
 from pathlib import Path
 from typing import (
@@ -112,92 +113,6 @@ class DataShuttle:
         self.cfg.init_paths()
 
         self._make_project_metadata_if_does_not_exist()
-
-    def _try_set_rclone_password(self):  # TODO: BETTER NAME!
-        """"""
-        input_ = input(
-            f"Your SSH key will be stored in the rclone config at:\n "  ## TODO: FIX T HIS IS NOT SSH
-            f"{rclone.get_full_config_filepath(self.cfg)}.\n\n"
-            f"Would you like to set a password using Windows credential manager? "
-            f"Press 'y' to set password or leave blank to skip."
-        )
-
-        set_password = input_ == "y"
-
-        if set_password:
-            try:
-                self.set_rclone_password()
-            except BaseException as e:
-                print(e)
-                # THIS PATH IS WRONG
-                config_path = rclone.get_full_config_filepath(self.cfg)
-
-                raise RuntimeError(
-                    f"Password set up failed. The config at {config_path} contains the private ssh key without a password.\n"
-                    f"Use set_rclone_password()` to attempt to set the password again (see stacktrace above). "
-                )
-
-            print("Password set successfully")
-
-    def set_rclone_password(self):
-        """"""
-        # TODO: CHECK CONNECTION METHOD
-        connection_method = self.cfg["connection_method"]
-
-        if self.cfg.rclone_has_password[connection_method]:
-            raise RuntimeError(
-                "This config file already has a password set. First, use `remove_rclone_password` to remove it."
-            )
-
-        rclone_config_path = rclone.get_full_config_filepath(
-            self.cfg
-        )  # change name to rclone config because this is getting confusing!
-
-        if not rclone_config_path.exists():
-            raise RuntimeError(
-                f"Rclone config file for: {connection_method} was not found. "
-                f"Make sure you set up the connection first with `setup_{connection_method}_connection()`"
-            )
-
-        password_filepath = rclone_password.get_password_filepath(self.cfg)
-
-        if password_filepath.exists():
-            password_filepath.unlink()
-
-        rclone_password.save_credentials_password(
-            password_filepath,
-        )
-
-        rclone_password.set_rclone_password(
-            password_filepath, rclone.get_full_config_filepath(self.cfg)
-        )
-
-        self.cfg.rclone_has_password[connection_method] = (
-            True  # HANDLE THIS PROPERLY
-        )
-        self.cfg.save_rclone_password_state()
-        print(self.cfg.rclone_has_password[connection_method])
-
-    def remove_rclone_password(self):
-        """"""
-        # TODO: CHECK CONNECTION METHOD
-        connection_method = self.cfg["connection_method"]
-
-        if not self.cfg.rclone_has_password[self.cfg["connection_method"]]:
-            raise RuntimeError(
-                f"The config for the current connection method: {self.cfg['connection_method']} does not have a password. Cannot remove."
-            )
-        config_filepath = rclone_password.get_password_filepath(self.cfg)
-        rclone_password.remove_rclone_password(
-            config_filepath, rclone.get_full_config_filepath(self.cfg)
-        )
-
-        self.cfg.rclone_has_password[connection_method] = (
-            False  # HANDLE THIS PROPERLY
-        )
-        self.cfg.save_rclone_password_state()
-
-        print("SAY SOMETHING LIKE PASSWORD REMOVED")
 
     # -------------------------------------------------------------------------
     # Public Folder Makers
@@ -890,6 +805,14 @@ class DataShuttle:
[…]
