### 2025-10-02 19872ee
Refactoring.
