+++ datashuttle/utils/utils.py
@@ -45,13 +45,15 @@ def log_and_message(message: str, use_rich: bool = False) -> None:
     print_message_to_user(message, use_rich)
 
 
-def log_and_raise_error(message: str, exception: Any) -> None:
+def log_and_raise_error(
+    message: str, exception: Any, from_error: Exception | None = None
+) -> None:
     """Log the message before raising the same message as an error."""
     if ds_logger.logging_is_active():
         logger = ds_logger.get_logger()
         logger.error(f"\n\n{' '.join(traceback.format_stack(limit=5))}")
         logger.error(message)
-    raise_error(message, exception)
+    raise_error(message, exception, from_error=from_error)
 
 
 def warn(message: str, log: bool) -> None:
@@ -72,14 +74,20 @@ def warn(message: str, log: bool) -> None:
     warnings.warn(message)
 
 
-def raise_error(message: str, exception) -> None:
+def raise_error(
+    message: str, exception, from_error: Exception | None = None
+) -> None:
     """Centralized way to raise an error.
 
     The logger is closed to ensure it is not still running
     if a function call raises an exception in a python environment.
     """
     ds_logger.close_log_filehandler()
-    raise exception(message)
+
+    if from_error:
+        raise exception(message) from from_error
+    else:
+        raise exception(message)
 
 
 def print_message_to_user(
