### 2025-10-02 274fe29
Refactoring more.
