### 2025-10-02 ebf073a
Refactoring more.
