### 2025-10-03 0ddb981
Tidy up rclone_password.py
