+++ datashuttle/utils/rclone.py
@@ -11,6 +11,7 @@ import platform
 import shlex
 import subprocess
 import tempfile
+from pathlib import Path
 from subprocess import CompletedProcess
 
 from datashuttle.configs import canonical_configs
@@ -114,7 +115,6 @@ def call_rclone_through_script_for_central_connection(
 
 
 def call_rclone_with_popen_for_central_connection(
-    cfg,
     command: str,
 ) -> subprocess.Popen:
     """Call rclone using `subprocess.Popen` for control over process termination.
@@ -126,18 +126,14 @@ def call_rclone_with_popen_for_central_connection(
     process explicitly.
     """
     command = "rclone " + command
-    lambda_func = lambda: subprocess.Popen(
+    process = subprocess.Popen(
         shlex.split(command), stdout=subprocess.PIPE, stderr=subprocess.PIPE
-    )
-
-    process = run_function_that_may_require_central_connection_password(
-        cfg, lambda_func
     )
     return process
 
 
 def await_call_rclone_with_popen_for_central_connection_raise_on_fail(
-    cfg, process: subprocess.Popen, log: bool = True
+    process: subprocess.Popen, log: bool = True
 ):
     """Await rclone the subprocess.Popen call.
 
@@ -161,11 +157,10 @@ def run_function_that_may_require_central_connection_password(
     cfg, lambda_func
 ):
     """ """
-    set_password = cfg.rclone_has_password[cfg["connection_method"]]
+    set_password = cfg.connection_method_rclone_config_has_password()
 
     if set_password:
-        config_filepath = rclone_password.get_password_filepath(cfg)
-        rclone_password.set_credentials_as_password_command(config_filepath)
+        rclone_password.set_credentials_as_password_command(cfg)
 
     results = lambda_func()
 
@@ -244,8 +239,9 @@ def setup_rclone_config_for_ssh(
     """
     key_escaped = private_key_str.replace("\n", "\\n")
 
-    rclone_config_filepath = cfg.get_rclone_config_filepath()
-    # TODO: do this for everything TODO: maybe this config file can be created before setup in case of old file
+    rclone_config_filepath = get_full_config_filepath(
+        cfg
+    )  # TODO: do this for everything TODO: maybe this config file can be created before setup in case of old file
     if rclone_config_filepath.exists():
         rclone_config_filepath.unlink()
 
@@ -265,12 +261,23 @@ def setup_rclone_config_for_ssh(
         log_rclone_config_output()
 
 
+def get_config_path():
+    """TODO PLACEHOLDER."""
+    return (
+        Path().home() / "AppData" / "Roaming" / "rclone"
+    )  #  # "$HOME/.config/rclone/rclone.conf")
+
+
+def get_full_config_filepath(cfg: Configs) -> Path:
+    return get_config_path() / f"{cfg.get_rclone_config_name()}.conf"
+
+
 def get_config_arg(cfg):
     """TODO PLACEHOLDER."""
     cfg.get_rclone_config_name()  # pass this? handle better...
 
     if cfg["connection_method"] in ["aws", "gdrive", "ssh"]:
-        return f'--config "{cfg.get_rclone_config_filepath()}"'
+        return f'--config "{get_full_config_filepath(cfg)}"'
     else:
         return ""
 
@@ -334,7 +341,6 @@ def setup_rclone_config_for_gdrive(
     )
 
     process = call_rclone_with_popen_for_central_connection(
-        cfg,
         f"config create "
         f"{rclone_config_name} "
         f"drive "
@@ -343,13 +349,13 @@ def setup_rclone_config_for_gdrive(
         f"scope drive "
         f"root_folder_id {cfg['gdrive_root_folder_id']} "
         f"{extra_args} "
-        f"{get_config_arg(cfg)}",
[…]
