### 2025-10-04 a4b2aa1
Editing setup_ssh.
