### 2025-10-04 2d3429e
Adding password to SSH GUI.
