### 2025-10-04 b32a979
Adding password to SSH GUI.
