### 2025-10-06 d5ae46d
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
