### 2025-10-06 c2b6e19
Rough version working on all screens.
