### 2025-10-06 05f573b
Rough version working on all screens.
