### 2025-10-06 6a12c54
Rough version working on all screens2.
