+++ .pre-commit-config.yaml
@@ -31,12 +31,12 @@ repos:
         - id: rst-directive-colons
         - id: rst-inline-touching-normal
     - repo: https://github.com/astral-sh/ruff-pre-commit
-      rev: v0.12.11
+      rev: v0.13.3
       hooks:
         - id: ruff
         - id: ruff-format
     - repo: https://github.com/pre-commit/mirrors-mypy
-      rev: v1.17.1
+      rev: v1.18.2
       hooks:
           - id: mypy
             additional_dependencies:
