+++ datashuttle/configs/config_class.py
@@ -1,12 +1,11 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Dict, Optional, Union, cast
+from typing import TYPE_CHECKING, Dict, Union, cast
 
 if TYPE_CHECKING:
     from collections.abc import ItemsView, KeysView, ValuesView
 
     from datashuttle.utils.custom_types import (
-        OverwriteExistingFiles,
         TopLevelFolder,
     )
 
@@ -20,6 +19,7 @@ from datashuttle.configs import (
     canonical_configs,
     canonical_folders,
     load_configs,
+    rclone_configs,
 )
 from datashuttle.utils import folders, utils
 
@@ -36,6 +36,9 @@ class Configs(UserDict):
         self, project_name: str, file_path: Path, input_dict: Union[dict, None]
     ) -> None:
         """Initialize the Configs class with project name, file path, and config dictionary.
+
+        This class also holds `RCloneConfigs` under the `.rclone` attribute to manage
+        the configs
 
         Parameters
         ----------
@@ -64,46 +67,7 @@ class Configs(UserDict):
         self.hostkeys_path: Path
         self.project_metadata_path: Path
 
-        self.rclone_password_state_file_path = (
-            self.file_path.parent / "rclone_ps_state.yaml"
-        )
-
-    def load_rclone_has_password(self):
-        assert self["connection_method"] in ["ssh", "aws", "gdrive"]
-
-        if self.rclone_password_state_file_path.is_file():
-            with open(self.rclone_password_state_file_path, "r") as file:
-                rclone_has_password = yaml.full_load(file)
-        else:
-            rclone_has_password = {
-                "ssh": False,
-                "gdrive": False,
-                "aws": False,
-            }
-
-            with open(self.rclone_password_state_file_path, "w") as file:
-                yaml.dump(rclone_has_password, file)
-
-        return rclone_has_password
-
-    def get_rclone_has_password(
-        self,
-    ):  # TODO: hmm this is used a lot... could hold state.. but nice to save...
-        """"""
-        rclone_has_password = self.load_rclone_has_password()
-
-        return rclone_has_password[self["connection_method"]]
-
-    def set_rclone_has_password(self, value):
-        """"""
-        assert self["connection_method"] in ["ssh", "aws", "gdrive"]
-
-        rclone_has_password = self.load_rclone_has_password()
-
-        rclone_has_password[self["connection_method"]] = value
-
-        with open(self.rclone_password_state_file_path, "w") as file:
-            yaml.dump(rclone_has_password, file)
+        self.rclone = rclone_configs.RCloneConfigs(self, self.file_path.parent)
 
     def setup_after_load(self) -> None:
         """Set up the config after loading it."""
@@ -286,51 +250,6 @@ class Configs(UserDict):
                 base_folder = self["central_path"] / top_level_folder
 
         return base_folder
-
-    def get_rclone_config_name(
-        self, connection_method: Optional[str] = None
-    ) -> str:
-        """Generate the rclone configuration name for the central project.
-
-        These configs are created by datashuttle but managed and stored by rclone.
-        """
-        if connection_method is None:
-            connection_method = self["connection_method"]
-
-        return f"central_{self.project_name}_{connection_method}"
-
-    def get_rclone_config_filepath(self) -> Path:
-        """"""
-        return (
-            canonical_folders.get_rclone_config_base_path()
-            / f"{self.get_rclone_config_name()}.conf"
-        )
-
-    def make_rclone_transfer_options(
-        self, overwrite_existing_files: OverwriteExistingFiles, dry_run: bool
-    ) -> Dict:
-        """Create a dictionary of rclone transfer options.
-
-        Originally these arguments were collected from configs, but now
[…]
