### 2025-10-07 e1395bc
Finished refactor of configs.
