+++ datashuttle/datashuttle_class.py
@@ -808,6 +808,11 @@ class DataShuttle:
         cluster. Once input, SSH private / public key pair
         will be setup.
         """
+        if self.cfg["connection_method"] != "ssh":
+            raise RuntimeError(
+                "configs `connection_method` must be 'ssh' to set up SSH connection."
+            )
+
         self._start_log(
             "setup-ssh-connection-to-central-server", local_vars=locals()
         )
@@ -828,7 +833,7 @@ class DataShuttle:
                 f"{self.cfg.rclone.get_rclone_config_filepath()}.\n\n"
             )
 
-            if not self.cfg.get_rclone_has_password():
+            if not self.cfg.rclone.get_rclone_has_password():
                 self._try_set_rclone_password()
 
             rclone.check_successful_connection_and_raise_error_on_fail(
@@ -861,6 +866,11 @@ class DataShuttle:
         Next, with the provided credentials, the final setup will be done. This
         opens up a browser if the user confirmed access to a browser.
         """
+        if self.cfg["connection_method"] != "gdrive":
+            raise RuntimeError(
+                "configs `connection_method` must be 'gdrive' to set up Google Drive connection."
+            )
+
         self._start_log(
             "setup-google-drive-connection-to-central-server",
             local_vars=locals(),
@@ -887,14 +897,12 @@ class DataShuttle:
             gdrive_client_secret, config_token
         )
 
-        print("got process")
-
         # TODO: do something with stderr stdout here, in general handle errors better!
         rclone.await_call_rclone_with_popen_for_central_connection_raise_on_fail(
             self.cfg, process, log=True
         )
 
-        if not self.cfg.get_rclone_has_password():
+        if not self.cfg.rclone.get_rclone_has_password():
             self._try_set_rclone_password()
 
         rclone.check_successful_connection_and_raise_error_on_fail(self.cfg)
@@ -918,6 +926,12 @@ class DataShuttle:
 
         Next, with the provided credentials, the final connection setup will be done.
         """
+        if self.cfg["connection_method"] != "aws":
+            raise RuntimeError(
+                "configs `connection_method` must be 'aws' to "
+                "set up Amazon Web Services S3 Bucket connection."
+            )
+
         self._start_log(
             "setup-aws-connection-to-central-server",
             local_vars=locals(),
@@ -927,7 +941,7 @@ class DataShuttle:
 
         self._setup_rclone_aws_config(aws_secret_access_key, log=True)
 
-        if not self.cfg.get_rclone_has_password():
+        if not self.cfg.rclone.get_rclone_has_password():
             self._try_set_rclone_password()
 
         rclone.check_successful_connection_and_raise_error_on_fail(self.cfg)
@@ -982,7 +996,7 @@ class DataShuttle:
 
     def set_rclone_password(self):
         """"""
-        if self.cfg.get_rclone_has_password():
+        if self.cfg.rclone.get_rclone_has_password():
             raise RuntimeError(
                 "This config file already has a password set. "
                 "First, use `remove_rclone_password` to remove it."
@@ -994,7 +1008,7 @@ class DataShuttle:
 
     def remove_rclone_password(self):
         """"""
-        if not self.cfg.get_rclone_has_password():
+        if not self.cfg.rclone.get_rclone_has_password():
             raise RuntimeError(
                 f"The config for the current connection method: {self.cfg['connection_method']} "
                 f"does not have a password. Cannot remove."
