### 2025-10-07 b13a984
Finished refactor of configs.
