### 2025-10-07 d360348
Add doc to setup_gdrive.
