+++ datashuttle/utils/rclone.py
@@ -156,7 +156,7 @@ def await_call_rclone_with_popen_for_central_connection_raise_on_fail(
         utils.log_and_raise_error(stderr.decode("utf-8"), ConnectionError)
 
     if log:
-        log_rclone_config_output()
+        log_rclone_config_output(cfg)
 
     return stdout, stderr
 
@@ -214,7 +214,7 @@ def setup_rclone_config_for_local_filesystem(
     call_rclone(f"config create {rclone_config_name} local", pipe_std=True)
 
     if log:
-        log_rclone_config_output()
+        log_rclone_config_output(cfg)
 
 
 def setup_rclone_config_for_ssh(
@@ -247,12 +247,7 @@ def setup_rclone_config_for_ssh(
     """
     key_escaped = private_key_str.replace("\n", "\\n")
 
-    rclone_config_filepath = get_full_config_filepath(
-        cfg
-    )  # TODO: do this for everything TODO: maybe this config file can be created before setup in case of old file
-    if rclone_config_filepath.exists():
-        rclone_config_filepath.unlink()
-        cfg.rclone.set_rclone_has_password(False)
+    delete_existing_rclone_config_file(cfg)
 
     command = (
         f"config create "
@@ -267,26 +262,37 @@ def setup_rclone_config_for_ssh(
     call_rclone(command, pipe_std=True)
 
     if log:
-        log_rclone_config_output()
-
-
-def get_config_path():
-    """TODO PLACEHOLDER."""
-    return (
-        Path().home() / "AppData" / "Roaming" / "rclone"
-    )  #  # "$HOME/.config/rclone/rclone.conf")
+        log_rclone_config_output(cfg)
 
 
 def get_full_config_filepath(cfg: Configs) -> Path:
-    return get_config_path() / f"{cfg.rclone.get_rclone_config_name()}.conf"
+    """ """
+    return (
+        canonical_folders.get_rclone_config_base_path()
+        / f"{cfg.rclone.get_rclone_config_name()}.conf"
+    )
+
+
+def delete_existing_rclone_config_file(cfg: Configs):
+    """ """
+    rclone_config_filepath = (
+        cfg.rclone.get_rclone_central_connection_config_filepath()
+    )
+
+    if delete_existing_file:
+        if rclone_config_filepath.exists():
+            rclone_config_filepath.unlink()
+            cfg.rclone.set_rclone_has_password(False)
 
 
 def get_config_arg(cfg):
     """TODO PLACEHOLDER."""
-    cfg.rclone.get_rclone_config_name()  # pass this? handle better...
+    rclone_config_path = (
+        cfg.rclone.get_rclone_central_connection_config_filepath()
+    )
 
     if cfg["connection_method"] in ["aws", "gdrive", "ssh"]:
-        return f'--config "{get_full_config_filepath(cfg)}"'
+        return f'--config "{rclone_config_path}"'
     else:
         return ""
 
@@ -349,12 +355,7 @@ def setup_rclone_config_for_gdrive(
         else ""
     )
 
-    rclone_config_filepath = get_full_config_filepath(
-        cfg
-    )  # TODO: do this for everything TODO: maybe this config file can be created before setup in case of old file
-    if rclone_config_filepath.exists():
-        rclone_config_filepath.unlink()
-        cfg.rclone.set_rclone_has_password(False)
+    delete_existing_rclone_config_file(cfg)
 
     command = (
         f"config create "
@@ -408,12 +409,7 @@ def setup_rclone_config_for_aws(
         else f" location_constraint {aws_region}"
     )
 
-    rclone_config_filepath = get_full_config_filepath(
-        cfg
-    )  # TODO: do this for everything TODO: maybe this config file can be created before setup in case of old file
[…]
