+++ datashuttle/tui/screens/setup_gdrive.py
@@ -27,8 +27,6 @@ class SetupGdriveScreen(ModalScreen):
     to enter a client secret. If the user has access to a browser, a Google Drive
     authentication page will open. Otherwise, the user is asked to run a rclone command
     and input a config token.
-
-
     """
 
     def __init__(self, interface: Interface) -> None:
@@ -36,7 +34,7 @@ class SetupGdriveScreen(ModalScreen):
         super(SetupGdriveScreen, self).__init__()
 
         self.interface = interface
-        self.stage: int = 0
+        self.no_browser_stage: None | str = "show_command_to_generate_code"
         self.setup_worker: Worker | None = None
         self.is_browser_available: bool = True
         self.gdrive_client_secret: Optional[str] = None
@@ -73,20 +71,24 @@ class SetupGdriveScreen(ModalScreen):
 
         This dialog window operates using 6 buttons:
 
-        1) "ok" button : Starts the connection setup process.
+        1) `setup_gdrive_ok_button` : Starts the connection setup process.
 
-        2) "yes" button : A "yes" answer to the availability of browser question. On click,
-            if "gdrive_client_id" is present in configs, the user is asked for client secret
+        2) `setup_gdrive_has_browser_yes_button`  : A "yes" answer to the availability of browser question.
+            On click, if "gdrive_client_id" is present in configs, the user is asked for client secret
             and proceeds to a browser authentication.
 
-        3) "no" button : A "no" answer to the availability of browser question. On click,
+        3) `setup_gdrive_no_button`  : A "no" answer to the availability of browser question. On click,
             prompts the user to enter a config token by running an rclone command.
 
-        4) "enter" button : To enter the client secret or config token.
+        4) `setup_gdrive_no_browser_enter_button` : To enter the client secret or config token.
 
-        5) "finish" button : To finish the setup.
+        5) `setup_gdrive_set_password_yes_button` : To set a password on the RClone config file
 
-        6) "cancel" button : To cancel the setup at any step before completion.
+        6) `setup_gdrive_set_password_no_button` : To skip setting a password on the RClone config file
+
+        7) `setup_gdrive_finish_button` button : To finish the setup.
+
+        8) "`setup_gdrive_cancel_button` : To cancel the setup at any step before completion.
         """
         if (
             event.button.id == "setup_gdrive_cancel_button"
@@ -94,7 +96,7 @@ class SetupGdriveScreen(ModalScreen):
         ):
             # see setup_gdrive_connection_and_update_ui()
             if self.setup_worker and self.setup_worker.is_running:
-                self.setup_worker.cancel()  # fix
+                self.setup_worker.cancel()
                 self.interface.terminate_gdrive_setup()
             self.dismiss()
 
@@ -107,27 +109,29 @@ class SetupGdriveScreen(ModalScreen):
                 self.ask_user_for_browser()
 
         elif event.button.id == "setup_gdrive_has_browser_yes_button":
-            self.remove_yes_no_browser_buttons()
+            self.query_one("#setup_gdrive_has_browser_yes_button").remove()
+            self.query_one("#setup_gdrive_no_button").remove()
             self.open_browser_and_setup_gdrive_connection(
                 self.gdrive_client_secret
             )
 
         elif event.button.id == "setup_gdrive_no_button":
             self.is_browser_available = False
-            self.remove_yes_no_browser_buttons()
+            self.query_one("#setup_gdrive_has_browser_yes_button").remove()
+            self.query_one("#setup_gdrive_no_button").remove()
             self.prompt_user_for_config_token()
 
         elif event.button.id == "setup_gdrive_no_browser_enter_button":
             if (
                 self.interface.project.cfg["gdrive_client_id"]
-                and self.stage == 0
+                and self.no_browser_stage == "show_command_to_generate_code"
[…]
