### 2025-10-07 860be41
Apply PR feedback and refactor tests.
