+++ tests/tests_transfers/ssh/test_ssh_suggest_next.py
@@ -1,5 +1,4 @@
 import platform
-import shutil
 
 import pytest
 
@@ -43,11 +42,4 @@ class TestSSHDriveSuggestNext(BaseSSHTransfer, TuiBase):
         """ """
         project = ssh_setup
 
-        test_utils.make_local_folders_with_files_in(
-            project, "rawdata", "sub-001", ["ses-001", "ses-002"]
-        )
-        project.upload_entire_project()
-
-        shutil.rmtree(project.get_local_path())
-
-        await self.check_next_sub_002_ses_003_in_tui(project)
+        await self.check_next_sub_ses_in_tui(project)
