### 2025-10-07 c531bb8
Move check_next_sub_ses_in_tui to tui base.
