+++ tests/tests_transfers/base_transfer.py
@@ -7,7 +7,6 @@ from pathlib import Path
 import pandas as pd
 import pytest
 
-from datashuttle.tui.app import TuiApp
 from datashuttle.utils import rclone
 
 from .. import test_utils
@@ -257,64 +256,3 @@ class BaseTransfer(BaseTest):
         local_path location in the test environment.
         """
         project.get_logging_path().mkdir(parents=True, exist_ok=True)
-
-    async def check_next_sub_ses_in_tui(self, project):
-        """A central function for testing next sub / ses in the TUI.
-
-        This test is shared between ssh, aws and gdrive tests that
-        use the same logic to test the get next sub / ses functionality.
-
-        First, sub / ses folders are created in the project and uploaded
-        centrally. Then, the folders are removed  from the local path.
-        In this way, we can be use that `include_central` (which searches
-        the remote for the next sub / ses) is behaving correctly and not just
-        reading the local path.
-
-        Because sub-001 is created, the suggested sub we expect is sub-002.
-        Because ses-002 is created in sub-001, the suggested ses we expect is ses-003.
-        """
-        test_utils.make_local_folders_with_files_in(
-            project, "rawdata", "sub-001", ["ses-001", "ses-002"]
-        )
-        project.upload_entire_project()
-
-        shutil.rmtree(project.get_local_path())
-
-        app = TuiApp()
-        async with app.run_test(size=self.tui_size()) as pilot:
-            await self.check_and_click_onto_existing_project(
-                pilot, project.project_name
-            )
-
-            # Turn on the central checkbox
-            await self.scroll_to_click_pause(
-                pilot, "#create_folders_settings_button"
-            )
-            await self.scroll_to_click_pause(
-                pilot, "#suggest_next_sub_ses_central_checkbox"
-            )
-            await self.scroll_to_click_pause(
-                pilot, "#create_folders_settings_close_button"
-            )
-
-            await self.fill_input(
-                pilot, "#create_folders_subject_input", "sub-001"
-            )
-
-            await self.double_click_input(pilot, "ses")
-
-            assert (
-                pilot.app.screen.query_one(
-                    "#create_folders_session_input"
-                ).value
-                == "ses-003"
-            )
-
-            await self.double_click_input(pilot, "sub")
-
-            assert (
-                pilot.app.screen.query_one(
-                    "#create_folders_subject_input"
-                ).value
-                == "sub-002"
-            )
