### 2025-10-08 424e1c6
Extend the gdrive setup message.
