### 2025-10-08 73995eb
Rework `_try_set_rclone_password`
