### 2025-10-08 631a2f5
Rework `_try_set_rclone_password`
