### 2025-10-08 611a440
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
