### 2025-10-09 5cf30aa
Fix message asking to use security, add try / except to async
