### 2025-10-09 cb89fe0
Fix message asking to use security, add try / except to async
