### 2025-10-09 e53259b
Add backward compatability message.
