+++ datashuttle/utils/rclone.py
@@ -14,6 +14,8 @@ import tempfile
 from pathlib import Path
 from subprocess import CompletedProcess
 
+from packaging import version
+
 from datashuttle.configs import canonical_configs
 from datashuttle.utils import rclone_password, utils
 
@@ -162,6 +164,24 @@ def run_function_that_may_require_central_connection_password(
     cfg, lambda_func
 ):
     """ """
+    from datashuttle import get_datashuttle_version  # avoid circular import
+
+    rclone_config_filepath = (
+        cfg.rclone.get_rclone_central_connection_config_filepath()
+    )
+
+    if not rclone_config_filepath.is_file():
+        if version.parse(get_datashuttle_version()) <= version.parse("0.7.1"):
+            raise RuntimeError(
+                f"The way RClone configs are managed has changed since version v0.7.1\n"
+                f"Please set up the {cfg['connection_method']} connection again."
+            )
+        else:
+            raise RuntimeError(
+                f"An unexpected error occurred. Could not find the rclone config file at: {rclone_config_filepath}\n"
+                f"Please set up the {cfg['connection_method']} connection again."
+            )
+
     set_password = cfg.rclone.get_rclone_has_password()
 
     if set_password:
@@ -262,6 +282,15 @@ def setup_rclone_config_for_ssh(
     if log:
         log_rclone_config_output(cfg)
 
+def delete_existing_rclone_config_file(cfg: Configs):
+    """ """
+    rclone_config_filepath = (
+        cfg.rclone.get_rclone_central_connection_config_filepath()
+    )
+
+    if rclone_config_filepath.exists():
+        rclone_config_filepath.unlink()
+        cfg.rclone.set_rclone_has_password(False)
 
 def setup_rclone_config_for_gdrive(
     cfg: Configs,
@@ -440,8 +469,8 @@ def check_successful_connection_and_raise_error_on_fail(cfg: Configs) -> None:
         )
 
 
-def log_rclone_config_output(cfg: Configs) -> None:
-    """Log the output from creating Rclone config."""
+def get_rclone_config_filepath(cfg: Configs) -> Path:
+    """"""
     if cfg["connection_method"] in ["aws", "ssh", "gdrive"]:
         config_filepath = (
             cfg.rclone.get_rclone_central_connection_config_filepath()
@@ -450,6 +479,12 @@ def log_rclone_config_output(cfg: Configs) -> None:
         output = call_rclone("config file", pipe_std=True)
         config_filepath = output.stdout.decode("utf-8")
 
+    return config_filepath
+
+
+def log_rclone_config_output(cfg: Configs) -> None:
+    """Log the output from creating Rclone config."""
+    config_filepath = get_rclone_config_filepath(cfg)
     utils.log(f"Successfully created rclone config. {config_filepath}")
 
 
