### 2025-10-09 f3c945b
Add backward compatability message.
