+++ datashuttle/utils/rclone.py
@@ -199,7 +199,7 @@ def run_function_that_may_require_central_connection_password(
 
 
 # -----------------------------------------------------------------------------
-# Setup
+# RClone Configs
 # -----------------------------------------------------------------------------
 
 
@@ -283,29 +283,6 @@ def setup_rclone_config_for_ssh(
 
     if log:
         log_rclone_config_output(cfg)
-
-
-def delete_existing_rclone_config_file(cfg: Configs):
-    """ """
-    rclone_config_filepath = (
-        cfg.rclone.get_rclone_central_connection_config_filepath()
-    )
-
-    if rclone_config_filepath.exists():
-        rclone_config_filepath.unlink()
-        cfg.rclone.set_rclone_has_password(False)
-
-
-def get_config_arg(cfg):
-    """TODO PLACEHOLDER."""
-    rclone_config_path = (
-        cfg.rclone.get_rclone_central_connection_config_filepath()
-    )
-
-    if cfg["connection_method"] in ["aws", "gdrive", "ssh"]:
-        return f'--config "{rclone_config_path}"'
-    else:
-        return ""
 
 
 def setup_rclone_config_for_gdrive(
@@ -432,6 +409,29 @@ def setup_rclone_config_for_aws(
 
     if log:
         log_rclone_config_output(cfg)
+
+
+def delete_existing_rclone_config_file(cfg: Configs):
+    """ """
+    rclone_config_filepath = (
+        cfg.rclone.get_rclone_central_connection_config_filepath()
+    )
+
+    if rclone_config_filepath.exists():
+        rclone_config_filepath.unlink()
+        cfg.rclone.set_rclone_has_password(False)
+
+
+def get_config_arg(cfg):
+    """TODO PLACEHOLDER."""
+    rclone_config_path = (
+        cfg.rclone.get_rclone_central_connection_config_filepath()
+    )
+
+    if cfg["connection_method"] in ["aws", "gdrive", "ssh"]:
+        return f'--config "{rclone_config_path}"'
+    else:
+        return ""
 
 
 def check_successful_connection_and_raise_error_on_fail(cfg: Configs) -> None:
