### 2025-10-13 9cca7f2
Update tests/tests_transfers/ssh/test_ssh_suggest_next.py
Co-authored-by: Shrey Singh <96627769+cs7-shrey@users.noreply.github.com>
