+++ tests/tests_transfers/ssh/test_ssh_suggest_next.py
@@ -24,7 +24,7 @@ class TestSSHDriveSuggestNext(BaseSSHTransfer, TuiBase):
     )
     def ssh_setup(self, setup_project_paths, setup_ssh_container_fixture):
         """
-        Setup pathtable and project for GDrive transfer tests.
+        Setup pathtable and project for SSH transfer tests.
         """
         project = test_utils.make_project(setup_project_paths["project_name"])
         ssh_test_utils.setup_project_for_ssh(
