### 2025-10-13 3b7915c
Add docstrings to datashuttle class.
