+++ datashuttle/tui/interface.py
@@ -636,7 +636,7 @@ class Interface:
     def try_setup_rclone_password(self):
         """"""
         try:
-            self.project._try_set_rclone_password()
+            self.project._try_encrypt_rclone_config()
             return True, None
         except BaseException as e:
             return False, str(e)
