### 2025-10-13 a2246c2
Begin adding docstrings and type hints.
