+++ datashuttle/utils/rclone_encryption.py
@@ -1,3 +1,7 @@
+"""Module for encrypthing the RClone config file. Methods based on:
+https://rclone.org/docs/#configuration-encryption
+"""
+
 from __future__ import annotations
 
 from typing import TYPE_CHECKING
@@ -9,13 +13,14 @@ import os
 import platform
 import shutil
 import subprocess
+from pathlib import Path
 
 from datashuttle.configs import canonical_folders
 from datashuttle.utils import utils
 
 
-def save_credentials_password(cfg):
-    """"""
+def save_credentials_password(cfg: Configs) -> None:
+    """Use the system password manager to set up a password for the Rclone config file encryption."""
     if platform.system() == "Windows":
         set_password_windows(cfg)
     elif platform.system() == "Linux":
@@ -24,9 +29,16 @@ def save_credentials_password(cfg):
         set_password_macos(cfg)
 
 
-def set_password_windows(cfg: Configs):
-    """"""
-    password_filepath = get_password_filepath(cfg)
+def set_password_windows(cfg: Configs) -> None:
+    """Generate and securely store a random password in a Windows Credential XML file.
+
+    Use PowerShell to create a random password associated with the name 'rclone'.
+    The password is stored as a PowerShell `PSCredential` object that can only
+    be decrypted by the same Windows user account that created it.
+
+    This password is later used to encrypt the Rclone config file.
+    """
+    password_filepath = get_windows_password_filepath(cfg)
 
     if password_filepath.exists():
         password_filepath.unlink()
@@ -57,8 +69,16 @@ def set_password_windows(cfg: Configs):
         )
 
 
-def set_password_linux(cfg):
-    """"""
+def set_password_linux(cfg: Configs) -> None:
+    """Generate and securely store a random password using the Linux `pass` utility.
+
+    This function generates a random password and stores it in the user's
+    GPG-encrypted password store via the `pass` command-line tool.
+
+    The `pass` utility must be installed and initialized with a GPG ID on the
+    current user account (via `pass init <gpg-id>`). If it is not initialized,
+    a RuntimeError will be raised.
+    """
     output = subprocess.run(
         "pass --help",
         shell=True,
@@ -104,8 +124,15 @@ def set_password_linux(cfg):
         )
 
 
-def set_password_macos(cfg: Configs):
-    """"""
+def set_password_macos(cfg: Configs) -> None:
+    """Generate and store a password using the macOS Keychain.
+
+    This function generates a random password and stores it in the macOS Keychain
+    using the built-in `security` command-line tool.
+
+    The password is generated using OpenSSL with 40 random base64 characters and
+    is securely saved to the user's login Keychain.
+    """
     output = subprocess.run(
         f"security add-generic-password -a datashuttle -s {cfg.rclone.get_rclone_config_name()} -w $(openssl rand -base64 40) -U",
         shell=True,
@@ -121,10 +148,22 @@ def set_password_macos(cfg: Configs):
         )
 
 
-def set_credentials_as_password_command(cfg):
-    """"""
+def set_credentials_as_password_command(cfg: Configs) -> None:
+    """Configure the RClone password retrieval command based on the operating system.
+
+    This function sets the `RCLONE_PASSWORD_COMMAND` environment variable so that
+    RClone can securely retrieve stored credentials
+
+    - Windows : Uses PowerShell to decrypt a previously exported `PSCredential`
+      object from the `.clixml` file created by `set_password_windows()`.
+    - Linux : Uses the `pass` command-line utility to fetch the stored password
+      from the user's GPG-encrypted password store.
+    - macOS : Uses the built-in `security` tool to read the password
[…]
