### 2025-10-13 c7d3fb1
Add docstrings and type hints.
