+++ datashuttle/tui/screens/setup_gdrive.py
@@ -338,7 +338,7 @@ class SetupGdriveScreen(ModalScreen):
     # ----------------------------------------------------------------------------------
 
     def set_rclone_encryption(self):
-        """"""
+        """Try and encrypt the Rclone config file and inform the user of success / failure."""
         success, output = self.interface.try_setup_rclone_encryption()
 
         if success:
