+++ datashuttle/tui/interface.py
@@ -634,7 +634,7 @@ class Interface:
     # ------------------------------------------------------------------------------------
 
     def try_setup_rclone_encryption(self):
-        """"""
+        """Try and encrypt the RClone config file for the current `connection_method`."""
         try:
             self.project._try_encrypt_rclone_config()
             return True, None
