### 2025-10-18 1c18f9d
Add mypy configuration file
Configure mypy settings for type checking.
