+++ mypy.ini
@@ -1,0 +1,13 @@
+[mypy]
+# Use Python 3.12 for type checking
+python_version = 3.12
+
+# Don’t show import errors yet (makes setup easier)
+ignore_missing_imports = True
+
+# Show clearer error messages
+show_error_codes = True
+warn_unused_ignores = True
+
+# Tell mypy which folders to check
+files = datashuttle, tests
