### 2025-10-19 50f2e11
docs: add Troubleshooting / FAQ page (#518)
