+++ docs/source/troubleshooting.rst
@@ -1,0 +1,6 @@
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents
+
+   get_started
+   usage
