+++ docs/source/index.rst
@@ -1,0 +1,8 @@
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents
+
+   get_started
+   usage
+
+   troubleshooting
