### 2025-10-20 4ef3d58
Add more suggest next sub ses tests (#597)
* Fix get next sub / ses on TUI and extend AWS, SSH, GDrive tests.

* Also remove aws folder.

* Apply PR feedback and refactor tests.

* Move check_next_sub_ses_in_tui to tui base.

* Update tests/tests_transfers/ssh/test_ssh_suggest_next.py

Co-authored-by: Shrey Singh <96627769+cs7-shrey@users.noreply.github.com>

* Update tests/tests_transfers/aws/test_aws_suggest_next.py

Co-authored-by: Shrey Singh <96627769+cs7-shrey@users.noreply.github.com>

---------

Co-authored-by: Shrey Singh <96627769+cs7-shrey@users.noreply.github.com>
