### 2025-10-20 447b2a1
Improve rclone transfer output detail.
