+++ datashuttle/datashuttle_class.py
@@ -315,7 +315,7 @@ class DataShuttle:
         overwrite_existing_files: OverwriteExistingFiles = "never",
         dry_run: bool = False,
         init_log: bool = True,
-    ) -> None:
+    ) -> list[str]:
         """Upload data from a local project to the central project folder.
 
         Parameters
@@ -360,7 +360,7 @@ class DataShuttle:
 
         self._check_top_level_folder(top_level_folder)
 
-        TransferData(
+        errors = TransferData(
             self.cfg,
             "upload",
             top_level_folder,
@@ -369,11 +369,12 @@ class DataShuttle:
             datatype,
             overwrite_existing_files,
             dry_run,
-            log=True,
-        )
+        ).run()
 
         if init_log:
             ds_logger.close_log_filehandler()
+
+        return errors
 
     @check_configs_set
     @check_is_not_local_project
@@ -386,7 +387,7 @@ class DataShuttle:
         overwrite_existing_files: OverwriteExistingFiles = "never",
         dry_run: bool = False,
         init_log: bool = True,
-    ) -> None:
+    ) -> list[str]:
         """Download data from the central project to the local project folder.
 
         Parameters
@@ -431,7 +432,7 @@ class DataShuttle:
 
         self._check_top_level_folder(top_level_folder)
 
-        TransferData(
+        errors = TransferData(
             self.cfg,
             "download",
             top_level_folder,
@@ -440,11 +441,12 @@ class DataShuttle:
             datatype,
             overwrite_existing_files,
             dry_run,
-            log=True,
-        )
+        ).run()
 
         if init_log:
             ds_logger.close_log_filehandler()
+
+        return errors
 
     # Specific top-level folder
     # ----------------------------------------------------------------------------------
@@ -457,7 +459,7 @@ class DataShuttle:
         self,
         overwrite_existing_files: OverwriteExistingFiles = "never",
         dry_run: bool = False,
-    ) -> None:
+    ) -> list[str]:
         """Upload all files in the `rawdata` top level folder.
 
         Parameters
@@ -474,7 +476,7 @@ class DataShuttle:
             transfer was taking place, but no files will be moved.
 
         """
-        self._transfer_top_level_folder(
+        return self._transfer_top_level_folder(
             "upload",
             "rawdata",
             overwrite_existing_files=overwrite_existing_files,
@@ -487,7 +489,7 @@ class DataShuttle:
         self,
         overwrite_existing_files: OverwriteExistingFiles = "never",
         dry_run: bool = False,
-    ) -> None:
+    ) -> list[str]:
         """Upload all files in the `derivatives` top level folder.
 
         Parameters
@@ -504,7 +506,7 @@ class DataShuttle:
             transfer was taking place, but no files will be moved.
 
         """
-        self._transfer_top_level_folder(
+        return self._transfer_top_level_folder(
             "upload",
             "derivatives",
             overwrite_existing_files=overwrite_existing_files,
@@ -517,7 +519,7 @@ class DataShuttle:
         self,
         overwrite_existing_files: OverwriteExistingFiles = "never",
         dry_run: bool = False,
-    ) -> None:
+    ) -> list[str]:
         """Download all files in the `rawdata` top level folder.
 
         Parameters
@@ -534,7 +536,7 @@ class DataShuttle:
             transfer was taking place, but no files will be moved..
 
         """
-        self._transfer_top_level_folder(
+        return self._transfer_top_level_folder(
             "download",
             "rawdata",
             overwrite_existing_files=overwrite_existing_files,
@@ -547,7 +549,7 @@ class DataShuttle:
         self,
         overwrite_existing_files: OverwriteExistingFiles = "never",
         dry_run: bool = False,
-    ) -> None:
[…]
