### 2025-10-21 9f642d7
Full draft implementation for TUI and API.
