### 2025-10-22 5cb61be
Extend to critical errors.
