### 2025-10-22 a81db02
Add test for `errors` output.
