### 2025-10-23 15d8102
Also show message if nothing was transferred.
