+++ datashuttle/tui/screens/modal_dialogs.py
@@ -242,26 +242,42 @@ class ConfirmAndAwaitTransferPopup(ModalScreen):
             if success:
                 errors = output
 
+                errors_message = ""
+
+                messagebox_kwargs = {}
+
+                no_transfer_col = (
+                    "blue"
+                    if self.app.theme == "textual-light"
+                    else "lightblue"
+                )
+
+                if errors["nothing_was_transferred_rawdata"] is True:
+                    errors_message += f"[{no_transfer_col}]\nNote! Nothing was transferred from rawdata[/{no_transfer_col}]\n"
+
+                if errors["nothing_was_transferred_derivatives"] is True:
+                    errors_message += f"[{no_transfer_col}]\nNote! Nothing was transferred from derivatives[/{no_transfer_col}]\n"
+
                 if any(errors["messages"]):
                     if errors["file_names"]:
-                        errors_message = (
+                        errors_message += (
                             "[red]Errors detected! in files:[/red]\n"
                         )
                         errors_message += "\n".join(errors["file_names"])
                     else:
-                        errors_message = "[red]Errors detected![/red]"
+                        errors_message += "[red]Errors detected![/red]"
                     errors_message += (
                         "[red]\n\nThe error messages are:[/red]\n"
                     )
                     errors_message += "\n\n".join(errors["messages"])
                     messagebox_kwargs = {"width": "75%", "height": "75%"}
-                else:
-                    errors_message = "No errors detected"
-                    messagebox_kwargs = {}
+
+                if errors_message == "":
+                    errors_message += "No errors detected"
 
                 message = (
-                    f"Transfer finished.\n\n"
-                    f"{errors_message}\n\n"
+                    f"Transfer finished.\n"
+                    f"{errors_message}\n"
                     f"Check the most recent logs for full details."
                 )
 
