### 2025-10-23 85f90f6
Docstrings and linting.
