+++ datashuttle/tui/screens/modal_dialogs.py
@@ -66,6 +66,12 @@ class MessageBox(ModalScreen):
         border_color
             Color of the MessageBox border (e.g. green if the message is positive).
 
+        height
+            The height of the messagebox.
+
+        width
+            The width of the messagebox.
+
         """
         super(MessageBox, self).__init__()
 
