### 2025-10-23 5afba61
Tidy up tests, add some docstrings.
