+++ tests/tests_transfers/local_filesystem/test_transfer.py
@@ -4,6 +4,7 @@ import time
 from pathlib import Path
 
 import pytest
+from filelock import FileLock
 
 from datashuttle.configs import canonical_folders
 from datashuttle.configs.canonical_configs import get_broad_datatypes
@@ -647,14 +648,18 @@ class TestFileTransfer(BaseTest):
         returned from every transfer function. `errors` is a variable
         that contains information about any errors that were encountered
         during transfer.
-
-        DOCS TODO this doesn't matter if its top level folder or whatever
         """
+
+        # Monkeypatch the error-parsing function so it returns
+        # predictable values.
+        import datashuttle
 
         def test_errors(top_level_folder):
             return {
                 "file_names": [f"{top_level_folder}/hello_world.txt"],
                 "messages": ["how are you?"],
+                "nothing_was_transferred_rawdata": None,
+                "nothing_was_transferred_derivatives": None,
             }
 
         def monkeypatch_parse_output(top_level_folder, b):
@@ -662,14 +667,13 @@ class TestFileTransfer(BaseTest):
             stderr = "stderr"
             return stdout, stderr, test_errors(top_level_folder)
 
-        import datashuttle
-
         monkeypatch.setattr(
             datashuttle.utils.rclone,
             "parse_rclone_copy_output",
             monkeypatch_parse_output,
         )
 
+        # Generate some test files so the transfer runs properly
         subs, sessions = test_utils.get_default_sub_sessions_to_test()
 
         for top_level_folder in ["rawdata", "derivatives"]:
@@ -681,10 +685,13 @@ class TestFileTransfer(BaseTest):
                 get_broad_datatypes(),
             )
 
+        # Run every transfer function and check that
+        # `errors` is returned correctly.
         specific_file = (
             lambda path_: f"{path_}/rawdata/{subs[0]}/{sessions[0]}/ephys/placeholder_file.txt"
         )
 
+        # All 'rawdata' functions
         for func in [
             lambda: project.upload_specific_folder_or_file(
                 specific_file(project.get_local_path())
@@ -699,9 +706,12 @@ class TestFileTransfer(BaseTest):
         ]:
             assert func() == test_errors("rawdata")
 
+        # All 'derivatives' functions
         for func in [project.upload_derivatives, project.download_derivatives]:
             assert func() == test_errors("derivatives")
 
+        # Entire project functions should merge the errors
+        # of rawdata and derivatives
         for func in [
             project.upload_entire_project,
             project.download_entire_project,
@@ -712,10 +722,19 @@ class TestFileTransfer(BaseTest):
                     "derivatives/hello_world.txt",
                 ],
                 "messages": ["how are you?", "how are you?"],
+                "nothing_was_transferred_rawdata": None,
+                "nothing_was_transferred_derivatives": None,
             }
 
     def test_errors_are_caught_and_logged(self, project):
-        """"""  # TODO DOC OMD TODO!
+        """
+        Create errors in the transfer by locking files, and check
+        the errors are correctly flagged in logs and `errors`. Also,
+        perform a transfer where no files are transferred, and check
+        this is flagged in logs and `errors`.
+        """
+
+        # Set up a folder to transfer
         subs, sessions = test_utils.get_default_sub_sessions_to_test()
 
         test_utils.make_and_check_local_project_folders(
@@ -726,6 +745,7 @@ class TestFileTransfer(BaseTest):
             get_broad_datatypes(),
         )
 
+        # Lock a file then perform the transfer, causing errors.
         relative_path = (
             Path("rawdata")
             / subs[0]
@@ -737,12 +757,11 @@ class TestFileTransfer(BaseTest):
 
         test_utils.delete_log_files(project.cfg.logging_path)
 
[…]
