### 2025-10-23 cafd40e
Update documentation: new images, better links to API. (#591)
* Update images.

* Set functions to API links.

* Update images.
