### 2025-10-23 56107b2
Rename 'Template validation' to 'Validation templates'.
