### 2025-10-24 17a51aa
Tidy up and fix tests.
