### 2025-10-27 7da6305
Introduce get_datetime_to_search_regexp function.
