### 2025-10-27 4c1bd4a
Introduce get_datetime_to_search_regexp function.
