### 2025-10-27 c60b5ff
Remove 'test_simple_wildcard_first'.
