### 2025-10-27 120c4d6
Use parameterisation for download test.
