### 2025-10-27 0a936a2
Add 'upload_or_download' for all key tests.
