### 2025-10-27 08d77d8
Small tidy ups.
