### 2025-10-27 2e45e2b
Support datetime directly as sub ses values @DATETIME@.
