### 2025-10-27 cc602a8
Add tests for new sub-@DATE@ ses-@DATE@ behaviour.
