### 2025-10-27 e04c8ae
Tests, different cases for locking file on windows vs. macos/linux.
