### 2025-10-27 06933ac
Extend tests to new sub case + some refactoring.
