### 2025-10-29 28335db
Fix docstring for linter.
