### 2025-10-29 3f01b0b
Improve error message on bad datatype combination. (#595)
