### 2025-10-29 6772736
Minor reformating on error messages.
