### 2025-10-29 d1a3c7a
Fix docstring.
