### 2025-10-30 b0dd3f3
Small tidy ups.
