+++ datashuttle/utils/rclone.py
@@ -3,6 +3,9 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Any, Dict, List, Literal, Optional
 
 if TYPE_CHECKING:
+    from pathlib import Path
+    from subprocess import CompletedProcess
+
     from datashuttle.configs.config_class import Configs
     from datashuttle.utils.custom_types import TopLevelFolder
 
@@ -11,8 +14,6 @@ import platform
 import shlex
 import subprocess
 import tempfile
-from pathlib import Path
-from subprocess import CompletedProcess
 
 from packaging import version
 
@@ -66,7 +67,7 @@ def call_rclone_for_central_connection(
 
 
 def call_rclone_through_script_for_central_connection(
-    cfg, command: str
+    cfg: Configs, command: str
 ) -> CompletedProcess:
     """Call rclone through a script.
 
@@ -74,6 +75,9 @@ def call_rclone_through_script_for_central_connection(
     Used for transfers due to generation of large call strings.
 
     Parameters
+    ----------
+    cfg
+        Datashuttle Configs class.
     ----------
     command
         Full command to run with RClone.
@@ -234,6 +238,9 @@ def setup_rclone_config_for_local_filesystem(
 
     Parameters
     ----------
+    cfg
+        datashuttle Configs class
+
     rclone_config_name
         canonical config name, generated by
         datashuttle.cfg.rclone.get_rclone_config_name()
