### 2025-10-30 eaa06b6
Add first rclone encryption test.
