### 2025-10-30 f9d3dbf
install pass on linux.
