+++ .github/workflows/code_test_and_deploy.yml
@@ -100,8 +100,10 @@ jobs:
       - name: Install pass on Linux
         if: runner.os == 'Linux'
         run: |
-          sudo apt-get update
-          sudo apt-get install -y pass
+          sudo apt-get update && sudo apt-get install -y pass gnupg
+          export GNUPGHOME=$(mktemp -d)
+          gpg --batch --passphrase '' --quick-gen-key "CI Key" default default never
+          pass init "CI Key"
 
       - name: RClone Encryption
         run: |
