### 2025-10-30 c1d1c65
Try a different pass set up command.
