+++ tests/tests_integration/test_rclone_encryption.py
@@ -1,3 +1,5 @@
+import os
+
 from datashuttle.utils import rclone_encryption
 
 from ..base import BaseTest
@@ -6,7 +8,7 @@ from ..tests_transfers.ssh import ssh_test_utils
 
 class TestRcloneEncryption(BaseTest):
     def test_set_and_remove_password(self, project):
-        """"""
+        """ """
         ssh_test_utils.setup_project_for_ssh(
             project,
         )
@@ -38,12 +40,16 @@ class TestRcloneEncryption(BaseTest):
 
         rclone_encryption.run_rclone_config_encrypt(project.cfg)
 
+        assert "RCLONE_PASSWORD_COMMAND" not in os.environ
+
         with open(rclone_config_path, "r", encoding="utf-8") as f:
             first_line = f.readline().strip()
 
         assert first_line == "# Encrypted rclone configuration File"
 
         rclone_encryption.remove_rclone_encryption(project.cfg)
+
+        assert "RCLONE_PASSWORD_COMMAND" not in os.environ
 
         with open(rclone_config_path, "r", encoding="utf-8") as f:
             first_line = f.readline().strip()
