+++ tests/tests_transfers/gdrive/test_gdrive_transfer.py
@@ -29,8 +29,9 @@ class TestGdriveTransfer(BaseTransfer):
 
         yield [pathtable, project]
 
-        rclone.call_rclone(
-            f"purge central_{project.project_name}_gdrive:{project.cfg['central_path'].parent} {rclone.get_config_arg(project.cfg)}"
+        rclone.call_rclone_for_central_connection(
+            project.cfg,
+            f"purge central_{project.project_name}_gdrive:{project.get_central_path()} {rclone.get_config_arg(project.cfg)}",
         )
 
     @pytest.mark.parametrize(
