### 2025-10-31 7f1f077
Refactor ssh connection set up tests.
