+++ datashuttle/datashuttle_class.py
@@ -1686,7 +1686,8 @@ class DataShuttle:
 
     def _setup_rclone_central_local_filesystem_config(self) -> None:
         rclone.setup_rclone_config_for_local_filesystem(
-            self.cfg, self.cfg.rclone.get_rclone_config_name("local_filesystem"),
+            self.cfg,
+            self.cfg.rclone.get_rclone_config_name("local_filesystem"),
         )
 
     def _setup_rclone_gdrive_config(
