### 2025-11-01 bbadcec
Finalise aws and other tests, revert changes made to setup_aws during erroneous 'fix linting' PR.
