+++ tests/tests_transfers/aws/test_tui_setup_aws.py
@@ -2,10 +2,12 @@ import os
 
 import pytest
 
+from datashuttle import DataShuttle
 from datashuttle.tui.app import TuiApp
 from datashuttle.tui.screens.project_manager import ProjectManagerScreen
 from datashuttle.utils import rclone, utils
 
+from ... import test_utils
 from ...tests_tui.tui_base import TuiBase
 from . import aws_test_utils
 
@@ -31,12 +33,18 @@ class TestTuiSetupAws(TuiBase):
 
         yield central_path, project_name
 
-        rclone.call_rclone(
-            f"purge central_{project_name}_aws:{central_path}"
-        )  # TODO: I think this will fail, needs config
+        project = DataShuttle(project_name)
+
+        rclone.call_rclone_for_central_connection(
+            project.cfg,
+            f"purge central_{project_name}_aws:{central_path} {rclone.get_config_arg(project.cfg)}",
+        )
 
     @pytest.mark.asyncio
-    async def test_aws_connection_setup(self, central_path_and_project):
+    @pytest.mark.parametrize("set_encryption", [True, False])
+    async def test_aws_connection_setup(
+        self, central_path_and_project, set_encryption
+    ):
         """Test AWS connection setup via the TUI.
 
         AWS connection details are filled in the configs tab. The setup
@@ -60,11 +68,39 @@ class TestTuiSetupAws(TuiBase):
             )
 
             assert (
-                "AWS Connection Successful!"
+                "Would you like to encrypt the RClone config file"
                 in pilot.app.screen.query_one(
                     "#setup_aws_messagebox_message"
                 ).renderable
             )
+
+            if set_encryption:
+                await self.scroll_to_click_pause(pilot, "#setup_aws_ok_button")
+
+                assert (
+                    "The rclone_encryption was successfully set. Setup complete!"
+                    in pilot.app.screen.query_one(
+                        "#setup_aws_messagebox_message"
+                    ).renderable
+                )
+
+                project = pilot.app.screen.interface.project
+
+                test_utils.check_rclone_file_is_encrypted(
+                    project.cfg.rclone.get_rclone_central_connection_config_filepath()
+                )
+
+            else:
+                await self.scroll_to_click_pause(
+                    pilot, "#setup_aws_cancel_button"
+                )
+
+                assert (
+                    "AWS Connection Successful!"
+                    in pilot.app.screen.query_one(
+                        "#setup_aws_messagebox_message"
+                    ).renderable
+                )
 
     @pytest.mark.asyncio
     async def test_aws_connection_setup_failed(self, central_path_and_project):
