+++ datashuttle/tui/screens/setup_aws.py
@@ -11,6 +11,8 @@ from textual.containers import Container, Horizontal, Vertical
 from textual.screen import ModalScreen
 from textual.widgets import Button, Input, Static
 
+from datashuttle.utils import rclone_encryption
+
 
 class SetupAwsScreen(ModalScreen):
     """Dialog window that sets up connection to an Amazon Web Service S3 bucket.
@@ -26,7 +28,7 @@ class SetupAwsScreen(ModalScreen):
         super(SetupAwsScreen, self).__init__()
 
         self.interface = interface
-        self.stage = 0
+        self.stage = "init"
 
     def compose(self) -> ComposeResult:
         """Set widgets on the SetupAwsScreen."""
@@ -52,18 +54,35 @@ class SetupAwsScreen(ModalScreen):
         self.query_one("#setup_aws_secret_access_key_input").visible = False
 
     def on_button_pressed(self, event: Button.Pressed) -> None:
-        """Handle button press on the screen."""
-        if event.button.id == "setup_aws_cancel_button":
-            self.dismiss()
+        """Handle button press on the screen.
 
-        if event.button.id == "setup_aws_ok_button":
-            if self.stage == 0:
+        The `setup_aws_ok_button` is used for all 'positive' events ('Yes, Ok')
+        and 'setup_aws_cancel_button' is used for 'negative' events ('No', 'Cancel').
+        The appropriate action to take on the button press is determined by the
+        current stage.
+
+        """
+        if event.button.id == "setup_aws_cancel_button":
+            if self.stage == "ask_rclone_encryption":
+                message = "AWS Connection Successful!"  #
+                self.query_one("#setup_aws_messagebox_message").update(message)
+                self.query_one("#setup_aws_ok_button").label = "Finish"
+                self.query_one("#setup_aws_cancel_button").remove()
+                self.stage = "finished"
+            else:
+                self.dismiss()
+
+        elif event.button.id == "setup_aws_ok_button":
+            if self.stage == "init":
                 self.prompt_user_for_aws_secret_access_key()
 
-            elif self.stage == 1:
+            elif self.stage == "use_secret_access_key":
                 self.use_secret_access_key_to_setup_aws_connection()
 
-            elif self.stage == 2:
+            elif self.stage == "ask_rclone_encryption":
+                self.set_rclone_encryption()
+
+            elif self.stage == "finished":
                 self.dismiss()
 
     def prompt_user_for_aws_secret_access_key(self) -> None:
@@ -73,10 +92,15 @@ class SetupAwsScreen(ModalScreen):
         self.query_one("#setup_aws_messagebox_message").update(message)
         self.query_one("#setup_aws_secret_access_key_input").visible = True
 
-        self.stage += 1
+        self.query_one("#setup_aws_ok_button")
+
+        self.stage = "use_secret_access_key"
 
     def use_secret_access_key_to_setup_aws_connection(self) -> None:
-        """Set up the AWS connection and inform user of success or failure."""
+        """Set up the AWS connection and failure.
+
+        If success, move onto the rclone_encryption screen.
+        """
         secret_access_key = self.query_one(
             "#setup_aws_secret_access_key_input"
         ).value
@@ -86,11 +110,13 @@ class SetupAwsScreen(ModalScreen):
         )
 
         if success:
-            message = "AWS Connection Successful!"
-            self.query_one(
-                "#setup_aws_secret_access_key_input"
-            ).visible = False
+            message = f"{rclone_encryption.get_explanation_message(self.interface.project.cfg)}"
+            self.query_one("#setup_aws_messagebox_message").update(message)
 
+            self.query_one("#setup_aws_secret_access_key_input").remove()
+            self.query_one("#setup_aws_ok_button").label = "Yes"
[…]
