### 2025-11-03 c2f4643
Refactor SSH connection to make adding passwords easier. (#602)
* Refactor SSH method.

* Tidy up.

* Update datashuttle/utils/rclone.py

Co-authored-by: Shrey Singh <96627769+cs7-shrey@users.noreply.github.com>

* Update datashuttle/utils/ssh.py

Co-authored-by: Shrey Singh <96627769+cs7-shrey@users.noreply.github.com>

* Fix docstring.

* Small tidy ups.

---------

Co-authored-by: Shrey Singh <96627769+cs7-shrey@users.noreply.github.com>
