### 2025-11-03 f4b6d4c
Add note on docs.
