### 2025-11-03 3153bff
Adjust tests to investigate failures.
