### 2025-11-03 bc27a76
Add back all tests.
