### 2025-11-03 b6818c0
Add workflow to run gdrive and aws tests on forks with a label.
