+++ tests/tests_transfers/aws/aws_test_utils.py
@@ -53,4 +53,8 @@ def has_aws_environment_variables():
     ]:
         if key not in os.environ:
             return False
+
+        if os.environ[key].strip() == "":
+            return False
+
     return True
