+++ tests/tests_transfers/gdrive/gdrive_test_utils.py
@@ -67,4 +67,8 @@ def has_gdrive_environment_variables():
     ]:
         if key not in os.environ:
             return False
+
+        if os.environ[key].strip() == "":
+            return False
+
     return True
