+++ tests/tests_transfers/aws/aws_test_utils.py
@@ -53,4 +53,9 @@ def has_aws_environment_variables():
     ]:
         if key not in os.environ:
             return False
+
+        # On CI triggered by forked repositories, secrets are empty
+        if os.environ[key].strip() == "":
+            return False
+
     return True
