+++ tests/tests_transfers/aws/test_aws_suggest_next.py
@@ -27,7 +27,7 @@ class TestAWSSuggestNext(BaseTransfer, TuiBase):
         yield project
 
         rclone.call_rclone(
-            f"purge central_{project.project_name}_gdrive:{project.cfg['central_path'].parent}"
+            f"purge central_{project.project_name}_gdrive:{project.cfg['central_path'].parent} {rclone.get_config_arg(project.cfg)}"
         )
 
     @pytest.mark.asyncio
