+++ .github/workflows/code_test_and_deploy.yml
@@ -111,7 +111,7 @@ jobs:
           AWS_REGION: ${{ secrets.AWS_REGION }}
           AWS_BUCKET_NAME: ${{ secrets.AWS_BUCKET_NAME }}
         run: |
-          pytest tests/tests_transfers/aws
+          pytest tests/tests_transfers/aws -v
 
 #      - name: All Other Tests
 #        run: |
