### 2025-11-04 683e376
Set DATASHUTTLE_GDRIVE_ROOT_FOLDER_ID in workflow
Add environment variable for Google Drive root folder ID.
