+++ .github/workflows/code_test_and_deploy.yml
@@ -113,7 +113,7 @@ jobs:
     if: github.event_name == 'push' && github.ref_type == 'tag'
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/download-artifact@v5
+    - uses: actions/download-artifact@v6
       with:
         name: artifact
         path: dist
