### 2025-11-05 06c77c8
Try moving to logging module to take advantage of the logging test machinery.
