### 2025-11-05 ad76736
Fix CI script.
