+++ tests/tests_transfers/aws/test_aws_transfer.py
@@ -27,8 +27,9 @@ class TestAwsTransfer(BaseTransfer):
 
         yield [pathtable, project]
 
-        rclone.call_rclone(
-            f"purge central_{project.project_name}_aws:{project.cfg['central_path'].parent} {rclone.get_config_arg(project.cfg)}"
+        rclone.call_rclone_for_central_connection(
+            project.cfg,
+            f"purge central_{project.project_name}_aws:{project.cfg['central_path'].parent} {rclone.get_config_arg(project.cfg)}",
         )
 
     @pytest.mark.parametrize(
