### 2025-11-05 e90077a
Revert CI changes.
