+++ .github/workflows/code_test_and_deploy.yml
@@ -89,20 +89,20 @@ jobs:
         # run SSH tests only on Linux because Windows and macOS
         # are already run within a virtual container and so cannot
         # run Linux containers because nested containerisation is disabled.
-#      - name: Test SSH (Linux only)
-#        if: runner.os == 'Linux'
-#        run: |
-#          sudo service mysql stop  # free up port 3306 for ssh tests
-#          pytest tests/tests_transfers/ssh
+      - name: Test SSH (Linux only)
+        if: runner.os == 'Linux'
+        run: |
+          sudo service mysql stop  # free up port 3306 for ssh tests
+          pytest tests/tests_transfers/ssh
 
-#      - name: Test Google Drive
-#        env:
-#          GDRIVE_CLIENT_ID: ${{ secrets.GDRIVE_CLIENT_ID }}
-#          GDRIVE_CLIENT_SECRET: ${{ secrets.GDRIVE_CLIENT_SECRET }}
-#          GDRIVE_ROOT_FOLDER_ID: ${{ secrets.GDRIVE_ROOT_FOLDER_ID }}
-#          GDRIVE_CONFIG_TOKEN: ${{ secrets.GDRIVE_CONFIG_TOKEN }}
-#        run: |
-#          pytest tests/tests_transfers/gdrive
+      - name: Test Google Drive
+        env:
+          GDRIVE_CLIENT_ID: ${{ secrets.GDRIVE_CLIENT_ID }}
+          GDRIVE_CLIENT_SECRET: ${{ secrets.GDRIVE_CLIENT_SECRET }}
+          GDRIVE_ROOT_FOLDER_ID: ${{ secrets.GDRIVE_ROOT_FOLDER_ID }}
+          GDRIVE_CONFIG_TOKEN: ${{ secrets.GDRIVE_CONFIG_TOKEN }}
+        run: |
+          pytest tests/tests_transfers/gdrive
 
       - name: Test AWS
         env:
@@ -113,9 +113,9 @@ jobs:
         run: |
           pytest tests/tests_transfers/aws -v
 
-#      - name: All Other Tests
-#        run: |
-#          pytest --ignore=tests/tests_transfers/ssh --ignore=tests/tests_transfers/gdrive --ignore=tests/tests_transfers/aws
+      - name: All Other Tests
+        run: |
+          pytest --ignore=tests/tests_transfers/ssh --ignore=tests/tests_transfers/gdrive --ignore=tests/tests_transfers/aws
 
   build_sdist_wheels:
     name: Build source distribution
