### 2025-11-05 0b8ca68
Improve explaination of templates and add missing closing bracket in docs.
