### 2025-11-06 a3f9ce9
Rename `name templates` to `validation templates` (#594)
* Rename 'name templates' to 'validation templates'.

* Update docs.

* Fix tests.

* Rename 'Template validation' to 'Validation templates'.

* Update images with tui renaming.

* Update backwards compatability tests.

* Improve explaination of templates and add missing closing bracket in docs.
