+++ datashuttle/tui/screens/create_folder_settings.py
@@ -29,7 +29,7 @@ from datashuttle.tui.tooltips import get_tooltip
 class CreateFoldersSettingsScreen(ModalScreen):
     """Settings for the Create Folders tab.
 
-    Handles setting datashuttle's `name_template`, as well
+    Handles setting datashuttle's `validation_template`, as well
     as the top-level-folder select and option to bypass all validation.
 
     Name Templates
@@ -37,8 +37,8 @@ class CreateFoldersSettingsScreen(ModalScreen):
     These are regexp templates that can be validated against
     during folder creation / project validation.
 
-    An input is provided to input a `name_template` for validation. When
-    the window is closed, the `name_template` is stored in datashuttle's
+    An input is provided to input a `validation_template` for validation. When
+    the window is closed, the `validation_template` is stored in datashuttle's
     persistent settings.
 
     The Create tab validation of input widgets is immediately updated
@@ -46,7 +46,7 @@ class CreateFoldersSettingsScreen(ModalScreen):
 
     Attributes
     ----------
-    Because the Input for `name_templates` is shared between subject
+    Because the Input for `validation_templates` is shared between subject
     and session, the values are held in the `input_values` attribute.
     These are loaded from `persistent_settings` on init.
 
@@ -89,8 +89,8 @@ class CreateFoldersSettingsScreen(ModalScreen):
         ses_on = not sub_on
 
         explanation = r"""
-        A 'Template' can be set check subject or session names are
-        formatted in a specific way.
+        You can define a custom “Template” to ensure subject or
+        session names follow a particular pattern or format.
 
         For example:
             sub-\d\d_id-.?.?.?_.*
@@ -141,9 +141,11 @@ class CreateFoldersSettingsScreen(ModalScreen):
                 Container(
                     Horizontal(
                         Checkbox(
-                            "Template validation",
+                            "Validation templates",
                             id="template_settings_validation_on_checkbox",
-                            value=self.interface.get_name_templates()["on"],
+                            value=self.interface.get_validation_templates()[
+                                "on"
+                            ],
                         ),
                         id="template_inner_horizontal_container",
                     ),
@@ -193,9 +195,9 @@ class CreateFoldersSettingsScreen(ModalScreen):
 
     def init_input_values_holding_variable(self) -> None:
         """Add the project Name Templates to the relevant Inputs."""
-        name_templates = self.interface.get_name_templates()
-        self.input_values["sub"] = name_templates["sub"]
-        self.input_values["ses"] = name_templates["ses"]
+        validation_templates = self.interface.get_validation_templates()
+        self.input_values["sub"] = validation_templates["sub"]
+        self.input_values["ses"] = validation_templates["ses"]
 
     def switch_template_container_disabled(self) -> None:
         """Switch the name template widgets disabled / enabled."""
@@ -205,7 +207,7 @@ class CreateFoldersSettingsScreen(ModalScreen):
         self.query_one("#template_inner_container").disabled = not is_on
 
     def fill_input_from_template(self) -> None:
-        """Fill the `name_templates` Input.
+        """Fill the `validation_templates` Input.
 
         The Input is shared between subject and session,
         depending on the current RadioSet value.
@@ -222,15 +224,15 @@ class CreateFoldersSettingsScreen(ModalScreen):
     def on_button_pressed(self, event: Button.Pressed) -> None:
         """Handle button press on the screen.
 
-        On close, update the `name_templates` stored in
+        On close, update the `validation_templates` stored in
         `persistent_settings` with those set on the TUI.
 
         Setting may error if templates are turned on but
         no template exists for either subject or session.
         """
         if event.button.id == "create_folders_settings_close_button":
[…]
