+++ datashuttle/tui/interface.py
@@ -35,7 +35,7 @@ class Interface:
     def __init__(self) -> None:
         """Initialise the Interface class."""
         self.project: DataShuttle
-        self.name_templates: Dict = {}
+        self.validation_templates: Dict = {}
         self.tui_settings: Dict = {}
 
         self.gdrive_rclone_setup_process: subprocess.Popen | None = None
@@ -182,7 +182,7 @@ class Interface:
                 top_level_folder,
                 sub_names,
                 ses_names,
-                self.get_name_templates(),
+                self.get_validation_templates(),
                 bypass_validation=False,
                 allow_letters_in_sub_ses_values=allow_letters_in_sub_ses_values,
             )
@@ -368,8 +368,8 @@ class Interface:
     # Name templates
     # ----------------------------------------------------------------------------------
 
-    def get_name_templates(self) -> Dict:
-        """Return the `name_templates` defining templates to validate against.
+    def get_validation_templates(self) -> Dict:
+        """Return the `validation_templates` defining templates to validate against.
 
         These are stored in a variable to avoid constantly
         reading these values from disk where they are stored in
@@ -377,19 +377,21 @@ class Interface:
         and the file contents are in sync, so when changed
         on the TUI side they are updated also, in `get_tui_settings`.
         """
-        if not self.name_templates:
-            self.name_templates = self.project.get_name_templates()
+        if not self.validation_templates:
+            self.validation_templates = self.project.get_validation_templates()
 
-        return self.name_templates
+        return self.validation_templates
 
-    def set_name_templates(self, name_templates: Dict) -> InterfaceOutput:
-        """Set the `name_templates` here and on disk.
+    def set_validation_templates(
+        self, validation_templates: Dict
+    ) -> InterfaceOutput:
+        """Set the `validation_templates` here and on disk.
 
-        See `get_name_templates` for more information.
+        See `get_validation_templates` for more information.
         """
         try:
-            self.project.set_name_templates(name_templates)
-            self.name_templates = name_templates
+            self.project.set_validation_templates(validation_templates)
+            self.validation_templates = validation_templates
             return True, None
 
         except BaseException as e:
@@ -398,7 +400,7 @@ class Interface:
     def get_tui_settings(self) -> Dict:
         """Return the "tui" field of `persistent_settings`.
 
-        Similar to `get_name_templates`, there are held on the
+        Similar to `get_validation_templates`, there are held on the
         class to avoid constantly reading from disk.
         """
         if not self.tui_settings:

[…]
