### 2025-11-06 8f3e444
Extend Mock Configs class.
