### 2025-11-12 afdf4e2
adapt conf.py and docs CI worflow for multi-version docs
