+++ .github/workflows/docs_build_and_deploy.yml
@@ -33,9 +33,14 @@ jobs:
     needs: build_sphinx_docs
     permissions:
       contents: write
-    if: github.event_name == 'push' && github.ref_type == 'tag'
+    # runs on releases, push to main, or manual dispatch on main
+    if: |
+      (github.event_name == 'push' && github.ref_type == 'tag') ||
+      (github.event_name == 'push' && github.ref == 'refs/heads/main') ||
+      (github.event_name == 'workflow_dispatch' && github.ref == 'refs/heads/main')
     runs-on: ubuntu-latest
     steps:
-      - uses: neuroinformatics-unit/actions/deploy_sphinx_docs@v2
+      - uses: neuroinformatics-unit/actions/deploy_sphinx_docs_multiversion@main
         with:
           secret_input: ${{ secrets.GITHUB_TOKEN }}
+          switcher-url: https://datashuttle.neuroinformatics.dev/latest/_static/switcher.json
