+++ tests/tests_integration/test_local_only_mode.py
@@ -9,7 +9,7 @@ from datashuttle.utils.custom_exceptions import (
 from .. import test_utils
 from ..base import BaseTest
 
-TEST_PROJECT_NAME = "test_project"
+TEST_PROJECT_NAME = "ds-unique-test-project-d375gd234vds2f"
 
 
 class TestLocalOnlyProject(BaseTest):
