+++ tests/tests_integration/test_logging.py
@@ -86,7 +86,7 @@ class TestLogging:
         Switch on datashuttle logging as required for
         these tests, then turn back off during tear-down.
         """
-        project_name = "test_project"
+        project_name = "ds-unique-test-project-d375gd234vds2f"
         test_utils.delete_project_if_it_exists(project_name)
         test_utils.set_datashuttle_loggers(disable=False)
 
@@ -290,7 +290,7 @@ class TestLogging:
         assert "Using config file from" in log
         assert "--include" in log
         assert "sub-11/ses-123/anat/**" in log
-        assert "/central/test_project/rawdata" in log
+        assert "/central/ds-unique-test-project-d375gd234vds2f/rawdata" in log
 
     @pytest.mark.parametrize("upload_or_download", ["upload", "download"])
     def test_logs_upload_and_download_folder_or_file(
