### 2025-11-18 60b30be
Fix validate_from_path.
