### 2025-11-18 fd91242
Fix perform_rclone_check.
