### 2025-11-19 d51d94e
Some tidy up and extend use cases for new encrpytion checker.
