### 2025-11-19 4ed8a8a
Apply some fixes from self review 1.
