+++ datashuttle/configs/rclone_configs.py
@@ -1,42 +1,44 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Dict, Optional
+from typing import TYPE_CHECKING, Optional
 
 if TYPE_CHECKING:
     from pathlib import Path
 
-    from datashuttle.utils.custom_types import (
-        OverwriteExistingFiles,
-    )
+    from datashuttle.configs.configs_class import Configs
 
 import yaml
 
 from datashuttle.configs import canonical_folders
-from datashuttle.utils import rclone_encryption, utils
+from datashuttle.utils import rclone_encryption
 
 
 class RCloneConfigs:
     """Class to manage the RClone configuration file.
 
     This is a file that RClone creates to hold all information about local and
-    remote transfer targets. For example, the ssh RClone config holds the private key.
+    central transfer targets. For example, the SSH RClone config holds the private key,
+    the GDrive rclone config holds the access token, etc.
 
     In datashuttle, local filesystem configs uses the Rclone default configuration file,
-    that RClone manages. However, remote transfers to ssh, aws and gdrive are held in
-    separate config files (set using RClone's --config argument). Then being separate
-    means these files can be separately encrypted.
+    that RClone manages, for backwards comatability reasons. However, SSH, AWS and GDrive
+    configs are stored in separate config files (set using RClone's --config argument).
+    Then being separate means these files can be separately encrypted.
 
     This class tracks the state on whether a RClone config is encrypted, as well
     as provides the default names for the rclone conf (e.g. central_<project_name>_<connection_method>).
 
     Parameters
     ----------
+    datashuttle_configs
+        Parent Configs class.
+
     config_base_class
         Path to the datashuttle configs folder where all configs for the project are stored.
 
     """
 
-    def __init__(self, datashuttle_configs, config_base_path):
+    def __init__(self, datashuttle_configs: Configs, config_base_path: Path):
         """Construct the class."""
         self.datashuttle_configs = datashuttle_configs
         self.rclone_encryption_state_file_path = (
@@ -73,8 +75,9 @@ class RCloneConfigs:
     def set_rclone_config_encryption_state(self, value: bool) -> None:
         """Store the current state of the rclone config encryption for the `connection_method`.
 
-        Note that this is stored to disk each call (rather than tracked locally) to ensure
-        it is updated live if updated through the Python API while the TUI is also running.
+        Note that this is stored to disk each call (rather than tracked in memory)
+        to ensure it is updated properly if changed through the Python API
+        while the TUI is also running.
         """
         assert rclone_encryption.connection_method_requires_encryption(
             self.datashuttle_configs["connection_method"]
@@ -118,27 +121,6 @@ class RCloneConfigs:
             canonical_folders.get_rclone_config_base_path()
             / f"{self.get_rclone_config_name()}.conf"
         )
-
-    def make_rclone_transfer_options(
-        self, overwrite_existing_files: OverwriteExistingFiles, dry_run: bool
-    ) -> Dict:
-        """Create a dictionary of rclone transfer options."""
-        allowed_overwrite = ["never", "always", "if_source_newer"]
-
-        if overwrite_existing_files not in allowed_overwrite:
-            utils.log_and_raise_error(
-                f"`overwrite_existing_files` not "
-                f"recognised, must be one of: "
-                f"{allowed_overwrite}",
-                ValueError,
-            )
-
-        return {
-            "overwrite_existing_files": overwrite_existing_files,
-            "show_transfer_progress": True,
-            "transfer_verbosity": "vv",
-            "dry_run": dry_run,
-        }
 
     def delete_existing_rclone_config_file(self) -> None:
         """Delete the Rclone config file if it exists."""
