### 2025-11-19 2c644a2
Rename to rclone_file_is_encrypted.
