### 2025-11-20 94b9037
Reverted 04233d5e67a6c982794ac9c2e3dc15f8227bcca5.
