+++ datashuttle/tui/screens/setup_ssh.py
@@ -152,7 +152,10 @@ class SetupSshScreen(ModalScreen):
         )
 
         if success:
-            message = "Connection successful! SSH key saved to the RClone config file."
+            message = (
+                f"Connection successful! SSH key "
+                f"saved to {self.interface.get_configs().ssh_key_path}"
+            )
             self.query_one("#setup_ssh_ok_button").label = "Finish"
             self.query_one("#setup_ssh_cancel_button").disabled = True
             self.stage += 1
