### 2025-11-20 31c80b0
Remove unecessary type hint ignore.
