### 2025-11-20 ccb718b
Rename preliminary_setup_gdrive_config_for_without_browser.
