+++ datashuttle/tui/interface.py
@@ -568,13 +568,11 @@ class Interface:
     ) -> InterfaceOutput:
         """Get the rclone message for Google Drive setup without a browser."""
         try:
-            output = (
-                rclone.preliminary_setup_gdrive_config_for_without_browser(
-                    self.project.cfg,
-                    gdrive_client_secret,
-                    self.project.cfg.rclone.get_rclone_config_name("gdrive"),
-                    log=False,
-                )
+            output = rclone.preliminary_setup_gdrive_config_without_browser(
+                self.project.cfg,
+                gdrive_client_secret,
+                self.project.cfg.rclone.get_rclone_config_name("gdrive"),
+                log=False,
             )
             return True, output
         except BaseException as e:
