+++ datashuttle/utils/gdrive.py
@@ -42,7 +42,7 @@ def prompt_and_get_config_token(
     with google drive and input the `config_token` generated by rclone. The `config_token` is
     then used to complete rclone's config setup for google drive.
     """
-    message = rclone.preliminary_setup_gdrive_config_for_without_browser(
+    message = rclone.preliminary_setup_gdrive_config_without_browser(
         cfg, gdrive_client_secret, rclone_config_name, log=log
     )
     input_ = utils.get_user_input(
