+++ datashuttle/tui/screens/setup_gdrive.py
@@ -29,7 +29,7 @@ class SetupGdriveScreen(ModalScreen):
 
     If the config contains a "gdrive_client_id", the user is prompted
     to enter a client secret. If the user has access to a browser, a Google Drive
-    authentication page will open. Otherwise, the user is asked to run a rclone command
+    authentication page will open. Otherwise, the user is asked to run an Rclone command
     and input a config token.
     """
 
