### 2025-11-20 0b2a9b2
Fix fancylog version.
