+++ pyproject.toml
@@ -16,7 +16,7 @@ dependencies = [
     "PyYAML",
     "requests",
     "rich",
-    "fancylog>=0.4.2",
+    "fancylog==0.4.5",
     "simplejson",
     "pyperclip",
     "textual==3.4.0",
