### 2025-11-20 13bd9c7
Not sure why pinning to fancylog version that was used in previous ci failed.
