### 2025-11-20 1700651
Revert change to create_folders conditional.
