+++ datashuttle/tui/tabs/create_folders.py
@@ -42,7 +42,6 @@ from datashuttle.tui.utils.tui_decorators import (
     require_double_click,
 )
 from datashuttle.tui.utils.tui_validators import NeuroBlueprintValidator
-from datashuttle.utils import rclone_encryption
 
 
 class CreateFoldersTab(TreeAndInputTab):
@@ -339,12 +338,9 @@ class CreateFoldersTab(TreeAndInputTab):
             in canonical_configs.get_connection_methods_list()
         )
 
-        if (
-            include_central
-            and rclone_encryption.connection_method_requires_encryption(
-                self.interface.project.cfg["connection_method"]
-            )
-        ):
+        if include_central and self.interface.project.cfg[
+            "connection_method"
+        ] in ["aws", "gdrive", "ssh"]:
             self.searching_central_popup_widget = (
                 SearchingCentralForNextSubSesPopup(prefix)
             )
