### 2025-11-20 7cea970
Small refactor.
