### 2025-11-20 2e3a710
Remove --ask-password.
