+++ datashuttle/utils/rclone_encryption.py
@@ -235,7 +235,7 @@ def run_rclone_config_encrypt(cfg: Configs) -> None:
         raise RuntimeError(
             f"\n--- STDOUT ---\n{output.stdout}\n"
             f"\n--- STDERR ---\n{output.stderr}\n"
-            "Could not remove the password from the RClone config. See the error message above."
+            "Could not encrypt the RClone config. See the error message above."
         )
 
     remove_credentials_as_password_command()
