### 2025-11-20 cec3c50
Improve a test docstring.
