+++ tests/tests_integration/test_rclone_encryption.py
@@ -1,4 +1,5 @@
 import os
+import textwrap
 
 from datashuttle.utils import rclone_encryption
 
@@ -8,8 +9,13 @@ from ..tests_transfers.ssh import ssh_test_utils
 
 
 class TestRcloneEncryption(BaseTest):
-    def test_set_and_remove_password(self, project):
-        """ """
+    def test_set_and_remove_rclone_config_encryption(self, project):
+        """Test that RClone config encryption is set up and torn down correctly.
+
+        First, we generate a mock RClone config file or a mock SSH project.
+        Then we check datashuttle functions for encryption, decryption and
+        cleaning up environment variables work as expected.
+        """
         ssh_test_utils.setup_project_for_ssh(
             project,
         )
@@ -20,8 +26,6 @@ class TestRcloneEncryption(BaseTest):
 
         if rclone_config_path.exists():
             rclone_config_path.unlink()
-
-        import textwrap
 
         config_content = textwrap.dedent(f"""\
         [{project.cfg.rclone.get_rclone_config_name()}]
@@ -49,6 +53,7 @@ class TestRcloneEncryption(BaseTest):
 
         assert "RCLONE_PASSWORD_COMMAND" not in os.environ
 
+        # Read the file contents to check it is no longer encrypted.
         with open(rclone_config_path, "r", encoding="utf-8") as f:
             first_line = f.readline().strip()
 
