### 2025-11-20 0f50c1b
Update datashuttle/configs/rclone_configs.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
