### 2025-11-20 ed1a89d
Update datashuttle/configs/rclone_configs.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
