### 2025-11-20 f5f44da
Update datashuttle/utils/rclone_encryption.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
