### 2025-11-20 b1a5032
Fix broken try / except.
