### 2025-11-20 d39c88b
Fix empty string case. (#634)
