### 2025-11-20 06b4a93
Replace direct raise with utils function.
