### 2025-11-20 6fdb357
Fixes to rclone_encryption.py
