### 2025-11-20 3d2cd66
Update datashuttle/datashuttle_class.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
