### 2025-11-20 531eb02
Update datashuttle/configs/canonical_folders.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
