### 2025-11-20 8f14c9e
Update datashuttle/datashuttle_class.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
