### 2025-11-20 e49b526
Fix broken datashuttle version check.
