+++ datashuttle/__init__.py
@@ -9,7 +9,3 @@ try:
 except PackageNotFoundError:
     # package is not installed
     pass
-
-
-def get_datashuttle_version():
-    return __version__
