### 2025-11-20 4d5450e
Update datashuttle/utils/rclone_encryption.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
