+++ datashuttle/utils/rclone_encryption.py
@@ -125,7 +125,7 @@ def set_password_linux(cfg: Configs) -> None:
         utils.log_and_raise_error(
             f"\n--- STDOUT ---\n{output.stdout}"
             f"\n--- STDERR ---\n{output.stderr}"
-            "Could encrypt the password from the RClone config. See the error message above.",
+            "Could not encrypt the password from the RClone config. See the error message above.",
             RuntimeError,
         )
 
