### 2025-11-20 33ce2d7
Update datashuttle/tui/screens/setup_aws.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
