### 2025-11-20 bf84a44
Update datashuttle/tui/screens/setup_aws.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
