### 2025-12-16 #8732 @adswa
I can probably squeeze this in over christmas :+1: feel free to assign me as reviewer.
