### 2025-12-17 #9642 @editorialbot
[REVIEW]: datashuttle: automated data management for experimental neuroscience
**Submitting author:** @JoeZiminski (<a href="http://orcid.org/0000-0003-4286-6868">Joseph Ziminski</a>)
**Repository:** https://github.com/neuroinformatics-unit/datashuttle
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.7.1
