### 2025-12-27 #9642 @adswa
- [x] **Community guidelines:** Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support -> https://github.com/neuroinformatics-unit/datashuttle/issues/661
