### 2025-12-30 neuroinformatics-unit/datashuttle#661 @adswa
[JOSS] Contributing guide lacks information on testing
Hey there :) :wave:
As part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/9642), I came here to ask for (better) contributor instructions for automated local testing.
I figured out how to run the tests (and they ran successfully) from your GitHub workflows, but I could not find detailed information in the [CONTRIBUTING GUIDE](https://datashuttle.neuroinformatics.dev/pages/community/contributing.html#contributing-documentation), where instructions suggest to rely on the GitHub actions running on submitted PRs. However, there is lot of value in being able to run tests locally -- to give newcomers the chance to contribute or simply test the software locally/on a  specific production system.
My suggestion would be extend the contributing guide with a brief section on how to invoke the tests in a development installation (and maybe advice which tests run 'local only' or require access to third party infrastructure (and how to configure that).
