### 2026-01-02 #9642 @JoeZiminski
# Changes to make to the JOSS paper

- [x] Extend the 'Statement of Need' to give a deeper review / comparison with other _tools_ used for standardisation and transfer (e.g. BIDS / NWB tools for standardisation, datajoint, datalad, and even rsync and dropbox for transfer)
