### 2026-01-04 #9642 @sneakers-the-rat
I think it makes sense to have an "author checklist" here that gets updated with TODO items that does the same thing as the reviewer checklists: open issues in the reviewed repo, and then make links back and forth from checklist here <-> opened issues so we can keep track of things without cluttering the review issue too much.

The reviewers were too on top of it and i was too slow to give my usual editorial spiel about keeping the review tidy, but looks like everyone is way ahead of me anyway!

Edit: also to be clear, that list of other tools i mentioned was just an example, dont feel like you need to address all/any of/only those tools, i trust the authors have a better sense of the landscape of related tools than i do.
