### 2026-02-03 #9642 @JoeZiminski
Hi @adswa @sneakers-the-rat thanks for your comments, a [section has been added](https://github.com/neuroinformatics-unit/datashuttle/pull/665) to the documentation on how to test locally, and a new paragraph (third paragraph in 'Statement of Need') has been added to the paper on existing tools.
