### 2026-03-19 #9642 @adswa
Sorry for the delay on my side - I set aside time between the years (and used that even), but when do you ever estimate the right amount of time for something in academia? Anyway, thanks for addressing the minor documentation/testing issue that I faced. I played with the software on my own system (Debian) and data (toy data), and also a remote system (SSH), and found everything to be working as promised. I did not try everything (e.g., I haven't tested S3 transfers), but I'm confident that the software and the effort behind it clearly warrants this publication.
