### 2026-04-21 #9642 @sneakers-the-rat
sorry for my delay, had a notification storm in the intervening time.

If @adswa is satisfied, has completed their checklist, and approves of the work, then that is fabulous! Thank you for your time. No need to test exhaustively, if you were able to run it and validate the claims using whatever transfer mechanism is familiar to you, then that is more than enough validation.

I believe we are still waiting on @likeajumprope - did you run into a blocking problem, or just need more time? Please let me know if you need anything, and my email is j@nny.fyi if you'd like to talk privately.

I'll give @likeajumprope a bit to respond, we need at least a second reviewer to complete things, but if they are busy/can't finish their review (which is fine! your time is volunteered not mandated, everyone gets busy) then we will have to look around and find another.
