### 2026-05-07 #9642 @sneakers-the-rat
Alright not getting takers on the fedi, let me search the reviewer DB and see if anyone can pick up the second review:

- @danielmk I see neuro data listed in your expertise,
- @Julie-Fabre you have neuroscience and quality control in your expertise, which seems relevant to the data validation part
- @jsheunis you also have RDM and QC in bio,

would any of you be interested in reviewing this package?
