### 2026-05-15 #9642 @danielmk
This is my review of the datashuttle submission. I read the manuscript, installed the software, worked through some tutorials, and worked through the JOSS checklist. Because everything worked as described and the manuscript is in good overall condition, I recommend acceptance of the submission.

I have just two minor comments on the manuscript that the authors might want to improve (not necessary for acceptance):
-	On line 32, “Neurodata Without Borders” is written out but never abbreviated, although it appears abbreviated as BIDS later in the manuscript.
-	The paragraph starting on 34 could use a rewrite. It is currently hard to read.
    -	The first two sentences both make the same very general statement about the ecosystem.
    -	BIDS has a different reference here than in the previous paragraph. Why does BIDS need a reference again, and what is the difference between this reference and the other reference?
    -	The citation for NWB is repeated, and a different one is added. The reference does not need repeating. And why does the Teeters reference appear here but not when NWB is introduced in the previous paragraph?
    -	In lines 40 and 41 the same reference (Gorgolewski et al., 2020) appears twice in the same sentence.

A quick report on what I did to verify functionality.
-	I installed the package with conda in a new environment.
-	I made a basic project structure outside the package, with deliberate formatting errors, and used the terminal interface’s “Validate Project From Path” to catch these errors.
-	I used the terminal UI to create a project without connection
-	I used the terminal UI to a local and a remote mounted drive, and downloaded / uploaded some ephys files and images from an old project.
-	I used the Python API to create a project with local connection and uploaded some data.

All of that worked without problems.

Please let me know if you have any questions or need anything else for the review.

Kind regards,
Daniel
