### 2026-06-05 #9642 @sneakers-the-rat
@JoeZiminski when the above request is addressed, and https://github.com/neuroinformatics-unit/datashuttle/pull/726 either merged or addressed, can you make an archival release and post the DOI and corresponding version here?

We don't have docs for this yet, but I drafted some for another recent submission here - https://github.com/openjournals/joss/pull/1547
